### 2021-09-23 #84 @whedon
[PRE REVIEW]: ReactiveMP.jl: A Julia Package for Reactive Message Passing-based Bayesian Inference
**Submitting author:** @bvdmitri (<a href="http://orcid.org/0000-0001-9655-7986">Dmitry Bagaev</a>)
**Repository:** <a href="https://github.com/biaslab/ReactiveMP.jl" target ="_blank">https://github.com/biaslab/ReactiveMP.jl</a>
**Version:**
