### 2021-09-23 #84 @bvdmitri
Paper is located in the different branch: https://github.com/biaslab/ReactiveMP.jl/tree/paper/julia_proceedings_2021
