### 2021-09-23 a4f1836
fix(): Fix GaussianMeanPrecision rules to use any gaussian parametrisation
