+++ src/rules/normal_mean_precision/mean.jl
@@ -12,4 +12,7 @@ end
 # --------------------------------- #
 @rule NormalMeanPrecision(:μ, Marginalisation) (q_out::Any, q_τ::Any) = NormalMeanPrecision(mean(q_out), mean(q_τ))
 
-@rule NormalMeanPrecision(:μ, Marginalisation) (m_out::UnivariateNormalDistributionsFamily, q_τ::Any) = NormalMeanPrecision(mean(m_out), cholinv( cov(m_out) + cholinv(mean(q_τ)) ))
+@rule NormalMeanPrecision(:μ, Marginalisation) (m_out::UnivariateNormalDistributionsFamily, q_τ::Any) = begin 
+    m_out_mean, m_out_cov = mean_cov(m_out)
+    return NormalMeanPrecision(m_out_mean, cholinv(m_out_cov + cholinv(mean(q_τ)) ))
+end
