### 2021-09-23 #84 @matbesancon
hi @bvdmitri, could you adapt you submission to the template:
https://github.com/JuliaCon/JuliaConSubmission.jl/
there are some important files in the folder that are not in yours, in particular the .latexmkrc file
