### 2021-09-23 dd02d0c
feat(): more efficient versions of marginal rules for gaussians
