### 2021-09-23 f9baf90
feat(): added more marginal rules for nmp parametrisation
