+++ src/rules/normal_mean_precision/marginals.jl
@@ -1,4 +1,3 @@
-export marginalrule
 
 @marginalrule NormalMeanPrecision(:out_μ_τ) (m_out::UnivariateNormalDistributionsFamily, m_μ::PointMass, m_τ::PointMass) = begin
     return (out = prod(ProdAnalytical(), NormalMeanPrecision(mean(m_μ), mean(m_τ)), m_out), μ = m_μ, τ = m_τ)
@@ -20,7 +19,7 @@ end
     return (out_μ = MvNormalWeightedMeanPrecision(xi, W), τ = m_τ)
 end
 
-@marginalrule NormalMeanPrecision(:out_μ) (m_out::UnivariateNormalDistributionsFamily, m_μ::UnivariateNormalDistributionsFamily, q_τ::GammaDistributionsFamily) = begin
+@marginalrule NormalMeanPrecision(:out_μ) (m_out::UnivariateNormalDistributionsFamily, m_μ::UnivariateNormalDistributionsFamily, q_τ::Any) = begin
     xi_out, W_out = weightedmean_precision(m_out)
     xi_μ, W_μ     = weightedmean_precision(m_μ)
 
