### 2021-09-23 d54dfaf
fix(): added .latexmkrc and new header
