+++ paper/header.tex
@@ -1,19 +1,17 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
-\title{My JuliaCon proceeding}
+\title{ReactiveMP.jl: A Julia Package for Reactive Message Passing-based Bayesian Inference}
 
-\author[1]{1st author}
-\author[1, 2]{2nd author}
-\author[2]{3rd author}
-\affil[1]{University}
-\affil[2]{National Lab}
+\author[1]{Dmitry Bagaev}
+\author[1]{Bert de Vries}
+\affil[1]{Electrical Eng. Dept., Eindhoven Univ. of Technology, Eindhoven, The Netherlands}
 
-\keywords{Julia, Optimization, Game theory, Compiler}
+\keywords{Bayesian Inference, Julia, Free Energy Minimisation, Message Passing, Reactive Programing, Variational Inference}
 
 \hypersetup{
-pdftitle = {My JuliaCon proceeding},
+pdftitle = {ReactiveMP.jl: A Julia Package for Reactive Message Passing-based Bayesian Inference},
 pdfsubject = {JuliaCon 2019 Proceedings},
-pdfauthor = {1st author, 2nd author, 3rd author},
-pdfkeywords = {Julia, Optimization, Game theory, Compiler},
+pdfauthor = {Dmitry Bagaev, Bert de Vries},
+pdfkeywords = {Bayesian Inference, Julia, Free Energy Minimisation, Message Passing, Reactive Programing, Variational Inference},
 }
 
