### 2021-09-23 e38e57d
add test for joint functions
