+++ test/nodes/flow/flow_models/test_flow_model.jl
@@ -285,6 +285,21 @@ using ReactiveMP
 
     end
 
+    @testset "Joint processing functions" begin
+        model = FlowModel(
+            (
+                InputLayer(10),
+                AdditiveCouplingLayer(PlanarFlow()),
+                AdditiveCouplingLayer(PlanarFlow(); permute=false)
+            )
+        )
+        compiled_model = compile(model)
+        x = randn(10)
+        @test forward_jacobian(compiled_model, x) == (forward(compiled_model, x), jacobian(compiled_model, x))
+        @test backward_inv_jacobian(compiled_model, x) == (backward(compiled_model, x), inv_jacobian(compiled_model, x))
+
+    end
+
     @testset "Utility Jacobian" begin
         
         # check utility functions jacobian (single layer)
