### 2021-09-23 7770a81
optimize rules for efficiency
