### 2021-09-24 a6e3974
fix incorrect inv_jacobian when dim > 2
