### 2021-09-24 f8c4e29
refactor flow model processing
