+++ src/nodes/flow/flow_models/flow_model.jl
@@ -1,5 +1,3 @@
-## TODO: create a forward-jacobian joint function that calculates both at the same time
-
 export FlowModel, CompiledFlowModel
 export compile, nr_params
 export getlayers, getforward, getbackward, getjacobian, getinv_jacobian
@@ -244,198 +242,6 @@ end
 backward!(input::Vector{T1}, layers::Tuple{}, output::Vector{T2}) where { T1 <: Real, T2 <: Real } = return
 
 
-# jacobian of the Flow model
-function _jacobian(model::CompiledFlowModel, input::Vector{T1}) where { T1 <: Real }
-
-    # fetch layers
-    dim    = getdim(model)
-    
-    # promote type for allocating output
-    T = promote_type(eltype(model), T1)
-
-    # allocate space for jacobian
-    J = zeros(T, dim, dim)
-    for k = 1:dim
-        J[k,k] = 1.0
-    end
-
-    # calculate jacobian
-    jacobian!(J, model, input)
-
-    # return result
-    return J
-
-end
-
-# when calling jacobian, redirect to _jacobian
-jacobian(model::CompiledFlowModel, input::Vector{T}) where { T <: Real } = _jacobian(model, input)
-
-# for broadcasting over jacobian, fix the model for multiple inputs
-Broadcast.broadcasted(::typeof(jacobian), model::CompiledFlowModel, input::Vector{Vector{T}}) where { T <: Real } = broadcast(_jacobian, Ref(model), input)
-
-# inplace jacobian of the Flow model
-function jacobian!(J::Matrix{T1}, model::CompiledFlowModel, input::Vector{T2}) where { T1 <: Real, T2 <: Real }
-
-    # fetch layers
-    layers = getlayers(model)
-    dim = getdim(model)
-    
-    # promote type for allocating output
-    T = promote_type(eltype(model), T2)
-
-    # allocate space for intermediate output and jacobian
-    input_new = zeros(T, dim)
-    input_new .= input
-    output = zeros(T, dim)
-    J_new = zeros(T, dim, dim)
-    J_old = zeros(T, dim, dim)
-    for k = 1:dim
-        J_old[k,k] = 1.0
-    end
-
-    # calculate jacobian over all layers
-    jacobian!(J, J_new, J_old, output, input_new, layers)
-
-end
-
-# inplace jacobian calculation for a tuple of layers
-function jacobian!(J::Matrix{T1}, J_new::Matrix{T2}, J_old::Matrix{T3}, output::Vector{T4}, input_new::Vector{T5}, layers::T) where { T1 <: Real, T2 <: Real,  T3 <: Real, T4 <: Real, T5 <: Real, T <: NTuple{N,AbstractLayer} where N}
-
-    # check for specialized jacobians that are fixed (e.g. Permutation Matrix)
-    if typeof(first(layers)) <: PermutationLayer
-        
-        # calculate total jacobian
-        mul!(J, jacobian(first(layers), input_new), J_old)
-        
-        # else fall back to standard calculation
-    else 
-        
-        # calculate new jacobian
-        jacobian!(J_new, first(layers), input_new)
-        
-        # calculate total jacobian
-        mul!(J, J_new, J_old)
-        
-    end
-
-    # update old jacobian
-    J_old .= J
-
-    # if we did not arrive at the last layer yet, then also perform a forward pass
-    if length(layers) > 1
-
-        # perform forward pass over last layer
-        forward!(output, first(layers), input_new)
-
-        # update intermediate output
-        input_new .= output
-
-    end
-
-    # calculate jacobian of remaining layers
-    jacobian!(J, J_new, J_old, output, input_new, Base.tail(layers))
-
-end
-
[…]
