### 2021-09-25 #84 @matbesancon
perfect it compiles. General remark, the abstract is a bit short and would benefit at least from additional definitions, context, and a section specifically on the package itself
