### 2021-09-27 #84 @matbesancon
these comments are for the whole paper. You submitted it in a short version (extended abstract) which are one page + references, so you still have some place to extend a bit
