### 2021-09-27 #84 @bvdmitri
@matbesancon I thought references have to be on the same page, If I extend my text references go on the page number 2. Is that allowed? If yes, then I would be happy to extend it.
