### 2021-09-27 #84 @matbesancon
yes that's allowed, see https://juliacon.github.io/proceedings-guide/author/
> one page + references

so the references are excluded from the page limit
