+++ paper/header.tex
@@ -6,12 +6,12 @@
 \author[1]{Bert de Vries}
 \affil[1]{Electrical Eng. Dept., Eindhoven Univ. of Technology, Eindhoven, The Netherlands}
 
-\keywords{Bayesian Inference, Julia, Free Energy Minimisation, Message Passing, Reactive Programing, Variational Inference}
+\keywords{Bayesian Inference, Julia, Factor Graphs, Graphical Models, Free Energy Minimisation, Message Passing, Reactive Programing, Variational Inference}
 
 \hypersetup{
 pdftitle = {ReactiveMP.jl: A Julia Package for Reactive Message Passing-based Bayesian Inference},
 pdfsubject = {JuliaCon 2019 Proceedings},
 pdfauthor = {Dmitry Bagaev, Bert de Vries},
-pdfkeywords = {Bayesian Inference, Julia, Free Energy Minimisation, Message Passing, Reactive Programing, Variational Inference},
+pdfkeywords = {Bayesian Inference, Julia, Factor Graphs, Graphical Models, Free Energy Minimisation, Message Passing, Reactive Programing, Variational Inference},
 }
 
