### 2021-09-28 #84 @bvdmitri
Dear @matbesancon, new version of my proposal is available on the same branch.
