### 2021-09-29 0d2dd94
fix(): Fix rules for better performance
