+++ src/rules/gcv/x.jl
@@ -2,18 +2,28 @@ export rule
 
 @rule GCV(:x, Marginalisation) (m_y::Any, q_z::Any, q_κ::Any, q_ω::Any) = begin
 
-    ksi = mean(q_κ) ^ 2 * var(q_z) + mean(q_z) ^ 2 * var(q_κ) + var(q_z) * var(q_κ)
-    A = exp(-mean(q_ω) + var(q_ω) / 2)
-    B = exp(-mean(q_κ) * mean(q_z) + ksi / 2)
+    y_mean, y_var = mean_var(m_y)
+    z_mean, z_var = mean_var(q_z)
+    κ_mean, κ_var = mean_var(q_κ)
+    ω_mean, ω_var = mean_var(q_ω)
+    
+    ksi = κ_mean ^ 2 * z_var + z_mean ^ 2 * κ_var + z_var * κ_var
+    A = exp(-ω_mean + ω_var / 2)
+    B = exp(-κ_mean * z_mean + ksi / 2)
 
-    return NormalMeanVariance(mean(m_y), var(m_y) + inv(A * B))
+    return NormalMeanVariance(y_mean, y_var + inv(A * B))
 end
 
 @rule GCV(:x, Marginalisation) (q_y::Any, q_z::Any, q_κ::Any, q_ω::Any) = begin
 
-    ksi = mean(q_κ) ^ 2 * var(q_z) + mean(q_z) ^ 2 * var(q_κ) + var(q_z) * var(q_κ)
-    A = exp(-mean(q_ω) + var(q_ω) / 2)
-    B = exp(-mean(q_κ) * mean(q_z) + ksi / 2)
+    y_mean        = mean(q_y)
+    z_mean, z_var = mean_var(q_z)
+    κ_mean, κ_var = mean_var(q_κ)
+    ω_mean, ω_var = mean_var(q_ω)
 
-    return NormalMeanVariance(mean(q_y),  inv(A * B))
+    ksi = κ_mean ^ 2 * z_var + z_mean ^ 2 * κ_var + z_var * κ_var
+    A = exp(-ω_mean + ω_var / 2)
+    B = exp(-κ_mean * z_mean + ksi / 2)
+
+    return NormalMeanVariance(y_mean, inv(A * B))
 end
