+++ src/rules/gcv/w.jl
@@ -2,34 +2,38 @@ export rule
 
 
 @rule GCV(:ω, Marginalisation) (q_y_x::Any, q_z::Any, q_κ::Any, meta::GCVMetadata) = begin
-    
-    m, v = mean(q_y_x), cov(q_y_x)
 
-    γ = mean(q_z) ^ 2 * var(q_κ) + mean(q_κ) ^ 2 * var(q_z) + var(q_z) * var(q_κ)
-    A = exp(-mean(q_κ) * mean(q_z) + γ / 2)
-    psi = (m[1] - m[2]) ^ 2 + v[1, 1] + v[2, 2] - v[1, 2] - v[2, 1]
+    y_x_mean, y_x_var = mean_cov(q_y_x)
+    z_mean, z_var     = mean_var(q_z)
+    κ_mean, κ_var     = mean_var(q_κ)
+
+    γ = z_mean ^ 2 * κ_var + κ_mean ^ 2 * z_var + z_var * κ_var
+    A = exp(-κ_mean * z_mean + γ / 2)
+    psi = (y_x_mean[1] - y_x_mean[2]) ^ 2 + y_x_var[1, 1] + y_x_var[2, 2] - y_x_var[1, 2] - y_x_var[2, 1]
 
     a = one(typeof(γ))
     b = psi * A
     c = -one(typeof(γ))
     d = zero(typeof(γ))
 
-    return ExponentialLinearQuadratic(get_approximation(meta),a, b, c, d)
+    return ExponentialLinearQuadratic(get_approximation(meta), a, b, c, d)
 end
 
 @rule GCV(:ω, Marginalisation) (q_y::Any, q_x::Any, q_z::Any, q_κ::Any, meta::GCVMetadata) = begin
     
-    my, vy = mean(q_y), cov(q_y)
-    mx, vx = mean(q_x), cov(q_x)
+    y_mean, y_var = mean_var(q_y)
+    x_mean, x_var = mean_var(q_x)
+    z_mean, z_var = mean_var(q_z)
+    κ_mean, κ_var = mean_var(q_κ)
 
-    γ = mean(q_z) ^ 2 * var(q_κ) + mean(q_κ) ^ 2 * var(q_z) + var(q_z) * var(q_κ)
-    A = exp(-mean(q_κ) * mean(q_z) + γ / 2)
-    psi = (my - mx) ^ 2 + vy + vx
+    γ = z_mean ^ 2 * κ_var + κ_mean ^ 2 * z_var + z_var * κ_var
+    A = exp(-κ_mean * z_mean + γ / 2)
+    psi = (y_mean - x_mean) ^ 2 + y_var + x_var
 
-    a = mean(q_z)
+    a = z_mean
     b = psi * A
     c = -a
-    d = var(q_z)
+    d = z_var
 
     return ExponentialLinearQuadratic(get_approximation(meta),a, b, c, d)
 end
