### 2021-09-29 9b6d755
feat(): Update GCV with gaussian note extensions
