+++ src/distributions/exp_linear_quadratic.jl
@@ -13,10 +13,13 @@ end
 ExponentialLinearQuadratic(approximation, a::Real, b::Real, c::Real, d::Real)             = ExponentialLinearQuadratic(approximation, promote(a, b, c, d)...)
 ExponentialLinearQuadratic(approximation, a::Integer, b::Integer, c::Integer, d::Integer) = ExponentialLinearQuadratic(approximation, float(a), float(b), float(c), float(d))
 
+mean_cov(dist::ExponentialLinearQuadratic) = mean_var(dist)
+mean_var(dist::ExponentialLinearQuadratic) = approximate_meancov(dist.approximation, (z) -> pdf(dist, z) * exp(0.5 * z^2), NormalMeanVariance(0.0, 1.0))
+
 Distributions.pdf(dist::ExponentialLinearQuadratic, x::Real)    = exp(logpdf(dist, x))
 Distributions.logpdf(dist::ExponentialLinearQuadratic, x::Real) = -0.5 * (dist.a * x + dist.b * exp(dist.c * x + dist.d * x ^ 2 / 2.0))
-Distributions.mean(dist::ExponentialLinearQuadratic)            = approximate_meancov(dist.approximation, (z) -> pdf(dist, z) * exp(0.5 * z^2), NormalMeanVariance(0.0,1.0))[1]
-Distributions.var(dist::ExponentialLinearQuadratic)             = approximate_meancov(dist.approximation, (z) -> pdf(dist, z) * exp(0.5 * z^2), NormalMeanVariance(0.0,1.0))[2]
+Distributions.mean(dist::ExponentialLinearQuadratic)            = mean_var(dist)[1]
+Distributions.var(dist::ExponentialLinearQuadratic)             = mean_var(dist)[2]
 Distributions.cov(dist::ExponentialLinearQuadratic)             = var(dist)
 
 precision(dist::ExponentialLinearQuadratic)    = inv(var(dist))
