### 2021-09-29 2053d2b
feat(): New rules for NormalMeanVariance
