+++ src/rules/normal_mean_variance/out.jl
@@ -11,3 +11,8 @@ end
 # Variational                       # 
 # --------------------------------- #
 @rule NormalMeanVariance(:out, Marginalisation) (q_μ::Any, q_v::Any) = NormalMeanVariance(mean(q_μ), mean(q_v))
+
+@rule NormalMeanVariance(:out, Marginalisation) (m_μ::UnivariateNormalDistributionsFamily, q_v::Any) = begin 
+    m_μ_mean, m_μ_cov = mean_cov(m_μ)
+    return NormalMeanVariance(m_μ_mean, m_μ_cov + mean(q_v))
+end
