+++ test/rules/normal_mean_precision/test_mean.jl
@@ -21,21 +21,21 @@ import ReactiveMP: @test_rules
     @testset "Belief Propagation: (m_out::UnivariateNormalDistributionsFamily, m_τ::PointMass)" begin
 
         @test_rules [ with_float_conversions = true ] NormalMeanPrecision(:μ, Marginalisation) [
-            (input = (m_out = NormalMeanPrecision(0.0, 1.0),  m_τ = PointMass(2.0)), output = NormalMeanPrecision(0.0, 2.0 / 3.0)),
-            (input = (m_out = NormalMeanPrecision(-1.0, 1.0), m_τ = PointMass(1.5)), output = NormalMeanPrecision(-1.0, 0.6)),
-            (input = (m_out = NormalMeanPrecision(2.0, 0.5),  m_τ = PointMass(1.0)), output = NormalMeanPrecision(2.0, 1.0 / 3.0)),
+            (input = (m_out = NormalMeanPrecision(0.0, 1.0),  m_τ = PointMass(2.0)), output = NormalMeanVariance(0.0, 3.0 / 2.0)),
+            (input = (m_out = NormalMeanPrecision(-1.0, 1.0), m_τ = PointMass(1.5)), output = NormalMeanVariance(-1.0, 10.0 / 6.0)),
+            (input = (m_out = NormalMeanPrecision(2.0, 0.5),  m_τ = PointMass(1.0)), output = NormalMeanVariance(2.0, 3.0 / 1.0)),
         ]
 
         @test_rules [ with_float_conversions = true ] NormalMeanPrecision(:μ, Marginalisation) [
-            (input = (m_out = NormalMeanVariance(0.0, 1.0),  m_τ = PointMass(2.0)), output = NormalMeanPrecision(0.0, 2.0 / 3.0)),
-            (input = (m_out = NormalMeanVariance(-1.0, 1.0), m_τ = PointMass(1.5)), output = NormalMeanPrecision(-1.0, 0.6)),
-            (input = (m_out = NormalMeanVariance(2.0, 0.5),  m_τ = PointMass(1.0)), output = NormalMeanPrecision(2.0, 2.0 / 3.0)),
+            (input = (m_out = NormalMeanVariance(0.0, 1.0),  m_τ = PointMass(2.0)), output = NormalMeanVariance(0.0, 3.0 / 2.0)),
+            (input = (m_out = NormalMeanVariance(-1.0, 1.0), m_τ = PointMass(1.5)), output = NormalMeanVariance(-1.0, 10.0 / 6.0)),
+            (input = (m_out = NormalMeanVariance(2.0, 0.5),  m_τ = PointMass(1.0)), output = NormalMeanVariance(2.0, 3.0 / 2.0)),
         ]
 
         @test_rules [ with_float_conversions = true ] NormalMeanPrecision(:μ, Marginalisation) [
-            (input = (m_out = NormalWeightedMeanPrecision(0.0, 1.0),  m_τ = PointMass(2.0)), output = NormalMeanPrecision(0.0, 2.0 / 3.0)),
-            (input = (m_out = NormalWeightedMeanPrecision(-1.0, 1.0), m_τ = PointMass(1.5)), output = NormalMeanPrecision(-1.0, 0.6)),
-            (input = (m_out = NormalWeightedMeanPrecision(2.0, 0.5),  m_τ = PointMass(1.0)), output = NormalMeanPrecision(4.0, 1.0 / 3.0)),
[…]
