### 2021-09-29 fd5849d
fix(): avoid using Float64 literals in rules
