### 2021-09-29 36d973e
feat(): gaussian extension for HGF structured
