+++ src/rules/gcv/gaussian_extension.jl
@@ -15,7 +15,7 @@ end
     return @call_marginalrule NormalMeanVariance(:out_μ) (m_out = NormalMeanVariance(m_out_mean, m_out_var), m_μ = m_μ, q_v = q_v, meta = meta)
 end
 
-@marginalrule NormalMeanPrecision(:out_μ) (m_out::ExponentialLinearQuadratic, m_μ::NormalMeanVariance, q_τ::PointMass, meta::Any) = begin 
+@marginalrule NormalMeanPrecision(:out_μ) (m_out::ExponentialLinearQuadratic, m_μ::UnivariateNormalDistributionsFamily, q_τ::PointMass, meta::Any) = begin 
     m_out_mean, m_out_var = mean_var(m_out)
     return @call_marginalrule NormalMeanPrecision(:out_μ) (m_out = NormalMeanVariance(m_out_mean, m_out_var), m_μ = m_μ, q_τ = q_τ, meta = meta)
 end
