+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"lib/math/#lib-math","page":"Math utils","title":"Math utilities","text":"","category":"section"},{"location":"lib/math/","page":"Math utils","title":"Math utils","text":"ReactiveMP package exports tiny and huge objects to represent tiny and huge numbers. These objects aren't really numbers and behave differently depending on the context. They do support any operation that is defined for Real numbers. For more info see Julia's documentation about promotion.","category":"page"},{"location":"lib/math/","page":"Math utils","title":"Math utils","text":"tiny\nhuge\nTinyNumber\nHugeNumber","category":"page"},{"location":"lib/math/#ReactiveMP.tiny","page":"Math utils","title":"ReactiveMP.tiny","text":"tiny\n\nAn instance of a TinyNumber. Behaviour and actual value of the tiny number depends on the context.\n\nExample\n\njulia> tiny\ntiny\n\njulia> 1 + tiny\n1.000000000001\n\njulia> tiny + 1\n1.000000000001\n\njulia> 1f0 + tiny\n1.000001f0\n\njulia> big\"1.0\" + tiny\n1.000000000000000000000001\n\njulia> big\"1\" + tiny\n1.000000000000000000000001\n\nSee also: huge, TinyNumber, HugeNumber\n\n\n\n\n\n","category":"constant"},{"location":"lib/math/#ReactiveMP.huge","page":"Math utils","title":"ReactiveMP.huge","text":"huge\n\nAn instance of a HugeNumber. Behaviour and actual value of the huge number depends on the context.\n\nExample\n\njulia> huge\nhuge\n\njulia> 1 + huge\n1.000000000001e12\n\njulia> huge + 1\n1.000000000001e12\n\njulia> 1f0 + huge\n1.000001f6\n\njulia> big\"1.0\" + huge\n1.000000000000000000000001e+24\n\njulia> big\"1\" + huge\n1.000000000000000000000001e+24\n\nSee also: tiny, TinyNumber, HugeNumber\n\n\n\n\n\n","category":"constant"},{"location":"lib/math/#ReactiveMP.TinyNumber","page":"Math utils","title":"ReactiveMP.TinyNumber","text":"TinyNumber <: Real\n\nTinyNumber represents (wow!) tiny number that can be used in a various computations without unnecessary type promotions.\n\nSee also: HugeNumber\n\n\n\n\n\n","category":"type"},{"location":"lib/math/#ReactiveMP.HugeNumber","page":"Math utils","title":"ReactiveMP.HugeNumber","text":"HugeNumber <: Real\n\nHugeNumber represents (wow!) huge number that can be used in a various computations without unnecessary type promotions.\n\nSee also: TinyNumber\n\n\n\n\n\n","category":"type"},{"location":"lib/node/#lib-node","page":"Factor nodes","title":"Nodes implementation","text":"","category":"section"},{"location":"lib/node/","page":"Factor nodes","title":"Factor nodes","text":"In message passing framework one of the most important concept is factor node.  Factor node represents a local function in a factorised representation of a generative model.","category":"page"},{"location":"lib/node/#lib-node-traits","page":"Factor nodes","title":"Node traits","text":"","category":"section"},{"location":"lib/node/","page":"Factor nodes","title":"Factor nodes","text":"Each factor node has to define as_node_functional_form trait function and to specify ValidNodeFunctionalForm singleton as a return object. By default as_node_functional_form returns UndefinedNodeFunctionalForm. Objects that do not specify this property correctly cannot be used in model specification.","category":"page"},{"location":"lib/node/","page":"Factor nodes","title":"Factor nodes","text":"note: Note
[…]
