+++ src/algebra/cholinv.jl
@@ -4,11 +4,11 @@ using LinearAlgebra
 using PositiveFactorizations
 
 cholinv(x)           = inv(cholesky(PositiveFactorizations.Positive, Hermitian(x)))
-cholinv(x::Diagonal) = Diagonal(inv.(diag(x)))
+cholinv(x::Diagonal) = inv(x)
 cholinv(x::Real)     = inv(x)
 
 cholsqrt(x)           = Matrix(cholesky(PositiveFactorizations.Positive, x).L)
-cholsqrt(x::Diagonal) = Diagonal(sqrt.(diag(x)))
+cholsqrt(x::Diagonal) = sqrt(x)
 cholsqrt(x::Real)     = sqrt(x)
 
 
