### 2021-10-05 870bd51
more notes
