### 2021-10-08 #91 @whedon
[REVIEW]: ReactiveMP.jl: A Julia Package for Reactive Message Passing-based Bayesian Inference
**Submitting author:** @bvdmitri (<a href="http://orcid.org/0000-0001-9655-7986">Dmitry Bagaev</a>)
**Repository:** <a href="https://github.com/biaslab/ReactiveMP.jl" target ="_blank">https://github.com/biaslab/ReactiveMP.jl</a>
**Version:** v1.3.1

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
