### 2021-10-09 ed3f031
fixed stuff and added AbstractVector rules
