### 2021-10-11 e9259a1
feat(): Fix type-instabilities in marginal rules
