+++ src/message.jl
@@ -154,7 +154,7 @@ Base.show(io::IO, ::VariationalMessage) = print(io, string("VariationalMessage(:
 getcache(vmessage::VariationalMessage)                    = vmessage.cache
 setcache!(vmessage::VariationalMessage, message::Message) = vmessage.cache = message
 
-compute_message(vmessage::VariationalMessage) = vmessage.mappingFn((vmessage.messages, getrecent(vmessage.marginals)))
+compute_message(vmessage::VariationalMessage) = vmessage.mappingFn((vmessage.messages, vmessage.marginals))
 
 function materialize!(vmessage::VariationalMessage)
     cache = getcache(vmessage)
@@ -177,12 +177,12 @@ as_message(vmessage::VariationalMessage) = materialize!(vmessage)
 reduce_messages(messages) = mapreduce(as_message, (left, right) -> multiply_messages(ProdAnalytical(), left, right), messages)
 
 ## Message Mapping structure
+## https://github.com/JuliaLang/julia/issues/42559
 ## Explanation: Julia cannot fully infer type of the lambda callback function in activate! method in node.jl file
 ## We create a lambda-like callable structure to improve type inference and make it more stable
 ## However it is not fully inferrable due to dynamic tags and variable constraints, but still better than just a raw lambda callback
 
-struct MessageMapping{F, E, T, C, N, M, A, R}
-    fform           :: E
+struct MessageMapping{F, T, C, N, M, A, R}
     vtag            :: T
     vconstraint     :: C
     msgs_names      :: N
@@ -191,20 +191,21 @@ struct MessageMapping{F, E, T, C, N, M, A, R}
     factornode      :: R
 end
 
-message_mapping_fform(m::MessageMapping{F}) where F = F
-message_mapping_fform(m::MessageMapping{F}) where F <: Function = m.fform
+message_mapping_fform(::MessageMapping{F}) where F = F
+message_mapping_fform(::MessageMapping{F}) where F <: Function = F.instance
 
 function MessageMapping(::Type{F}, vtag::T, vconstraint::C, msgs_names::N, marginals_names::M, meta::A, factornode::R) where { F, T, C, N, M, A, R }
-    return MessageMapping{F, Nothing, T, C, N, M, A, R}(nothing, vtag, vconstraint, msgs_names, marginals_names, meta, factornode)
+    return MessageMapping{F, T, C, N, M, A, R}(vtag, vconstraint, msgs_names, marginals_names, meta, factornode)
 end
 
-function MessageMapping(fn::E, vtag::T, vconstraint::C, msgs_names::N, marginals_names::M, meta::A, factornode::R) where { E <: Function, T, C, N, M, A, R} 
-    return MessageMapping{E, E, T, C, N, M, A, R}(fn, vtag, vconstraint, msgs_names, marginals_names, meta, factornode)
+function MessageMapping(::F, vtag::T, vconstraint::C, msgs_names::N, marginals_names::M, meta::A, factornode::R) where { F <: Function, T, C, N, M, A, R} 
+    return MessageMapping{F, T, C, N, M, A, R}(vtag, vconstraint, msgs_names, marginals_names, meta, factornode)
 end
 
 function (mapping::MessageMapping)(dependencies)
+
     messages  = dependencies[1]
-    marginals = dependencies[2]
+    marginals = getrecent(dependencies[2])
 
     # Message is clamped if all of the inputs are clamped
     is_message_clamped = __check_all(is_clamped, messages) && __check_all(is_clamped, marginals)
