+++ docs/src/examples/overview.md
@@ -1,0 +1,8 @@
+# [Examples overview](@id examples-overview)
+
+This section contains a set of examples for Bayesian Inference with `ReactiveMP` package in various probabilistic models.
+
+!!! note
+    This section is WIP and more examples will be added over time.
+
+- [Gaussian Linear Dynamical System](@ref examples-linear-gaussian-state-space-model): An example of inference procedure for Gaussian Linear Dynamical System with multivariate noisy observations using Belief Propagation (Sum Product) algorithm. Reference: [Simo Sarkka, Bayesian Filtering and Smoothing](https://users.aalto.fi/~ssarkka/pub/cup_book_online_20131111.pdf)
