### 2021-10-11 d170b53
fix(): Remove unnecessary dot product rules
