+++ src/rules/dot_product/out.jl
@@ -1,26 +1,10 @@
-export rule
-import LinearAlgebra: dot
 
-@rule typeof(dot)(:out, Marginalisation, symmetrical = [:in1, :in2]) (m_in1::PointMass{ <: AbstractVector }, m_in2::NormalDistributionsFamily) = begin
-    NormalMeanVariance(mean(m_in1)'*mean(m_in2), first(mean(m_in1)'*cov(m_in2)*mean(m_in1)))
+@rule typeof(dot)(:out, Marginalisation) (m_in1::NormalDistributionsFamily, m_in2::PointMass, meta::AbstractCorrection) = begin 
+    return @call_rule typeof(dot)(:out, Marginalisation) (m_in1 = m_in2, m_in2 = m_in1, meta = meta)
 end
 
-@rule typeof(dot)(:out, Marginalisation, symmetrical = [:in1, :in2]) (m_in1::NormalDistributionsFamily, m_in2::PointMass{ <: AbstractVector }) = begin
-    NormalMeanVariance(mean(m_in2)'*mean(m_in1), first(mean(m_in2)'*cov(m_in1)*mean(m_in2)))
-end
-
-@rule typeof(dot)(:out, Marginalisation) (m_in1::PointMass{ <: AbstractVector }, m_in2::NormalDistributionsFamily) = begin
-    MeanVariance(mean(m_in1)'*mean(m_in2), first(mean(m_in1)'*cov(m_in2)*mean(m_in1)))
-end
-
-@rule typeof(dot)(:out, Marginalisation) (m_in1::NormalDistributionsFamily, m_in2::PointMass{ <: AbstractVector }) = begin
-    NormalMeanVariance(mean(m_in2)'*mean(m_in1), first(mean(m_in2)'*cov(m_in1)*mean(m_in2)))
-end
-
-@rule typeof(dot)(:out, Marginalisation) (m_in1::PointMass, m_in2::NormalDistributionsFamily) = begin
-    NormalMeanVariance(mean(m_in1)'*mean(m_in2), first(mean(m_in1)'*cov(m_in2)*mean(m_in1)))
-end
-
-@rule typeof(dot)(:out, Marginalisation) (m_in1::NormalDistributionsFamily, m_in2::PointMass) = begin
-    NormalMeanVariance(mean(m_in2)'*mean(m_in1), first(mean(m_in2)'*cov(m_in1)*mean(m_in2)))
+@rule typeof(dot)(:out, Marginalisation) (m_in1::PointMass, m_in2::NormalDistributionsFamily, meta::AbstractCorrection) = begin 
+    A = mean(m_in1)
+    in2_mean, in2_cov = mean_cov(m_in2)
+    return NormalMeanVariance(dot(A, in2_mean), A' * in2_cov * A)
 end
