+++ src/nodes/dot_product.jl
@@ -1,5 +1,3 @@
-export make_node
-export dot
 import LinearAlgebra: dot
 
 @node typeof(dot) Deterministic [ out, in1, in2 ]
