+++ src/rules/dot_product/marginals.jl
@@ -1,29 +1,14 @@
-export marginalrule
 
-@marginalrule typeof(dot)(:in1_in2) (m_out::UnivariateNormalDistributionsFamily, m_in1::PointMass{ <: AbstractVector }, m_in2::MultivariateNormalDistributionsFamily) = begin
-    m, V = mean(m_out), cov(m_out)
-    x = mean(m_in1)
-    q_in2 = prod(ProdAnalytical(), m_in2, MvNormalWeightedMeanPrecision(x * weightedmean(m_out), x * precision(m_out) * x'))
+@marginalrule typeof(dot)(:in1_in2) (m_out::NormalDistributionsFamily, m_in1::PointMass, m_in2::NormalDistributionsFamily, meta::AbstractCorrection) = begin
+
+    # Forward message towards `in2` edge
+    mf_in2 = @call_rule typeof(dot)(:in2, Marginalisation) (m_out = m_out, m_in1 = m_in1, meta = meta)
+    q_in2  = prod(ProdAnalytical(), m_in2, mf_in2)
+
     return (in1 = m_in1, in2 = q_in2)
 end
 
-@marginalrule typeof(dot)(:in1_in2) (m_out::UnivariateNormalDistributionsFamily, m_in1::PointMass, m_in2::UnivariateNormalDistributionsFamily) = begin
-    m, V = mean(m_out), cov(m_out)
-    x = mean(m_in1)
-    q_in2 = prod(ProdAnalytical(), m_in2, NormalWeightedMeanPrecision(x * weightedmean(m_out), x * precision(m_out) * x'))
-    return (in1 = m_in1, in2 = q_in2)
-end
-
-@marginalrule typeof(dot)(:in1_in2) (m_out::UnivariateNormalDistributionsFamily, m_in1::MultivariateNormalDistributionsFamily, m_in2::PointMass{ <: AbstractVector }) = begin
-    m, V = mean(m_out), cov(m_out)
-    x = mean(m_in2)
-    q_in2 = prod(ProdAnalytical(), m_in1, MvNormalWeightedMeanPrecision(x * weightedmean(m_out), x * precision(m_out) * x'))
-    return (in1 = m_in2, in2 = q_in2)
-end
-
-@marginalrule typeof(dot)(:in1_in2) (m_out::UnivariateNormalDistributionsFamily, m_in1::UnivariateNormalDistributionsFamily, m_in2::PointMass) = begin
-    m, V = mean(m_out), cov(m_out)
-    x = mean(m_in2)
-    q_in2 = prod(ProdAnalytical(), m_in1, NormalWeightedMeanPrecision(x * weightedmean(m_out), x * precision(m_out) * x'))
-    return (in1 = m_in2, in2 = q_in2)
+@marginalrule typeof(dot)(:in1_in2) (m_out::NormalDistributionsFamily, m_in1::NormalDistributionsFamily, m_in2::PointMass, meta::AbstractCorrection) = begin
+    symmetric = @call_marginalrule typeof(dot)(:in1_in2) (m_out = m_out, m_in1 = m_in2, m_in2 = m_in1, meta = meta)
+    return (in1 = symmetric[:in2], in2 = symmetric[:in1])
 end
