+++ src/rules/dot_product/in2.jl
@@ -1,34 +1,11 @@
-export rule
-
-@rule typeof(dot)(:in2, Marginalisation) (m_out::UnivariateNormalDistributionsFamily, m_in1::PointMass{ <: AbstractVector }, meta::AbstractCorrection) = begin
-    x  = mean(m_in1)
-    xi, w = weightedmean_precision(m_out)
-    xi = x * xi
-    w  = correction!(meta, x * w * x')
-    return MvNormalWeightedMeanPrecision(xi, w)
-end
 
 @rule typeof(dot)(:in2, Marginalisation) (m_out::UnivariateNormalDistributionsFamily, m_in1::PointMass, meta::AbstractCorrection) = begin
-    x  = mean(m_in1)
-    xi, w = weightedmean_precision(m_out)
-    xi = x * xi
-    w  = correction!(meta, x * w * x')
-    return NormalWeightedMeanPrecision(xi, w)
-end
-
-@rule typeof(dot)(:in2, Marginalisation) (m_out::PointMass{ <: AbstractVector }, m_in1::UnivariateNormalDistributionsFamily, meta::AbstractCorrection) = begin
-    x  = mean(m_out)
-    xi, w = weightedmean_precision(m_in1)
-    xi = x * xi
-    w  = correction!(meta, x * w * x')
-    return MvNormalWeightedMeanPrecision(xi, w)
-end
-
-@rule typeof(dot)(:in2, Marginalisation) (m_out::PointMass, m_in1::UnivariateNormalDistributionsFamily, meta::AbstractCorrection) = begin
-    x  = mean(m_out)
-    xi, w = weightedmean_precision(m_in1)
-    xi = x * xi
-    w  = correction!(meta, x * w * x')
-    return NormalWeightedMeanPrecision(xi, w)
+    A  = mean(m_in1)
+    out_wmean, out_prec = weightedmean_precision(m_out)
+    
+    ξ = A * out_wmean
+    W = correction!(meta, A * out_prec * A')
+    
+    return convert(promote_variate_type(variate_form(m_in1), NormalWeightedMeanPrecision), ξ, W)
 end
 

[…]
