+++ test/rules/dot_product/test_in1.jl
@@ -3,6 +3,7 @@ module RulesDotProductIn1Test
 using Test
 using ReactiveMP
 using Random
+
 import ReactiveMP: @test_rules
 import LinearAlgebra: dot
 
@@ -10,41 +11,45 @@ import LinearAlgebra: dot
 
      @testset "Belief Propagation: (m_out::UnivariateNormalDistributionsFamily, m_in2::PointMass)" begin
 
-        @test_rules [ with_float_conversions = false ] typeof(dot)(:in1, Marginalisation) [
+        @test_rules [ with_float_conversions = true ] typeof(dot)(:in1, Marginalisation) [
+            (input = (m_out = NormalMeanVariance(2.0, 2.0), m_in2 = PointMass(-1.0), meta = NoCorrection()), output = NormalWeightedMeanPrecision(-1.0, 0.5)),
+            (input = (m_out = NormalMeanPrecision(1.0, 1.0), m_in2 = PointMass(-2.0), meta = NoCorrection()), output = NormalWeightedMeanPrecision(-2.0, 4.0)),
+            (input = (m_out = NormalWeightedMeanPrecision(2.0, 1.0), m_in2 = PointMass(-1.0), meta = NoCorrection()), output = NormalWeightedMeanPrecision(-2.0, 1.0)),
+        ]
+
+        @test_rules [ with_float_conversions = true ] typeof(dot)(:in1, Marginalisation) [
             (input = (m_out = NormalMeanVariance(2.0, 2.0), m_in2 = PointMass(-1.0), meta = TinyCorrection()), output = NormalWeightedMeanPrecision(-1.0, 0.5)),
-            (input = (m_out = NormalMeanVariance(1.0, 1.0), m_in2 = PointMass(-2.0), meta = TinyCorrection()), output = NormalWeightedMeanPrecision(-2.0, 4.0)),
-            (input = (m_out = NormalMeanVariance(2.0, 1.0), m_in2 = PointMass(-1.0), meta = TinyCorrection()), output = NormalWeightedMeanPrecision(-2.0, 1.0)),
+            (input = (m_out = NormalMeanPrecision(1.0, 1.0), m_in2 = PointMass(-2.0), meta = TinyCorrection()), output = NormalWeightedMeanPrecision(-2.0, 4.0)),
+            (input = (m_out = NormalWeightedMeanPrecision(2.0, 1.0), m_in2 = PointMass(-1.0), meta = TinyCorrection()), output = NormalWeightedMeanPrecision(-2.0, 1.0)),
+        ]
+
+        @test_rules [ with_float_conversions = true ] typeof(dot)(:in1, Marginalisation) [
+            (input = (m_out = NormalMeanVariance(2.0, 2.0), m_in2 = PointMass(0.0), meta = NoCorrection()), output = NormalWeightedMeanPrecision(0.0, 0.0)),
+            (input = (m_out = NormalMeanPrecision(1.0, 1.0), m_in2 = PointMass(0.0), meta = NoCorrection()), output = NormalWeightedMeanPrecision(0.0, 0.0)),
+            (input = (m_out = NormalWeightedMeanPrecision(2.0, 1.0), m_in2 = PointMass(0.0), meta = NoCorrection()), output = NormalWeightedMeanPrecision(0.0, 0.0)),
+        ]
+
+        @test_rules [ with_float_conversions = true ] typeof(dot)(:in1, Marginalisation) [
[…]
