+++ test/rules/dot_product/test_out.jl
@@ -3,6 +3,7 @@ module RulesDotProductOutTest
 using Test
 using ReactiveMP
 using Random
+
 import ReactiveMP: @test_rules
 import LinearAlgebra: dot
 
@@ -11,9 +12,15 @@ import LinearAlgebra: dot
      @testset "Belief Propagation: (m_in1::PointMass, m_in2::NormalDistributionsFamily)" begin
 
         @test_rules [ with_float_conversions = true ] typeof(dot)(:out, Marginalisation) [
-            (input = (m_in1 = PointMass(1.0), m_in2 = NormalMeanVariance(2.0, 2.0)), output = NormalMeanVariance(2.0, 2.0)),
-            (input = (m_in1 = PointMass(-1.0), m_in2 = NormalMeanVariance(3.0, 1.0)), output = NormalMeanVariance(-3.0, 1.0)),
-            (input = (m_in1 = PointMass(2.0), m_in2 = NormalMeanVariance(4.0, 2.0)), output = NormalMeanVariance(8.0, 8.0)),
+            (input = (m_in1 = PointMass(1.0), m_in2 = NormalMeanVariance(2.0, 2.0), meta = NoCorrection()), output = NormalMeanVariance(2.0, 2.0)),
+            (input = (m_in1 = PointMass(-1.0), m_in2 = NormalMeanPrecision(3.0, 1.0), meta = NoCorrection()), output = NormalMeanVariance(-3.0, 1.0)),
+            (input = (m_in1 = PointMass(2.0), m_in2 = NormalWeightedMeanPrecision(2.0, 0.5), meta = NoCorrection()), output = NormalMeanVariance(8.0, 8.0)),
+        ]
+
+        @test_rules [ with_float_conversions = true ] typeof(dot)(:out, Marginalisation) [
+            (input = (m_in1 = PointMass(1.0), m_in2 = NormalMeanVariance(2.0, 2.0), meta = TinyCorrection()), output = NormalMeanVariance(2.0, 2.0)),
+            (input = (m_in1 = PointMass(-1.0), m_in2 = NormalMeanPrecision(3.0, 1.0), meta = TinyCorrection()), output = NormalMeanVariance(-3.0, 1.0)),
+            (input = (m_in1 = PointMass(2.0), m_in2 = NormalWeightedMeanPrecision(2.0, 0.5), meta = TinyCorrection()), output = NormalMeanVariance(8.0, 8.0)),
         ]
 
     end
@@ -21,29 +28,35 @@ import LinearAlgebra: dot
     @testset "Belief Propagation: (m_in1::NormalDistributionsFamily, m_in2::PointMass)" begin
 
         @test_rules [ with_float_conversions = true ] typeof(dot)(:out, Marginalisation) [
-            (input = (m_in1 = NormalMeanVariance(2.0, 2.0), m_in2 = PointMass(4.0)), output = NormalMeanVariance(8.0, 32.0)),
-            (input = (m_in1 = NormalMeanVariance(2.0, 3.0), m_in2 = PointMass(2.0)), output = NormalMeanVariance(4.0, 12.0)),
-            (input = (m_in1 = NormalMeanVariance(4.0, 2.0), m_in2 = PointMass(1.0)), output = NormalMeanVariance(4.0, 2.0)),
+            (input = (m_in1 = NormalMeanVariance(2.0, 2.0), m_in2 = PointMass(4.0), meta = NoCorrection()), output = NormalMeanVariance(8.0, 32.0)),
[…]

[…]
