+++ test/rules/dot_product/test_marginals.jl
@@ -68,4 +68,5 @@ import LinearAlgebra: dot
     end
 
 end
+
 end
