+++ test/rules/dot_product/test_in1.jl
@@ -53,3 +53,5 @@ import LinearAlgebra: dot
     end
 
 end
+
+end
