+++ test/rules/dot_product/test_out.jl
@@ -62,4 +62,5 @@ import LinearAlgebra: dot
     end
 
 end
+
 end
