+++ test/rules/dot_product/test_in2.jl
@@ -53,4 +53,5 @@ import LinearAlgebra: dot
     end
 
 end
+
 end
