### 2021-10-12 ea1f9db
feat(): VMP rule for gmp node, fix for marginal rule
