### 2021-10-12 525714f
feat(): interfaceindices function for abstract factor node
