### 2021-10-12 f4afe53
feat(): update HGF rule specifications
