### 2021-10-13 0432961
feat(): tests and extra methods for exp lin quad distribution
