+++ src/distributions/exp_linear_quadratic.jl
@@ -1,7 +1,7 @@
 export ExponentialLinearQuadratic
 
 import Distributions: pdf, logpdf, ContinuousUnivariateDistribution
-
+import Base: eltype
 struct ExponentialLinearQuadratic{ A <: AbstractApproximationMethod, T <: Real } <: ContinuousUnivariateDistribution
     approximation :: A
     a :: T
@@ -10,15 +10,30 @@ struct ExponentialLinearQuadratic{ A <: AbstractApproximationMethod, T <: Real }
     d :: T
 end
 
+Base.eltype(::Type{ <: ExponentialLinearQuadratic{A, T} }) where { A, T } = T
+Base.eltype(::ExponentialLinearQuadratic{A, T})            where { A, T } = T
+
 ExponentialLinearQuadratic(approximation, a::Real, b::Real, c::Real, d::Real)             = ExponentialLinearQuadratic(approximation, promote(a, b, c, d)...)
 ExponentialLinearQuadratic(approximation, a::Integer, b::Integer, c::Integer, d::Integer) = ExponentialLinearQuadratic(approximation, float(a), float(b), float(c), float(d))
 
 mean_cov(dist::ExponentialLinearQuadratic) = mean_var(dist)
-mean_var(dist::ExponentialLinearQuadratic) = approximate_meancov(dist.approximation, (z) -> pdf(dist, z) * exp(0.5 * z^2), NormalMeanVariance(0.0, 1.0))
 
-function mean_invcov(dist::ExponentialLinearQuadratic) 
+function mean_var(dist::ExponentialLinearQuadratic) 
+    # This is equivalent to (x) -> pdf(dist, x) * exp(x^2 / 2)
+    adjusted_pdf = let a = dist.a, b = dist.b, c = dist.c, d = dist.d
+        (x) -> exp(-(a * x - x^2 + b * exp(c * x + d * x^2 / 2)) / 2)
+    end
+    return approximate_meancov(dist.approximation, adjusted_pdf, NormalMeanVariance(zero(eltype(dist)), one(eltype(dist))))
+end
+
+mean_invcov(dist::ExponentialLinearQuadratic)      = mean_cov(dist) .|> (identity, inv)
+mean_std(dist::ExponentialLinearQuadratic)         = mean_var(dist) .|> (identity, sqrt)
+weightedmean_cov(dist::ExponentialLinearQuadratic) = weightedmean_var(dist)
+weightedmean_std(dist::ExponentialLinearQuadratic) = weightedmean_var(dist) .|> (identity, sqrt)
+
+function weightedmean_var(dist::ExponentialLinearQuadratic)
     m, v = mean_cov(dist)
-    return (m, inv(v))
+    return (inv(v) * m, v)
 end
 
 function weightedmean_invcov(dist::ExponentialLinearQuadratic)
@@ -27,18 +42,26 @@ function weightedmean_invcov(dist::ExponentialLinearQuadratic)
 end
 
 Distributions.pdf(dist::ExponentialLinearQuadratic, x::Real)    = exp(logpdf(dist, x))
-Distributions.logpdf(dist::ExponentialLinearQuadratic, x::Real) = -0.5 * (dist.a * x + dist.b * exp(dist.c * x + dist.d * x ^ 2 / 2.0))
+Distributions.logpdf(dist::ExponentialLinearQuadratic, x::Real) = -(dist.a * x + dist.b * exp(dist.c * x + dist.d * x ^ 2 / 2)) / 2
 Distributions.mean(dist::ExponentialLinearQuadratic)            = mean_var(dist)[1]
 Distributions.var(dist::ExponentialLinearQuadratic)             = mean_var(dist)[2]
[…]
