+++ src/rules/normal_mean_variance/marginals.jl
@@ -11,7 +11,7 @@ end
     xi_out, W_out = weightedmean_precision(m_out)
     xi_μ, W_μ     = weightedmean_precision(m_μ)
 
-    W_bar = inv(mean(q_v))
+    W_bar = inv(mean(m_v))
 
     W  = [ W_out + W_bar -W_bar; -W_bar W_μ + W_bar ]
     xi = [ xi_out; xi_μ ]
