+++ docs/src/examples/hierarchical_gaussian_filter.md
@@ -1,0 +1,225 @@
+# [Example: Hierarchical Gaussian Filter](@id examples-hgf)
+
+In this demo the goal is to perform approximate variational Bayesian Inference for Univariate Hierarchical Gaussian Filter (HGF).
+
+Simple HGF model can be defined as:
+
+```math
+\begin{equation}
+  \begin{aligned}
+    x^{(j)}_k & \sim \, \mathcal{N}(x^{(j)}_{k - 1}, f_k(x^{(j - 1)}_k)) \\
+    y_k & \sim \, \mathcal{N}(x^{(j)}_k, \tau_k)
+  \end{aligned}
+\end{equation}
+```
+
+where ``j`` is an index of layer in hierarchy, ``k`` is a time step and ``f_k`` is a variance activation function. `ReactiveMP.jl` export Gaussian Controlled Variance (GCV) node with ``f_k = \exp(\kappa x + \omega)`` variance activation function. By default uses Gauss-Hermite cubature with a prespecified number of approximation points in the cubature. We can change the number of points in Gauss-Hermite cubature with the help of metadata structures in `ReactiveMP.jl`. 
+
+```math
+\begin{equation}
+  \begin{aligned}
+    z_k & \sim \, \mathcal{N}(z_{k - 1}, \mathcal{\tau_z}) \\
+    x_k & \sim \, \mathcal{N}(x_{k - 1}, \exp(\kappa z_k + \omega)) \\
+    y_k & \sim \, \mathcal{N}(x_k, \mathcal{\tau_y})
+  \end{aligned}
+\end{equation}
+```
+
+In this experiment we will create a single time step of the graph and perform variational message passing filtering alrogithm to estimate hidden states of the system.
+For a more rigorous introduction to Hierarchical Gaussian Filter we refer to [Ismail Senoz, Online Message Passing-based Inference in the Hierarchical Gaussian Filter](https://ieeexplore.ieee.org/document/9173980) paper.
+
+For simplicity we will consider ``\tau_z``, ``\tau_y``, ``\kappa`` and ``\omega`` known and fixed.
+
+To model this process in `ReactiveMP`, first, we start with importing all needed packages:
+
+```@example hgf
+using Rocket, ReactiveMP, GraphPPL, Distributions
+using BenchmarkTools, Random, Plots
+```
+
+Next step, is to generate some synthetic data:
+
+```@example hgf
+function generate_data(rng, k, w, zv, yv)
+    z_prev = 0.0
+    x_prev = 0.0
+
+    z = Vector{Float64}(undef, n)
+    v = Vector{Float64}(undef, n)
+    x = Vector{Float64}(undef, n)
+    y = Vector{Float64}(undef, n)
+
+    for i in 1:n
+        z[i] = rand(rng, Normal(z_prev, sqrt(zv)))
+        v[i] = exp(k * z[i] + w)
+        x[i] = rand(rng, Normal(x_prev, sqrt(v[i])))
+        y[i] = rand(rng, Normal(x[i], sqrt(yv)))
+
+        z_prev = z[i]
+        x_prev = x[i]
+    end 
+    
+    return z, x, y
+end
+```
+
+```@example hgf
+# Seed for reproducibility
+seed = 123
+
+rng = MersenneTwister(seed)
+
+# Parameters of HGF process
+real_k = 1.0
+real_w = 0.0
+z_variance = abs2(0.5)
+y_variance = abs2(1.0)
+
+# Number of observations
+n = 300
+
+z, x, y = generate_data(rng, real_k, real_w, z_variance, y_variance)
+nothing #hide
+```
+
+Lets plot our synthetic dataset. Lines represent our hidden states we want to estimate using noisy observations.
+
+```@example hgf
+let 
+    pz = plot(title = "Hidden States Z")
[…]
