+++ versions.js
@@ -3,3 +3,5 @@ var DOC_VERSIONS = [
   "v1.0",
   "dev",
 ];
+var DOCUMENTER_NEWEST = "v1.0.5";
+var DOCUMENTER_STABLE = "stable";
