### 2021-10-14 9a8141c
feat(): more efficient AR rules
