+++ src/rules/autoregressive/marginals.jl
@@ -1,23 +1,26 @@
-export marginalrule
 
 @marginalrule AR(:y_x) (m_y::NormalDistributionsFamily, m_x::NormalDistributionsFamily, q_θ::NormalDistributionsFamily, q_γ::GammaShapeRate, meta::ARMeta) = begin
     return ar_y_x_marginal(getstype(meta), m_y, m_x, q_θ, q_γ, meta)
 end
 
 function ar_y_x_marginal(::ARsafe, m_y::NormalDistributionsFamily, m_x::NormalDistributionsFamily, q_θ::NormalDistributionsFamily, q_γ::GammaShapeRate, meta::ARMeta)
-    mθ, Vθ = mean(q_θ), cov(q_θ)
+    mθ, Vθ = mean_cov(q_θ)
+    mγ = mean(q_γ)
 
     mA = as_companion_matrix(mθ)
-    mW = ar_precision(getvform(meta), getorder(meta), mean(q_γ))
+    mW = ar_precision(getvform(meta), getorder(meta), mγ)
 
-    b_my, b_Vy = mean(m_y), cov(m_y)
-    f_mx, f_Vx = mean(m_x), cov(m_x)
+    b_my, b_Vy = mean_cov(m_y)
+    f_mx, f_Vx = mean_cov(m_x)
 
-    D = cholinv(f_Vx) + mean(q_γ) * Vθ
-    W = [cholinv(b_Vy)+mW -mW*mA; -(mA'*mW) D+mA'*mW*mA]
-    m = cholinv(W)*[cholinv(b_Vy)*b_my; cholinv(f_Vx)*f_mx]
+    inv_b_Vy = cholinv(b_Vy)
+    inv_f_Vx = cholinv(f_Vx)
 
-    return MvNormalMeanPrecision(m, W)
+    D = inv_f_Vx + mγ * Vθ
+    W = [inv_b_Vy+mW -mW*mA; -(mA'*mW) D+mA'*mW*mA]
+    ξ = [inv_b_Vy * b_my; inv_f_Vx * f_mx]
+
+    return MvNormalWeightedMeanPrecision(ξ, W)
 end
 
 function ar_y_x_marginal(::ARunsafe, m_y::NormalDistributionsFamily, m_x::NormalDistributionsFamily, q_θ::NormalDistributionsFamily, q_γ::GammaShapeRate, meta::ARMeta)
