### 2021-10-14 e565b9a
feat(): fixed memory allocation in ar_slice
