+++ src/rules/autoregressive/theta.jl
@@ -5,16 +5,15 @@
 
     myx, Vyx = mean_cov(q_y_x)
     my, Vy   = ar_slice(F, myx, 1:order), ar_slice(F, Vyx, 1:order, 1:order)
-    mx, Vx   = ar_slice(F, myx, order+1:2order), ar_slice(F, Vyx, order+1:2order, order+1:2order)
-    Vyx      = ar_slice(F, Vyx, order+1:2order, 1:order)
+    mx, Vx   = ar_slice(F, myx, (order + 1):2order), ar_slice(F, Vyx, (order + 1):2order, (order + 1):2order)
+    Vyx      = ar_slice(F, Vyx, (order + 1):2order, 1:order)
 
     mγ = mean(q_γ)
 
-    D = mγ * (Vx + mx * mx')
+    W = mγ * (Vx + mx * mx')
     c = ar_unit(getvform(meta), order)
 
-    Vθ = cholinv(D)
-    mθ = Vθ * (Vyx + mx * my') * mγ * c
+    ξ = (Vyx + mx * my') * mγ * c
     
-    return convert(promote_variate_type(F, NormalMeanVariance), mθ, Vθ)
+    return convert(promote_variate_type(F, NormalWeightedMeanPrecision), ξ, W)
 end
