### 2021-10-14 908c287
feat(): lazy average energy correction
