### 2021-10-14 ea8903a
feat(): more performance improvements for AR rules
