### 2021-10-15 7a54150
feat(): use BLAS in autoregressive rules
