### 2021-10-15 216e277
project(): add compat entry for LazyArrays.jl
