+++ Project.toml
@@ -31,6 +31,7 @@ DomainIntegrals = "0.3.2"
 DomainSets = "0.5.2"
 FastGaussQuadrature = "0.4"
 ForwardDiff = "0.10"
+LazyArrays = "0.21, 0.22"
 LoopVectorization = "0.12"
 MacroTools = "0.5"
 Optim = "1.0.0"
