+++ docs/make.jl
@@ -22,6 +22,7 @@ makedocs(
             "Overview"                         => "examples/overview.md",
             "Linear Gaussian Dynamical System" => "examples/linear_gaussian_state_space_model.md",
             "Hierarchical Gaussian Filter"     => "examples/hierarchical_gaussian_filter.md",
+            "Autoregressive Model"             => "examples/autoregressive.md",
         ],
         "Library" => [
             "Messages"     => "lib/message.md",
