+++ docs/src/examples/autoregressive.md
@@ -1,0 +1,280 @@
+# [Example: Autoregressive model](@id examples-autoregressive)
+
+In this example we are going to perform an automated Variational Bayesian Inference for autoregressive model that can be represented as following:
+
+```math
+\begin{aligned}
+p(\gamma) &= \mathrm{Gamma}(\gamma|a, b),\\
+p(\mathbf{\theta}) &= \mathcal{N}(\mathbf{\theta}|\mathbf{\mu}, \Sigma),\\
+p(x_t|\mathbf{x}_{t-1:t-k}) &= \mathcal{N}(x_t|\mathbf{\theta}^{T}\mathbf{x}_{t-1:t-k}, \gamma^{-1}),\\
+p(y_t|x_t) &= \mathcal{N}(y_t|x_t, \tau^{-1}),
+\end{aligned}
+```
+
+where ``x_t`` is a current state of our system, ``\mathbf{x}_{t-1:t-k}`` is a sequence of ``k`` previous states, ``k`` is an order of autoregression process, ``\mathbf{\theta}`` is a vector of transition coefficients, ``\gamma`` is a precision of state transition process, ``y_k`` is a noisy observation of ``x_k`` with precision ``\tau``.
+
+For a more rigorous introduction to Bayesian inference in Autoregressive models we refer to [Albert Podusenko, Message Passing-Based Inference for Time-Varying Autoregressive Models](https://www.mdpi.com/1099-4300/23/6/683).
+
+We start with importing all needed packages:
+
+```@example ar
+using Rocket, ReactiveMP, GraphPPL
+using Distributions, LinearAlgebra, Random, Plots, BenchmarkTools
+```
+
+Lets generate some synthetic dataset, we use a predefined set of coeffcients for $k$ = 5:
+
+```@example ar
+# The following coefficients correspond to stable poles
+coefs_ar_5 = [0.10699399235785655, -0.5237303489793305, 0.3068897071844715, -0.17232255282458891, 0.13323964347539288]
+nothing #hide
+```
+
+```@example ar
+function generate_ar_data(rng, n, θ, γ, τ)
+    order        = length(θ)
+    states       = Vector{Vector{Float64}}(undef, n + 3order)
+    observations = Vector{Float64}(undef, n + 3order)
+    
+    γ_std = sqrt(inv(γ))
+    τ_std = sqrt(inv(γ))
+    
+    states[1] = randn(rng, order)
+    
+    for i in 2:(n + 3order)
+        states[i]       = vcat(rand(rng, Normal(dot(θ, states[i - 1]), γ_std)), states[i-1][1:end-1])
+        observations[i] = rand(rng, Normal(states[i][1], τ_std))
+    end
+    
+    return states[1+3order:end], observations[1+3order:end]
+end
+```
+
+```@example ar
+# Seed for reproducibility
+seed = 123
+rng  = MersenneTwister(seed)
+
+# Number of observations in synthetic dataset
+n = 500
+
+# AR process parameters
+real_γ = 5.0
+real_τ = 5.0
+real_θ = coefs_ar_5
+
+states, observations = generate_ar_data(rng, n, real_θ, real_γ, real_τ)
+nothing #hide
+```
+
+Lets plot our synthetic dataset:
+
+```@example ar
+plot(first.(states), label = "Hidden states")
+scatter!(observations, label = "Observations")
+```
+
[…]
