### 2021-10-19 348955d
fix type stability nr_params
