### 2021-10-20 a0eae6b
fix(): Fix algebra helpers for static arrays
