### 2021-10-20 1c47299
fix(): Bug fix for infinite recursion in structured VMP
