+++ src/node.jl
@@ -685,6 +685,7 @@ function getmarginal!(factornode::FactorNode, localmarginal::FactorNodeLocalMarg
         return as_marginal_observable(vmarginal, skip_strategy)
     else
         cmarginal = MarginalObservable()
+        setstream!(localmarginal, cmarginal)
 
         message_dependencies  = tuple(getclusterinterfaces(factornode, clusterindex)...)
         marginal_dependencies = tuple(TupleTools.deleteat(localmarginals(factornode), clusterindex)...)
@@ -701,8 +702,6 @@ function getmarginal!(factornode::FactorNode, localmarginal::FactorNodeLocalMarg
         marginalout = combineLatest((msgs_observable, marginals_observable), PushNew()) |> discontinue() |> map(Marginal, mapping)
 
         connect!(cmarginal, marginalout) # MarginalObservable has RecentSubject by default, there is no need to share_recent() here
-
-        setstream!(localmarginal, cmarginal)
 
         return as_marginal_observable(cmarginal, skip_strategy)
     end
