### 2021-10-21 c52975d
experimental(): eq nodes
