+++ src/variable.jl
@@ -2,6 +2,7 @@ export AbstractVariable, degree
 export is_clamped, is_marginalisation, is_moment_matching
 export FoldLeftProdStrategy, FoldRightProdStrategy, CustomProdStrategy
 export getmarginal, getmarginals, setmarginal!, setmarginals!, name, as_variable
+export setmessage!, setmessages!
 
 using Rocket
 
@@ -58,24 +59,44 @@ end
 getmarginals(variables::AbstractArray{ <: AbstractVariable })                                      = getmarginals(variables, SkipInitial())
 getmarginals(variables::AbstractArray{ <: AbstractVariable }, skip_strategy::MarginalSkipStrategy) = collectLatest(map(v -> getmarginal(v, skip_strategy), variables))
 
-# Setters
+## Setters
 
-function setmarginal!(variable::AbstractVariable, marginal)
-    setmarginal!(getmarginal(variable, IncludeAll()), marginal)
-end
+### Marginals
+
+setmarginal!(variable::AbstractVariable, marginal) = setmarginal!(getmarginal(variable, IncludeAll()), marginal)
 
 setmarginals!(variables::AbstractArray{ <: AbstractVariable }, marginal::Distribution)    = _setmarginals!(Base.HasLength(), variables, Iterators.repeated(marginal, length(variables)))
 setmarginals!(variables::AbstractArray{ <: AbstractVariable }, marginals)                 = _setmarginals!(Base.IteratorSize(marginals), variables, marginals)
 
 function _setmarginals!(::Base.IteratorSize, variables::AbstractArray{ <: AbstractVariable }, marginals)
     foreach(zip(variables, marginals)) do (variable, marginal)
-        setmarginal!(getmarginal(variable, IncludeAll()), marginal)
+        setmarginal!(variable, marginal)
     end
 end
 
 function _setmarginals!(::Any, variables::AbstractArray{ <: AbstractVariable }, marginals)
     error("setmarginals!() failed. Default value is neither an iterable object nor a distribution.")
 end
+
+### Messages
+
+setmessage!(variable::AbstractVariable, index::Int, message) = setmessage!(messageout(variable, index), message)
+setmessage!(variable::AbstractVariable, message)             = foreach(i -> setmessage!(variable, i, message), 1:degree(variable))
+
+setmessages!(variables::AbstractArray{ <: AbstractVariable }, message::Distribution)    = _setmessages!(Base.HasLength(), variables, Iterators.repeated(message, length(variables)))
+setmessages!(variables::AbstractArray{ <: AbstractVariable }, messages)                 = _setmessages!(Base.IteratorSize(messages), variables, messages)
+
+function _setmessages!(::Base.IteratorSize, variables::AbstractArray{ <: AbstractVariable }, messages)
+    foreach(zip(variables, messages)) do (variable, message)
+        setmessage!(variable, message)
+    end
+end
+
+function _setmessages!(::Any, variables::AbstractArray{ <: AbstractVariable }, marginals)
+    error("setmessages!() failed. Default value is neither an iterable object nor a distribution.")
+end
+
+##
 
 function name(variable::AbstractVariable)
     return variable.name
