### 2021-10-22 8794d18
experimental(): equality nodes chain
