### 2021-10-22 eb39821
experimental(): fix empirical equality chain size
