### 2021-10-22 7bf25da
fix(): bug fix with share()
