+++ src/nodes/equality.jl
@@ -4,68 +4,105 @@ import Base: length
 ## Equality node is a special case and has a special implementation
 ## It should not be used during model creation but instead is a part of variable node implementation
 
-## Equality node
+struct EqualityLeftOutbound end
+struct EqualityRightOutbound end
 
+"""
+    EqualityNode
+ 
+Generic idea of an equality node is to keep track of intermediate `left` and `right` messages 
+in a convenient manner so they can be reused during `prod` for random variables with large `degree`.
+
+ <- left    -------     right ->
+     ----- |   =   | ------
+            -------
+               |
+               | outbound
+
+# Attributes:
+- `left`: An observable that indicates that `left` message can be computed, always sends `missing` since we are not interested in actual value. Can be treated as notification.
+- `right`: An observable that indicates that `right` message can be computed, always sends `missing` since we are not interested in actual value. Can be treated as notification.
+- `cache_left`: Keeps cached value for left outbound message, node itself does not track if cache is valid or not. Cache integrity is tracked by `EqualityChain`.
+- `cache_right`: Keeps cached value for right outbound message, node itself does not track if cache is valid or not. Cache integrity is tracked by `EqualityChain`.
+
+See also: [`EqualityChain`](@ref)
+"""
 mutable struct EqualityNode 
-    index        :: Int
-    cache_left   :: Union{Nothing, Message}
-    cache_right  :: Union{Nothing, Message}
     left         :: LazyObservable{Missing}
     right        :: LazyObservable{Missing}
-    inbound      :: MessageObservable{AbstractMessage}
+    cache_left   :: Message
+    cache_right  :: Message
+
+    EqualityNode() = new(lazy(Missing), lazy(Missing), Message(missing, true, true), Message(missing, true, true))
 end
 
-EqualityNode(index::Int, inbound::MessageObservable{AbstractMessage}) = EqualityNode(index, nothing, nothing, lazy(Missing), lazy(Missing), inbound)
+getoutbound(::EqualityLeftOutbound, node::EqualityNode)  = node.left
+getoutbound(::EqualityRightOutbound, node::EqualityNode) = node.right
 
-getindex(node::EqualityNode)   = node.index
-getleft(node::EqualityNode)    = node.left
-getright(node::EqualityNode)   = node.right
-getinbound(node::EqualityNode) = node.inbound
+setoutbound!(::EqualityLeftOutbound, node::EqualityNode, left)   = set!(node.left, left)
+setoutbound!(::EqualityRightOutbound, node::EqualityNode, right) = set!(node.right, right)
 
-setleft!(node::EqualityNode, left)   = set!(node.left, left)
-setright!(node::EqualityNode, right) = set!(node.right, right)
+getcache(::EqualityLeftOutbound, node::EqualityNode)  = node.cache_left
+getcache(::EqualityRightOutbound, node::EqualityNode) = node.cache_right
 
-get_cache_left(node::EqualityNode)  = node.cache_left
-get_cache_right(node::EqualityNode) = node.cache_right
-
-set_cache_left!(node::EqualityNode, cache)  = node.cache_left = cache
-set_cache_right!(node::EqualityNode, cache) = node.cache_right = cache
-
-function invalidate!(node::EqualityNode)
-    node.cache_left  = nothing
-    node.cache_right = nothing
-    return nothing
-end
+setcache!(::EqualityLeftOutbound, node::EqualityNode, cache::Message)  = node.cache_left = cache
+setcache!(::EqualityRightOutbound, node::EqualityNode, cache::Message) = node.cache_right = cache
 
 # Equality chain
 
[…]
