+++ src/nodes/equality.jl
@@ -1,11 +1,14 @@
 
 import Base: length
+import Base: @propagate_inbounds
 
 ## Equality node is a special case and has a special implementation
 ## It should not be used during model creation but instead is a part of variable node implementation
 
-struct EqualityLeftOutbound end
-struct EqualityRightOutbound end
+abstract type EqualityNodeOutboundType end
+
+struct EqualityLeftOutbound <: EqualityNodeOutboundType end
+struct EqualityRightOutbound <: EqualityNodeOutboundType end
 
 """
     EqualityNode
@@ -69,57 +72,42 @@ Base.length(chain::EqualityChain) = chain.length
 
 prod(chain::EqualityChain, left, right) = chain.prod_fn((left, right))
 
+@propagate_inbounds getnode(chain::EqualityChain, node_index) = chain.nodes[node_index]
+
 function invalidate!(chain::EqualityChain, index) 
     fill_bitarray!(view(chain.cacheleft, firstindex(chain.cacheleft):index), false)
     fill_bitarray!(view(chain.cacheright, index:lastindex(chain.cacheright)), false)
     return nothing
 end
 
-getleft(chain::EqualityChain, node_index::Int)    = (1 < node_index <= length(chain)) ? (@inbounds getoutbound(EqualityLeftOutbound(), chain.nodes[node_index])) : (of(missing))
-getright(chain::EqualityChain, node_index::Int)   = (1 <= node_index < length(chain)) ? (@inbounds getoutbound(EqualityRightOutbound(), chain.nodes[node_index])) : (of(missing))
-getinbound(chain::EqualityChain, node_index::Int) = @inbounds chain.inputmsgs[node_index]
+__check_indices(::EqualityLeftOutbound, chain::EqualityChain, node_index)  = 1 < node_index <= length(chain)
+__check_indices(::EqualityRightOutbound, chain::EqualityChain, node_index) = 1 <= node_index < length(chain)
 
-is_left_cached(chain, node_index::Int)  = @inbounds chain.cacheleft[node_index]
-is_right_cached(chain, node_index::Int) = @inbounds chain.cacheright[node_index]
+@propagate_inbounds getoutbound(type::EqualityNodeOutboundType, chain::EqualityChain, node_index) = __check_indices(type, chain, node_index) ? getoutbound(type, getnode(chain, node_index)) : (of(missing))
+@propagate_inbounds getinbound(chain::EqualityChain, node_index)                                  = chain.inputmsgs[node_index]
 
-function cache_left!(chain::EqualityChain, node_index::Int, node::EqualityNode, cache) 
-    setcache!(EqualityLeftOutbound(), node, cache)
-    @inbounds chain.cacheleft[node_index] = true
-    return nothing
-end
+@propagate_inbounds iscached(::EqualityLeftOutbound, chain::EqualityChain, node_index)  = chain.cacheleft[node_index]
+@propagate_inbounds iscached(::EqualityRightOutbound, chain::EqualityChain, node_index) = chain.cacheright[node_index]
 
-function cache_right!(chain::EqualityChain, node_index::Int, node::EqualityNode, cache) 
-    setcache!(EqualityRightOutbound(), node, cache)
-    @inbounds chain.cacheright[node_index] = true
-    return nothing
-end
+@propagate_inbounds setcache!(::EqualityLeftOutbound, chain::EqualityChain, node_index)  = chain.cacheleft[node_index] = true
+@propagate_inbounds setcache!(::EqualityRightOutbound, chain::EqualityChain, node_index) = chain.cacheright[node_index] = true
 
-function materialize_left!(chain::EqualityChain, node_index::Int)
-    if (1 < node_index <= length(chain))
[…]
