+++ src/nodes/equality.jl
@@ -1,6 +1,7 @@
 
-import Base: length
+import Base: length, map
 import Base: @propagate_inbounds
+import Rocket: tap
 
 ## Equality node is a special case and has a special implementation
 ## It should not be used during model creation but instead is a part of variable node implementation
@@ -53,18 +54,22 @@ setcache!(::EqualityRightOutbound, node::EqualityNode, cache::Message) = node.ca
 
 # Equality chain
 
-struct EqualityChain{F}
+"""
+    EqualityChain
+"""
+struct EqualityChain{P, F}
     length     :: Int
     nodes      :: Vector{EqualityNode}
     inputmsgs  :: Vector{MessageObservable{AbstractMessage}}
     cacheleft  :: BitVector
     cacheright :: BitVector
+    pipeline   :: P
     prod_fn    :: F
 
-    function EqualityChain(inputmsgs::Vector{MessageObservable{AbstractMessage}}, prod_fn::F) where { F } 
+    function EqualityChain(inputmsgs::Vector{MessageObservable{AbstractMessage}}, pipeline::P, prod_fn::F) where { P, F } 
         n     = length(inputmsgs)
         nodes = map(_ -> EqualityNode(), 1:n)
-        return new{F}(n, nodes, inputmsgs, falses(n), falses(n), prod_fn)     
+        return new{P, F}(n, nodes, inputmsgs, falses(n), falses(n), pipeline, prod_fn)     
     end
 end
 
@@ -72,13 +77,9 @@ Base.length(chain::EqualityChain) = chain.length
 
 prod(chain::EqualityChain, left, right) = chain.prod_fn((left, right))
 
-@propagate_inbounds getnode(chain::EqualityChain, node_index) = chain.nodes[node_index]
+getpipeline(chain::EqualityChain) = chain.pipeline
 
-function invalidate!(chain::EqualityChain, index) 
-    fill_bitarray!(view(chain.cacheleft, firstindex(chain.cacheleft):index), false)
-    fill_bitarray!(view(chain.cacheright, index:lastindex(chain.cacheright)), false)
-    return nothing
-end
+@propagate_inbounds getnode(chain::EqualityChain, node_index) = chain.nodes[node_index]
 
 __check_indices(::EqualityLeftOutbound, chain::EqualityChain, node_index)  = 1 < node_index <= length(chain)
 __check_indices(::EqualityRightOutbound, chain::EqualityChain, node_index) = 1 <= node_index < length(chain)
@@ -112,18 +113,54 @@ nextindex(::EqualityRightOutbound, node_index) = node_index - 1
     end
 end
 
-function activate!(model, chain::EqualityChain, outputmsgs::AbstractVector)
+##
+
+struct ChainInvalidationCallback 
+    index      :: Int
+    cacheleft  :: BitVector
+    cacheright :: BitVector
+
+    function ChainInvalidationCallback(index::Int, chain::EqualityChain)
+        return new(index, chain.cacheleft, chain.cacheright)
+    end
+end
+
+Rocket.tap(callback::ChainInvalidationCallback) = Rocket.TapOperator{ChainInvalidationCallback}(callback)
+
+function (callback::ChainInvalidationCallback)(_)
+    fill_bitarray!(view(callback.cacheleft, firstindex(callback.cacheleft):(callback.index)), false)
+    fill_bitarray!(view(callback.cacheright, (callback.index):lastindex(callback.cacheright)), false)
+end
+
+## 
+
+struct ChainOutboundMapping 
+    index :: Int
+    chain :: EqualityChain
+end
+
+function (mapping::ChainOutboundMapping)(_)
+    from_left  = materialize!(EqualityRightOutbound(), mapping.chain, nextindex(EqualityRightOutbound(), mapping.index))
[…]
