### 2021-10-25 466c4b7
feat(): variable collection type info
