+++ src/variables/random.jl
@@ -1,11 +1,14 @@
 export RandomVariable, randomvar
 
+import Base: show
+
 ## Random variable implementation
 
 mutable struct RandomVariable <: AbstractVariable
     name                :: Symbol
-    inputmsgs           :: Vector{MessageObservable{AbstractMessage}} 
-    outputmsgs          :: Union{Nothing, Vector{MessageObservable{Message}}}
+    collection_type     :: AbstractVariableCollectionType
+    input_messages      :: Vector{MessageObservable{AbstractMessage}} 
+    output_messages     :: Union{Nothing, Vector{MessageObservable{Message}}}
     cache               :: Union{Nothing, EqualityChain}
     marginal            :: Union{Nothing, MarginalObservable}
     pipeline            :: AbstractPipelineStage
@@ -15,24 +18,35 @@ mutable struct RandomVariable <: AbstractVariable
     form_check_strategy
 end
 
-function randomvar(name::Symbol; pipeline = EmptyPipelineStage(), prod_constraint = ProdAnalytical(), prod_strategy = FoldLeftProdStrategy(), form_constraint = UnspecifiedFormConstraint(), form_check_strategy = FormConstraintCheckPickDefault()) 
-    return RandomVariable(name, Vector{MessageObservable{AbstractMessage}}(), nothing, nothing, nothing, pipeline, prod_constraint, prod_strategy, form_constraint, form_check_strategy)
+Base.show(io::IO, randomvar::RandomVariable) = print(io, "RandomVariable(", indexed_name(randomvar), ")")
+
+function randomvar(name::Symbol, collection_type::AbstractVariableCollectionType = VariableIndividual(); pipeline = EmptyPipelineStage(), prod_constraint = ProdAnalytical(), prod_strategy = FoldLeftProdStrategy(), form_constraint = UnspecifiedFormConstraint(), form_check_strategy = FormConstraintCheckPickDefault()) 
+    return RandomVariable(name, collection_type, Vector{MessageObservable{AbstractMessage}}(), nothing, nothing, nothing, pipeline, prod_constraint, prod_strategy, form_constraint, form_check_strategy)
 end
 
 function randomvar(name::Symbol, dims::Tuple; pipeline = EmptyPipelineStage(), prod_constraint = ProdAnalytical(), prod_strategy = FoldLeftProdStrategy(), form_constraint = UnspecifiedFormConstraint(), form_check_strategy = FormConstraintCheckPickDefault())
     return randomvar(name, dims...; pipeline = pipeline, prod_constraint = prod_constraint, prod_strategy = prod_strategy, form_constraint = form_constraint, form_check_strategy = form_check_strategy)
 end
 
-function randomvar(name::Symbol, dims::Vararg{Int}; pipeline = EmptyPipelineStage(), prod_constraint = ProdAnalytical(), prod_strategy = FoldLeftProdStrategy(), form_constraint = UnspecifiedFormConstraint(), form_check_strategy = FormConstraintCheckPickDefault())
-    vars = Array{RandomVariable}(undef, dims)
-    for i in CartesianIndices(axes(vars))
-        @inbounds vars[i] = randomvar(Symbol(name, :_, Symbol(join(i.I, :_))); pipeline = pipeline, prod_constraint = prod_constraint, prod_strategy = prod_strategy, form_constraint = form_constraint, form_check_strategy = form_check_strategy)
+function randomvar(name::Symbol, length::Int; pipeline = EmptyPipelineStage(), prod_constraint = ProdAnalytical(), prod_strategy = FoldLeftProdStrategy(), form_constraint = UnspecifiedFormConstraint(), form_check_strategy = FormConstraintCheckPickDefault())
+    vars = Vector{RandomVariable}(undef, length)
[…]
