### 2021-10-26 7915e32
fix(): refactor variable marginal observables
