+++ src/variables/constant.jl
@@ -57,10 +57,14 @@ messagein(constvar::ConstVariable, ::Int)  = error("It is not possible to get a 
 
 get_pipeline_stages(::ConstVariable) = EmptyPipelineStage()
 
-_getmarginal(constvar::ConstVariable) = of(Marginal(constvar.constant, true, false))
+_getmarginal(constvar::ConstVariable)      = of(Marginal(constvar.constant, true, false))
+_setmarginal!(::ConstVariable, observable) = error("It is not possible to set a marginal stream for `ConstVariable`")
+_makemarginal(::ConstVariable)             = error("It is not possible to make marginal stream for `ConstVariable`")
 
-_setmarginal!(::ConstVariable, ::MarginalObservable) = error("It is not possible to set a marginal stream for constvar")
-_makemarginal(::ConstVariable)                       = error("It is not possible to make marginal stream for constvar")
+# For _getmarginal
+function Rocket.getrecent(observable::SingleObservable{ <: Marginal })
+    return observable.value
+end
 
 function setmessagein!(constvar::ConstVariable, ::Int, messagein) 
     constvar.nconnected += 1
