+++ src/variables/random.jl
@@ -8,9 +8,10 @@ mutable struct RandomVariable <: AbstractVariable
     name                :: Symbol
     collection_type     :: AbstractVariableCollectionType
     input_messages      :: Vector{MessageObservable{AbstractMessage}} 
-    output_messages     :: Union{Nothing, Vector{MessageObservable{Message}}}
-    cache               :: Union{Nothing, EqualityChain}
-    marginal            :: Union{Nothing, MarginalObservable}
+    output_messages     :: Vector{MessageObservable{Message}}
+    output_initialised  :: Bool
+    output_cache        :: Union{Nothing, EqualityChain}
+    marginal            :: MarginalObservable
     pipeline            :: AbstractPipelineStage
     prod_constraint     
     prod_strategy       
@@ -21,7 +22,7 @@ end
 Base.show(io::IO, randomvar::RandomVariable) = print(io, "RandomVariable(", indexed_name(randomvar), ")")
 
 function randomvar(name::Symbol, collection_type::AbstractVariableCollectionType = VariableIndividual(); pipeline = EmptyPipelineStage(), prod_constraint = ProdAnalytical(), prod_strategy = FoldLeftProdStrategy(), form_constraint = UnspecifiedFormConstraint(), form_check_strategy = FormConstraintCheckPickDefault()) 
-    return RandomVariable(name, collection_type, Vector{MessageObservable{AbstractMessage}}(), nothing, nothing, nothing, pipeline, prod_constraint, prod_strategy, form_constraint, form_check_strategy)
+    return RandomVariable(name, collection_type, Vector{MessageObservable{AbstractMessage}}(), Vector{MessageObservable{Message}}(), false, nothing, MarginalObservable(), pipeline, prod_constraint, prod_strategy, form_constraint, form_check_strategy)
 end
 
 function randomvar(name::Symbol, dims::Tuple; pipeline = EmptyPipelineStage(), prod_constraint = ProdAnalytical(), prod_strategy = FoldLeftProdStrategy(), form_constraint = UnspecifiedFormConstraint(), form_check_strategy = FormConstraintCheckPickDefault())
@@ -64,20 +65,18 @@ getlastindex(randomvar::RandomVariable) = degree(randomvar) + 1
 messagein(randomvar::RandomVariable, index::Int)  = @inbounds randomvar.input_messages[index]
 
 function messageout(randomvar::RandomVariable, index::Int) 
-    if randomvar.output_messages === nothing
+    if randomvar.output_initialised === false
         initialize_output_messages!(randomvar)
-        return messageout(randomvar, index)
-    else
-        return @inbounds randomvar.output_messages[index]
     end
+    return @inbounds randomvar.output_messages[index]
 end
 
 get_pipeline_stages(randomvar::RandomVariable)        = randomvar.pipeline
 add_pipeline_stage!(randomvar::RandomVariable, stage) = randomvar.pipeline = (randomvar.pipeline + stage)
 
-_getmarginal(randomvar::RandomVariable)                                = randomvar.marginal
-_setmarginal!(randomvar::RandomVariable, marginal::MarginalObservable) = randomvar.marginal = marginal
-_makemarginal(randomvar::RandomVariable)                               = collectLatest(AbstractMessage, Marginal, randomvar.input_messages, marginal_prod_fn(randomvar))
+_getmarginal(randomvar::RandomVariable)              = randomvar.marginal
+_setmarginal!(randomvar::RandomVariable, observable) = connect!(_getmarginal(randomvar), observable)
+_makemarginal(randomvar::RandomVariable)             = collectLatest(AbstractMessage, Marginal, randomvar.input_messages, marginal_prod_fn(randomvar))
 
[…]

[…]
