### 2021-11-03 4066089
fix(): Fix for point mass infinite entropy typing
