+++ test/distributions/test_pointmass.jl
@@ -6,146 +6,155 @@ using Distributions
 using Random
 using SpecialFunctions
 
-import ReactiveMP: ∞
+import ReactiveMP: ∞, tiny, huge
+import ReactiveMP.MacroHelpers: @test_inferred
 
 @testset "PointMass" begin
 
+    rng = MersenneTwister(1234)
+
     @testset "Real-based PointMass" begin
+        for T in (Float16, Float32, Float64, BigFloat)
+            scalar = rand(rng, T)
+            dist   = PointMass(scalar)
 
-        scalar = rand()
-        dist   = PointMass(scalar)
+            @test variate_form(dist) === Univariate
+            @test_throws BoundsError dist[2]
+            @test_throws BoundsError dist[2, 2]
 
-        @test variate_form(dist) === Univariate
-        @test_throws BoundsError dist[2]
-        @test_throws BoundsError dist[2, 2]
+            @test insupport(dist, scalar)
+            @test !insupport(dist, scalar + tiny)
+            @test !insupport(dist, scalar - tiny)
 
-        @test insupport(dist, scalar)
-        @test !insupport(dist, scalar + 1e-6)
-        @test !insupport(dist, scalar - 1e-6)
+            @test @test_inferred(T, pdf(dist, scalar))        == one(T)
+            @test @test_inferred(T, pdf(dist, scalar + tiny)) == zero(T)
+            @test @test_inferred(T, pdf(dist, scalar - tiny)) == zero(T)
 
-        @test pdf(dist, scalar)        === 1.0
-        @test pdf(dist, scalar + 1e-6) === 0.0
-        @test pdf(dist, scalar - 1e-6) === 0.0
+            @test @test_inferred(T, logpdf(dist, scalar))        == zero(T)
+            @test @test_inferred(T, logpdf(dist, scalar + tiny)) == convert(T, -Inf)
+            @test @test_inferred(T, logpdf(dist, scalar - tiny)) == convert(T, -Inf)
 
-        @test logpdf(dist, scalar)        === 0.0
-        @test logpdf(dist, scalar + 1e-6) === -Inf
-        @test logpdf(dist, scalar - 1e-6) === -Inf
+            @test_throws MethodError insupport(dist, ones(T, 2))
+            @test_throws MethodError insupport(dist, ones(T, 2, 2))
+            @test_throws MethodError pdf(dist, ones(T, 2))
+            @test_throws MethodError pdf(dist, ones(T, 2, 2))
+            @test_throws MethodError logpdf(dist, ones(T, 2))
+            @test_throws MethodError logpdf(dist, ones(T, 2, 2))
 
-        @test_throws MethodError insupport(dist, [ 1.0, 1.0 ])
-        @test_throws MethodError insupport(dist, [ 1.0 1.0; 1.0 1.0 ])
-        @test_throws MethodError pdf(dist, [ 1.0, 1.0 ])
-        @test_throws MethodError pdf(dist, [ 1.0 1.0; 1.0 1.0 ])
-        @test_throws MethodError logpdf(dist, [ 1.0, 1.0 ])
-        @test_throws MethodError logpdf(dist, [ 1.0 1.0; 1.0 1.0 ])
+            @test (@inferred entropy(dist)) === -∞
 
-        @test entropy(dist) === -∞
+            @test @test_inferred(T, mean(dist))         == scalar
+            @test @test_inferred(T, var(dist))          == zero(T)
+            @test @test_inferred(T, std(dist))          == zero(T)
+            @test @test_inferred(T, cov(dist))          == zero(T)
+            @test @test_inferred(T, precision(dist))    == convert(T, Inf)
+            @test @test_inferred(Int, ndims(dist))      == 1
+            @test @test_inferred(Type{T}, eltype(dist)) == T
 
-        @test mean(dist)       === scalar
[…]
