+++ Project.toml
@@ -46,14 +46,17 @@ julia = "1.5.0"
 
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 BenchmarkCI = "20533458-34a3-403d-a444-e18f38190b5b"
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 Coverage = "a2441757-f6aa-5fb2-8edb-039e3f45d037"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+GraphPPL = "b3f8163a-e979-4e85-b43e-1f63d8c8b42c"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 PkgBenchmark = "32113eaa-f34f-5b0d-bd6c-c81e245fc73d"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TestSetExtensions = "98d24dd4-01ad-11ea-1b02-c9a08f80db04"
 
 [targets]
-test = ["Test", "InteractiveUtils", "TestSetExtensions", "Coverage", "Documenter", "BenchmarkCI", "BenchmarkTools", "PkgBenchmark", "Aqua"]
+test = ["Test", "InteractiveUtils", "TestSetExtensions", "Coverage", "Dates", "Documenter", "GraphPPL", "Plots", "BenchmarkCI", "BenchmarkTools", "PkgBenchmark", "Aqua"]
