+++ test/models/test_lgssm.jl
@@ -2,7 +2,7 @@ module ReactiveMPModelsLGSSMTest
 
 using Test
 using Rocket, ReactiveMP, GraphPPL, Distributions
-using BenchmarkTools, Random, Plots, Dates
+using BenchmarkTools, Random, Plots, Dates, LinearAlgebra
 
 ## Model definition
 ## -------------------------------------------- ##
@@ -25,19 +25,44 @@ using BenchmarkTools, Random, Plots, Dates
 
     return x, y
 end
+
+@model function multivariate_lgssm_model(n, x0, A, B, Q, P)
+    
+    # We create constvar references for better efficiency
+    cA = constvar(A)
+    cB = constvar(B)
+    cQ = constvar(Q)
+    cP = constvar(P)
+    
+    # `x` is a sequence of hidden states
+    x = randomvar(n)
+    # `y` is a sequence of "clamped" observations
+    y = datavar(Vector{Float64}, n)
+    
+    x_prior ~ MvNormalMeanCovariance(mean(x0), cov(x0))
+    x_prev = x_prior
+    
+    for i in 1:n
+        x[i] ~ MvNormalMeanCovariance(cA * x_prev, cQ)
+        y[i] ~ MvNormalMeanCovariance(cB * x[i], cP)
+        x_prev = x[i]
+    end
+    
+    return x, y
+end
 ## -------------------------------------------- ##
 ## Inference definition
 ## -------------------------------------------- ##
 function univariate_lgssm_inference(data, x0, c, P)
     n = length(data)
-    
-    model, (x, y) = univariate_lgssm_model(n, x0, c, P);
 
-    x_buffer = buffer(Marginal, n)
-    fe       = 0.0
+    model, (x, y) = univariate_lgssm_model(n, x0, c, P);
     
+    x_buffer = buffer(Marginal, n)
+    fe       = keep(Float64)
+
     x_sub = subscribe!(getmarginals(x), x_buffer)
-    f_sub = subscribe!(score(Float64, BetheFreeEnergy(), model), (v) -> fe = v)
+    f_sub = subscribe!(score(Float64, BetheFreeEnergy(), model), fe)
     
     update!(y, data)
     
@@ -46,10 +71,27 @@ function univariate_lgssm_inference(data, x0, c, P)
     return x_buffer, fe
 end
 
+function multivariate_lgssm_inference(data, x0, A, B, Q, P)
+    n = length(data)
+
+    model, (x, y) = multivariate_lgssm_model(n, x0, A, B, Q, P, options = (limit_stack_depth = 100, ));
+
+    xbuffer = buffer(Marginal, n)
+    fe      = keep(Float64)
+
+    x_sub = subscribe!(getmarginals(x), xbuffer)
+    f_sub  = subscribe!(score(Float64, BetheFreeEnergy(), model), fe)
+
+    update!(y, data)
+
+    unsubscribe!((x_sub, f_sub))
+    
+    return xbuffer, fe
+end
+
 @testset "Linear Gaussian State Space Model" begin
 
     @testset "Univariate" begin 
-
         ## -------------------------------------------- ##
         ## Data creation
         ## -------------------------------------------- ##
@@ -67,7 +109,7 @@ end
         @test length(x_estimated) === n
         @test all((hidden .- 3 .* std.(x_estimated)) .< mean.(x_estimated) .< (hidden .+ 3 .* std.(x_estimated)))
         @test all(var.(x_estimated) .> 0.0)
-        @test fe ≈ 1896.15674252578
+        @test length(fe) === 1 && first(fe) ≈ 1896.15674252578
         ## -------------------------------------------- ##
         ## Form debug output
         base_output = joinpath(pwd(), "_output", "models")
[…]
