### 2021-11-03 b216006
Added PointMass to Normal Mixture
Added option of PointMass for the Normal Mixture node, for when the means and precisions are known.
