+++ src/rules/normal_mixture/switch.jl
@@ -21,3 +21,17 @@ end
     end
     return Categorical(clamp.(softmax(U), tiny, one(eltype(U)) - tiny))
 end
+
+@rule NormalMixture{N}(:switch, Marginalisation) (q_out::Any, q_m::NTuple{N, PointMass{T}}, q_p::NTuple{N, PointMass{T}}) where { N, T <: Real } = begin
+    U = map(zip(q_m, q_p)) do (m, p)
+        return -score(AverageEnergy(), NormalMeanPrecision, Val{ (:out, :μ, :τ) }, map((q) -> Marginal(q, false , false), (q_out, m, p)), nothing)
+    end
+    return Categorical(clamp.(softmax(U), tiny, one(eltype(U)) - tiny))
+end
+
+@rule NormalMixture{N}(:switch, Marginalisation) (q_out::Any, q_m::NTuple{N, PointMass}, q_p::NTuple{N, PointMass}) where { N } = begin
+    U = map(zip(q_m, q_p)) do (m, p)
+        return -score(AverageEnergy(), MvNormalMeanPrecision, Val{ (:out, :μ, :Λ) }, map((q) -> Marginal(q, false , false), (q_out, m, p)), nothing)
+    end
+    return Categorical(clamp.(softmax(U), tiny, one(eltype(U)) - tiny))
+end
