+++ test/models/test_hgf.jl
@@ -1,0 +1,180 @@
+module ReactiveMPDistributionsTest
+
+using Test
+
+using Rocket, ReactiveMP, GraphPPL, Distributions
+using BenchmarkTools, Random, Plots, Dates, LinearAlgebra
+
+## Model definition
+## -------------------------------------------- ##
+@model [ default_factorisation = MeanField() ] function hgf_online_model(real_k, real_w, z_variance, y_variance)
+    
+    # Priors from previous time step for `z`
+    zt_min_mean = datavar(Float64)
+    zt_min_var  = datavar(Float64)
+    
+    # Priors from previous time step for `x`
+    xt_min_mean = datavar(Float64)
+    xt_min_var  = datavar(Float64)
+
+    zt_min ~ NormalMeanVariance(zt_min_mean, zt_min_var)
+    xt_min ~ NormalMeanVariance(xt_min_mean, xt_min_var)
+
+    # Higher layer is modelled as a random walk 
+    zt ~ NormalMeanVariance(zt_min, z_variance) where { q = q(zt, zt_min)q(z_variance) }
+    
+    # Lower layer is modelled with `GCV` node
+    gcv_node, xt ~ GCV(xt_min, zt, real_k, real_w) where { q = q(xt, xt_min)q(zt)q(κ)q(ω) }
+    
+    # Noisy observations 
+    y = datavar(Float64)
+    y ~ NormalMeanVariance(xt, y_variance)
+    
+    return zt, xt, y, gcv_node, xt_min_mean, xt_min_var, zt_min_mean, zt_min_var
+end
+## -------------------------------------------- ##
+## Inference definition
+## -------------------------------------------- ##
+function hgf_online_inference(data, vmp_iters, real_k, real_w, z_variance, y_variance)
+    n = length(data)
+    
+    # We don't want to save all marginals from all VMP iterations
+    # but only last one after all VMP iterations per time step
+    # Rocket.jl exports PendingScheduler() object that postpones 
+    # any update unless manual `resolve!()` has been called
+    ms_scheduler = PendingScheduler()
+    
+    mz = keep(Marginal)
+    mx = keep(Marginal)
+    fe = ScoreActor(Float64)
+
+    model, (zt, xt, y, gcv_node, xt_min_mean, xt_min_var, zt_min_mean, zt_min_var) = hgf_online_model(real_k, real_w, z_variance, y_variance)
+
+    # Initial priors
+    current_zt_mean, current_zt_var = 0.0, 10.0
+    current_xt_mean, current_xt_var = 0.0, 10.0
+    
+    s_mz = subscribe!(getmarginal(zt) |> schedule_on(ms_scheduler), mz)
+    s_mx = subscribe!(getmarginal(xt) |> schedule_on(ms_scheduler), mx)
+    s_fe = subscribe!(score(Float64, BetheFreeEnergy(), model), fe)
+
+    # Initial marginals to start VMP procedire
+    setmarginal!(gcv_node, :y_x, MvNormalMeanCovariance([ 0.0, 0.0 ], [ 5.0, 5.0 ]))
+    setmarginal!(gcv_node, :z, NormalMeanVariance(0.0, 5.0))
+    
+    # For each observations we perofrm `vmp_iters` VMP iterations
+    for i in 1:n
+        
+        for _ in 1:vmp_iters
+            update!(y, data[i])
+            update!(zt_min_mean, current_zt_mean)
+            update!(zt_min_var, current_zt_var)
+            update!(xt_min_mean, current_xt_mean)
+            update!(xt_min_var, current_xt_var)
+        end
+        
+        # After all VMP iterations we release! `PendingScheduler`
+        # as well as release! `ScoreActor` to indicate new time step
+        release!(ms_scheduler)
+        release!(fe)
+        
[…]
