+++ test/models/test_lgssm.jl
@@ -1,6 +1,6 @@
 module ReactiveMPModelsLGSSMTest
 
-using Test
+using Test, InteractiveUtils
 using Rocket, ReactiveMP, GraphPPL, Distributions
 using BenchmarkTools, Random, Plots, Dates, LinearAlgebra
 
