### 2021-11-04 025b916
test(): added ar model test
