+++ test/models/test_ar.jl
@@ -1,0 +1,237 @@
+module ReactiveMPModelsAutoregressiveTest
+
+using Test, InteractiveUtils
+using Rocket, ReactiveMP, GraphPPL, Distributions
+using BenchmarkTools, Random, Plots, Dates, LinearAlgebra
+
+## Model definition
+## -------------------------------------------- ##
+@model [ default_factorisation = MeanField() ] function lar_model(T::Type{ Multivariate }, n, order, c, stype, τ)
+
+    # Parameter priors
+    γ  ~ GammaShapeRate(1.0, 1.0)
+    θ  ~ MvNormalMeanPrecision(zeros(order), diageye(order))
+
+    # We create a sequence of random variables for hidden states
+    x = randomvar(n)
+    # As well a sequence of observartions
+    y = datavar(Float64, n)
+
+    ct = constvar(c)
+    # We assume observation noise to be known
+    cτ = constvar(τ)
+
+    # Prior for first state
+    x0 ~ MvNormalMeanPrecision(zeros(order), diageye(order))
+
+    x_prev = x0
+
+    # AR process requires extra meta information
+    meta = ARMeta(Multivariate, order, stype)
+
+    for i in 1:n
+        # Autoregressive node uses structured factorisation assumption between states
+        x[i] ~ AR(x_prev, θ, γ) where { q = q(y, x)q(γ)q(θ), meta = meta }
+        y[i] ~ NormalMeanPrecision(dot(ct, x[i]), cτ)
+        x_prev = x[i]
+    end
+
+    return x, y, θ, γ
+end
+@model [ default_factorisation = MeanField() ] function lar_model(T::Type{Univariate}, n, order, c, stype, τ)
+
+    # Parameter priors
+    γ  ~ GammaShapeRate(1.0, 1.0)
+    θ  ~ NormalMeanPrecision(0.0, 1.0)
+
+    # We create a sequence of random variables for hidden states
+    x = randomvar(n)
+    # As well a sequence of observartions
+    y = datavar(Float64, n)
+
+    ct = constvar(c)
+    # We assume observation noise to be known
+    cτ = constvar(τ)
+
+    # Prior for first state
+    x0 ~ NormalMeanPrecision(0.0, 1.0)
+
+    x_prev = x0
+
+    # AR process requires extra meta information
+    meta = ARMeta(Univariate, order, stype)
+
+    for i in 1:n
+        x[i] ~ AR(x_prev, θ, γ) where { q = q(y, x)q(γ)q(θ), meta = meta }
+        y[i] ~ NormalMeanPrecision(ct * x[i], cτ)
+        x_prev = x[i]
+    end
+
+    return x, y, θ, γ
+end
+## -------------------------------------------- ##
+## Inference definition
+## -------------------------------------------- ##
+function init_marginals!(::Type{ Multivariate }, order, γ, θ)
+    setmarginal!(γ, GammaShapeRate(1.0, 1.0))
+    setmarginal!(θ, MvNormalMeanPrecision(zeros(order), diageye(order)))
+end
+
+function init_marginals!(::Type{ Univariate }, order, γ, θ)
+    setmarginal!(γ, GammaShapeRate(1.0, 1.0))
+    setmarginal!(θ, NormalMeanPrecision(0.0, 1.0))
+end
+
+function inference(data, order, artype, stype, niter, τ)
+
+    # We build a full graph based on nber of observatios
+    n = length(data)
+
+    # Depending on the order of AR process `c` is
+    # either a nber or a vector
+    c = ReactiveMP.ar_unit(artype, order)
+
+    # Note that to run inference for huge model it might be necessary to pass extra
[…]
