### 2021-11-04 3225b6d
test(): added ar and gmm tests
