+++ test/models/test_gmm.jl
@@ -1,0 +1,323 @@
+module ReactiveMPModelsTemplateTest
+
+using Test, InteractiveUtils
+using Rocket, ReactiveMP, GraphPPL, Distributions
+using BenchmarkTools, Random, Plots, Dates, LinearAlgebra
+
+## Model definition
+## -------------------------------------------- ##
+@model [ default_factorisation = MeanField() ] function univariate_gaussian_mixture_model(n)
+    
+    s ~ Beta(1.0, 1.0)
+    
+    m1 ~ NormalMeanVariance(-2.0, 1e3)
+    w1 ~ GammaShapeRate(0.01, 0.01)
+    
+    m2 ~ NormalMeanVariance(2.0, 1e3)
+    w2 ~ GammaShapeRate(0.01, 0.01)
+    
+    z = randomvar(n)
+    y = datavar(Float64, n)
+    
+    for i in 1:n
+        z[i] ~ Bernoulli(s)
+        y[i] ~ NormalMixture(z[i], (m1, m2), (w1, w2))
+    end
+    
+    return s, m1, w1, m2, w2, z, y
+end
+
+@model function multivariate_gaussian_mixture_model(rng, L, nmixtures, n)
+    
+    z = randomvar(n)
+    m = randomvar(nmixtures)
+    w = randomvar(nmixtures)
+    
+    basis_v = [ 1.0, 0.0 ]
+    
+    for i in 1:nmixtures
+        # Assume we now only approximate location of cluters's mean
+        approximate_angle_prior = ((2π + + rand(rng)) / nmixtures) * (i - 1)
+        approximate_basis_v  = L / 2 * (basis_v .+ rand(rng, 2))
+        approximate_rotation = [ cos(approximate_angle_prior) -sin(approximate_angle_prior); sin(approximate_angle_prior) cos(approximate_angle_prior) ]
+        mean_mean_prior     = approximate_rotation * approximate_basis_v
+        mean_mean_cov       = [ 1e6 0.0; 0.0 1e6 ]
+        
+        m[i] ~ MvNormalMeanCovariance(mean_mean_prior, mean_mean_cov)
+        w[i] ~ Wishart(3, [ 1e2 0.0; 0.0 1e2 ])
+    end
+    
+    s ~ Dirichlet(ones(nmixtures))
+
+    y = datavar(Vector{Float64}, n)
+    
+    means = tuple(m...)
+    precs = tuple(w...)
+    
+    for i in 1:n
+        z[i] ~ Categorical(s) where { q = MeanField() }
+        y[i] ~ NormalMixture(z[i], means, precs) where { q = MeanField() }
+    end
+    
+    return s, z, m, w, y
+end
+## -------------------------------------------- ##
+## Inference definition
+## -------------------------------------------- ##
+function inference_univariate(data, n_its)
+    n = length(data)
+    model, (s, m1, w1, m2, w2, z, y) = univariate_gaussian_mixture_model(n);
+    
+    mswitch = keep(Marginal)
+    mm1 = keep(Marginal)
+    mm2 = keep(Marginal)
+    mw1 = keep(Marginal)
+    mw2 = keep(Marginal)
+    
+    fe = keep(Float64)
+    
+    m1sub = subscribe!(getmarginal(m1), mm1)
+    m2sub = subscribe!(getmarginal(m2), mm2)
+    w1sub = subscribe!(getmarginal(w1), mw1)
+    w2sub = subscribe!(getmarginal(w2), mw2)
+    switchsub = subscribe!(getmarginal(s), mswitch)
+    fesub = subscribe!(score(Float64, BetheFreeEnergy(), model), fe)
+    
[…]
