+++ test/runtests.jl
@@ -146,6 +146,7 @@ end
     addtests("models/test_hgf.jl")
     addtests("models/test_ar.jl")
     addtests("models/test_gmm.jl")
+    addtests("models/test_hmm.jl")
 
 end
 
