+++ test/models/test_hmm.jl
@@ -1,0 +1,142 @@
+module ReactiveMPModelsHMMTest
+
+using Test, InteractiveUtils
+using Rocket, ReactiveMP, GraphPPL, Distributions
+using BenchmarkTools, Random, Plots, Dates, LinearAlgebra
+
+## Model definition
+## -------------------------------------------- ##
+@model [ default_factorisation = MeanField() ] function transition_model(n)
+    
+    A ~ MatrixDirichlet(ones(3, 3)) 
+    B ~ MatrixDirichlet([ 10.0 1.0 1.0; 1.0 10.0 1.0; 1.0 1.0 10.0 ])
+    
+    s_0 ~ Categorical(fill(1.0 / 3.0, 3))
+    
+    s = randomvar(n)
+    x = datavar(Vector{Float64}, n)
+    
+    s_prev = s_0
+    
+    for t in 1:n
+        s[t] ~ Transition(s_prev, A) where { q = q(out, in)q(a) }
+        x[t] ~ Transition(s[t], B)
+        s_prev = s[t]
+    end
+    
+    return s, x, A, B
+end
+## -------------------------------------------- ##
+## Inference definition
+## -------------------------------------------- ##
+function inference(data, vmp_iters)
+    n = length(data)
+    
+    model, (s, x, A, B) = transition_model(n, options = (limit_stack_depth = 500, ))
+    
+    sbuffer = keep(Vector{Marginal})
+    Abuffer = keep(Marginal)
+    Bbuffer = keep(Marginal)
+    fe      = ScoreActor(Float64)
+    
+    ssub  = subscribe!(getmarginals(s), sbuffer)
+    Asub  = subscribe!(getmarginal(A), Abuffer)
+    Bsub  = subscribe!(getmarginal(B), Bbuffer)
+    fesub = subscribe!(score(Float64, BetheFreeEnergy(), model), fe)
+    
+    setmarginal!(A, vague(MatrixDirichlet, 3, 3))
+    setmarginal!(B, vague(MatrixDirichlet, 3, 3))
+    
+    foreach(s) do svar
+        setmarginal!(svar, vague(Categorical, 3))
+    end
+    
+    for i in 1:vmp_iters
+        update!(x, data)
+    end
+    
+    unsubscribe!(ssub)
+    unsubscribe!(Asub)
+    unsubscribe!(Bsub)
+    unsubscribe!(fesub)
+    
+    return map(getvalues, (sbuffer, Abuffer, Bbuffer, fe))
+end
+
+@testset "Hidden Markov Model" begin
+
+    @testset "Full graph" begin 
+        ## -------------------------------------------- ##
+        ## Data creation
+        ## -------------------------------------------- ##
+        function rand_vec(distribution::Categorical) 
+            k = ncategories(distribution)
+            s = zeros(k)
+            s[ rand(distribution) ] = 1.0
+            s
+        end
+        
+        function generate_data(n_samples; seed = 124)
+            Random.seed!(seed)
+            
+            # Transition probabilities (some transitions are impossible)
+            A = [0.9 0.0 0.1; 0.1 0.9 0.0; 0.0 0.1 0.9] 
+            # Observation noise
+            B = [0.9 0.05 0.05; 0.05 0.9 0.05; 0.05 0.05 0.9] 
+            # Initial state
+            s_0 = [1.0, 0.0, 0.0] 
+            # Generate some data
+            s = Vector{Vector{Float64}}(undef, n_samples) # one-hot encoding of the states
+            x = Vector{Vector{Float64}}(undef, n_samples) # one-hot encoding of the observations
+            
+            s_prev = s_0
+            
+            for t = 1:n_samples
+                a = A * s_prev
[…]
