### 2021-11-04 ee5390d
test(): use StableRNGs for data generation
