+++ Project.toml
@@ -55,9 +55,9 @@ GraphPPL = "b3f8163a-e979-4e85-b43e-1f63d8c8b42c"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 PkgBenchmark = "32113eaa-f34f-5b0d-bd6c-c81e245fc73d"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TestSetExtensions = "98d24dd4-01ad-11ea-1b02-c9a08f80db04"
-StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 
 [targets]
 test = ["Test", "InteractiveUtils", "TestSetExtensions", "Coverage", "Dates", "Documenter", "GraphPPL", "Plots", "BenchmarkCI", "BenchmarkTools", "PkgBenchmark", "Aqua", "StableRNGs"]
