### 2021-11-04 31f7da9
test(): improve benchmark stability
