+++ test/models/test_lgssm.jl
@@ -131,7 +131,7 @@ end
         savefig(p, plot_output)
         ## -------------------------------------------- ##
         ## Create output benchmarks
-        benchmark = @benchmark univariate_lgssm_inference($data, $x0_prior, 1.0, $P)
+        benchmark = @benchmark univariate_lgssm_inference($data, $x0_prior, 1.0, $P) seconds=15
         open(benchmark_output, "w") do io
             show(io, MIME("text/plain"), benchmark)
             versioninfo(io)
@@ -204,7 +204,7 @@ end
         savefig(px, plot_output)
         ## -------------------------------------------- ##
         ## Create output benchmarks
-        benchmark = @benchmark multivariate_lgssm_inference($y, $x0, $A, $B, $Q, $P)
+        benchmark = @benchmark multivariate_lgssm_inference($y, $x0, $A, $B, $Q, $P) seconds=15
         open(benchmark_output, "w") do io
             show(io, MIME("text/plain"), benchmark)
             versioninfo(io)
