### 2021-11-05 df63cfb
test(): add pure ar test
