+++ test/models/test_ar.jl
@@ -6,6 +6,21 @@ using BenchmarkTools, Random, Plots, Dates, LinearAlgebra, StableRNGs
 
 ## Model definition
 ## -------------------------------------------- ##
+@model [ default_factorisation = MeanField() ] function ar_model(n, order)
+
+    x = datavar(Vector{Float64}, n)
+    y = datavar(Float64, n)
+
+    γ ~ GammaShapeRate(1.0, 1.0)
+    θ ~ MvNormalMeanPrecision(zeros(order), diageye(order))
+
+    for i in 1:n
+        y[i] ~ NormalMeanPrecision(dot(x[i], θ), γ)
+    end
+
+    return x, y, θ, γ
+end
+
 @model [ default_factorisation = MeanField() ] function lar_model(T::Type{ Multivariate }, n, order, c, stype, τ)
 
     # Parameter priors
@@ -72,17 +87,42 @@ end
 ## -------------------------------------------- ##
 ## Inference definition
 ## -------------------------------------------- ##
-function init_marginals!(::Type{ Multivariate }, order, γ, θ)
+function ar_inference(inputs, outputs, order, niter)
+    n = length(outputs)
+
+    model, (x, y, θ, γ) = ar_model(n, order, options = (limit_stack_depth = 500, ))
+
+    γ_buffer = keep(Marginal)
+    θ_buffer = keep(Marginal)
+    fe       = keep(Float64)
+
+    γ_sub = subscribe!(getmarginal(γ), γ_buffer)
+    θ_sub = subscribe!(getmarginal(θ), θ_buffer)
+    f_sub = subscribe!(score(Float64, BetheFreeEnergy(), model), fe)
+
+    setmarginal!(γ, GammaShapeRate(1.0, 1.0))
+
+    for _ in 1:niter
+        update!(x, inputs)
+        update!(y, outputs)
+    end
+
+    unsubscribe!((γ_sub, θ_sub, f_sub))
+
+    return γ_buffer, θ_buffer, fe
+end
+
+function lar_init_marginals!(::Type{ Multivariate }, order, γ, θ)
     setmarginal!(γ, GammaShapeRate(1.0, 1.0))
     setmarginal!(θ, MvNormalMeanPrecision(zeros(order), diageye(order)))
 end
 
-function init_marginals!(::Type{ Univariate }, order, γ, θ)
+function lar_init_marginals!(::Type{ Univariate }, order, γ, θ)
     setmarginal!(γ, GammaShapeRate(1.0, 1.0))
     setmarginal!(θ, NormalMeanPrecision(0.0, 1.0))
 end
 
-function inference(data, order, artype, stype, niter, τ)
+function lar_inference(data, order, artype, stype, niter, τ)
 
     # We build a full graph based on nber of observatios
     n = length(data)
@@ -108,7 +148,7 @@ function inference(data, order, artype, stype, niter, τ)
     xsub  = subscribe!(getmarginals(x), x_buffer)
     fesub = subscribe!(score(Float64, BetheFreeEnergy(), model), fe)
 
-    init_marginals!(artype, order, γ, θ)
+    lar_init_marginals!(artype, order, γ, θ)
 
     # We update data several times to perform several VMP iterations
     for i in 1:niter
@@ -123,14 +163,57 @@ end
 
 @testset "Autoregressive model" begin
 
-    @testset "Full graph inference" begin 
+    @testset "Autoregressive model" begin
+        ## -------------------------------------------- ##
+        ## Data creation
+        ## -------------------------------------------- ##
+        function ar_ssm(series, order)
+            inputs = [reverse!(series[1:order])]
+            outputs = [series[order + 1]]
+            for x in series[order+2:end]
[…]
