### 2021-11-05 b1b61df
fix(): Fix for stackoverflow for large equality chains
