### 2021-11-05 bbfeb50
update(): Bump version to 1.1.0
