+++ versions.js
@@ -1,7 +1,8 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.1",
   "v1.0",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.0.6";
+var DOCUMENTER_NEWEST = "v1.1.0";
 var DOCUMENTER_STABLE = "stable";
