+++ src/nodes/normal_mixture.jl
@@ -141,14 +141,14 @@ end
     end
 end
 
-@average_energy NormalMixture (q_out::Any, q_switch::Any, q_m::NTuple{N, PointMass{T}}, q_p::NTuple{N, PointMass{T}}) where { N, T <: Real } = begin
+@average_energy NormalMixture (q_out::Any, q_switch::Any, q_m::NTuple{N, PointMass{T} where T <: Real}, q_p::NTuple{N, PointMass{T} where T <: Real}) where N = begin
     z_bar = probvec(q_switch)
     return mapreduce(+, 1:N, init = 0.0) do i
         return z_bar[i] * score(AverageEnergy(), NormalMeanPrecision, Val{ (:out, :μ, :τ) }, map((q) -> Marginal(q, false, false), (q_out, q_m[i], q_p[i])), nothing)
     end
 end
 
-@average_energy NormalMixture (q_out::Any, q_switch::Any, q_m::NTuple{N, PointMass}, q_p::NTuple{N, PointMass}) where N = begin
+@average_energy NormalMixture (q_out::Any, q_switch::Any, q_m::NTuple{N, PointMass{T} where T <: AbstractVector}, q_p::NTuple{N, PointMass{T} where T <: AbstractMatrix}) where N = begin
     z_bar = probvec(q_switch)
     return mapreduce(+, 1:N, init = 0.0) do i
         return z_bar[i] * score(AverageEnergy(), MvNormalMeanPrecision, Val{ (:out, :μ, :Λ) }, map((q) -> Marginal(q, false, false), (q_out, q_m[i], q_p[i])), nothing)
