+++ src/rules/normal_mixture/switch.jl
@@ -22,14 +22,14 @@ end
     return Categorical(clamp.(softmax(U), tiny, one(eltype(U)) - tiny))
 end
 
-@rule NormalMixture{N}(:switch, Marginalisation) (q_out::Any, q_m::NTuple{N, PointMass{T}}, q_p::NTuple{N, PointMass{T}}) where { N, T <: Real } = begin
+@rule NormalMixture{N}(:switch, Marginalisation) (q_out::Any, q_m::NTuple{N, PointMass{T} where T <: Real}, q_p::NTuple{N, PointMass{T} where T <: Real}) where { N } = begin
     U = map(zip(q_m, q_p)) do (m, p)
         return -score(AverageEnergy(), NormalMeanPrecision, Val{ (:out, :μ, :τ) }, map((q) -> Marginal(q, false , false), (q_out, m, p)), nothing)
     end
     return Categorical(clamp.(softmax(U), tiny, one(eltype(U)) - tiny))
 end
 
-@rule NormalMixture{N}(:switch, Marginalisation) (q_out::Any, q_m::NTuple{N, PointMass}, q_p::NTuple{N, PointMass}) where { N } = begin
+@rule NormalMixture{N}(:switch, Marginalisation) (q_out::Any, q_m::NTuple{N, PointMass{T} where T <: AbstractVector}, q_p::NTuple{N, PointMass{T} where T <: AbstractMatrix}) where { N } = begin
     U = map(zip(q_m, q_p)) do (m, p)
         return -score(AverageEnergy(), MvNormalMeanPrecision, Val{ (:out, :μ, :Λ) }, map((q) -> Marginal(q, false , false), (q_out, m, p)), nothing)
     end
