### 2021-11-09 dc9ec5b
test(): tests added for out edge of gaussian mixture node
