+++ src/macrohelpers.jl
@@ -120,11 +120,15 @@ function expression_convert_eltype(eltype::Type{T}, expr::Expr) where T
     if @capture(expr, f_(args__)) 
         return :(ReactiveMP.convert_eltype($f, $T, $expr))
     elseif @capture(expr, (elems__, ))
-        entries = map(elems) do elem
-            @capture(elem, (name_ = value_)) || error("Invalid expression specification in expression_convert_eltype() function: $expr. Expression should be in the form of a constructor call or tuple of (name = value) elements.")
-            return (name, value)
+        if @capture(first(elems), (name_ = value_))
+            entries = map(elems) do elem
+                @capture(elem, (name_ = value_)) || error("Invalid expression specification in expression_convert_eltype() function: $expr. Expression should be in the form of a constructor call or tuple of (name = value) elements.")
+                return (name, value)
+            end
+            return Expr(:tuple, map((entry) -> :($(first(entry)) = $(expression_convert_eltype(eltype, last(entry)))), entries)...)
+        else 
+            return Expr(:tuple, map(elem -> :($(expression_convert_eltype(eltype, elem))), elems)...)
         end
-        return Expr(:tuple, map((entry) -> :($(first(entry)) = $(expression_convert_eltype(eltype, last(entry)))), entries)...)
     end
     error("Invalid expression specification in expression_convert_eltype() function: $expr. Expression should be in the form of a constructor call or tuple of (name = value) elements.")
 end
