### 2021-11-11 6bcff48
generalize update rule gaussian
