+++ src/nodes/flow/flow_models/flow_model.jl
@@ -128,7 +128,7 @@ nr_params(layers::Tuple{}) = return 0
 
 
 # forward pass through the Flow model
-function _forward(model::CompiledFlowModel, input::Vector{T1}) where { T1 <: Real }
+function _forward(model::CompiledFlowModel, input::AbstractVector{T1}) where { T1 <: Real }
 
     # promote type for allocating output
     T = promote_type(eltype(model), T1)
@@ -145,13 +145,13 @@ function _forward(model::CompiledFlowModel, input::Vector{T1}) where { T1 <: Rea
 end
 
 # when calling forward, redirect to _forward
-forward(model::CompiledFlowModel, input::Vector{T}) where { T <: Real } = _forward(model, input)
+forward(model::CompiledFlowModel, input::AbstractVector{T}) where { T <: Real } = _forward(model, input)
 
 # for broadcasting over forward, fix the model for multiple inputs
-Broadcast.broadcasted(::typeof(forward), model::CompiledFlowModel, input::Vector{Vector{T}}) where { T <: Real } = broadcast(_forward, Ref(model), input)
+Broadcast.broadcasted(::typeof(forward), model::CompiledFlowModel, input::AbstractVector{AbstractVector{T}}) where { T <: Real } = broadcast(_forward, Ref(model), input)
 
 # inplace forward pass through the Flow model
-function forward!(output::Vector{T1}, model::CompiledFlowModel, input::Vector{T2}) where { T1 <: Real, T2 <: Real }
+function forward!(output::AbstractVector{T1}, model::CompiledFlowModel, input::AbstractVector{T2}) where { T1 <: Real, T2 <: Real }
 
     # fetch layers
     layers = getlayers(model)
@@ -169,7 +169,7 @@ function forward!(output::Vector{T1}, model::CompiledFlowModel, input::Vector{T2
 end
 
 # inplace forward pass through a tuple of layers
-function forward!(output::Vector{T1}, layers::T, input::Vector{T2}) where { T1 <: Real, T2 <: Real,  T <: NTuple{N,AbstractLayer} where N}
+function forward!(output::AbstractVector{T1}, layers::T, input::AbstractVector{T2}) where { T1 <: Real, T2 <: Real,  T <: NTuple{N,AbstractLayer} where N}
 
     # perform pass over first layer
     forward!(output, first(layers), input)
@@ -183,11 +183,11 @@ function forward!(output::Vector{T1}, layers::T, input::Vector{T2}) where { T1 <
 end
 
 # when no layers are left, stop the inplace recursion
-forward!(output::Vector{T1}, layers::Tuple{}, input::Vector{T2}) where { T1 <: Real, T2 <: Real } = return
+forward!(output::AbstractVector{T1}, layers::Tuple{}, input::AbstractVector{T2}) where { T1 <: Real, T2 <: Real } = return
 
 
 # backward pass through the Flow model
-function _backward(model::CompiledFlowModel, output::Vector{T1}) where { T1 <: Real }
+function _backward(model::CompiledFlowModel, output::AbstractVector{T1}) where { T1 <: Real }
 
     # promote type for allocating output
     T = promote_type(eltype(model), T1)
@@ -204,13 +204,13 @@ function _backward(model::CompiledFlowModel, output::Vector{T1}) where { T1 <: R
 end
 
 # when calling backward, redirect to _backward
-backward(model::CompiledFlowModel, output::Vector{T}) where { T <: Real } = _backward(model, output)
[…]
