+++ src/nodes/flow/layers/permutation_layer.jl
@@ -63,7 +63,7 @@ eltype(layer::PermutationLayer{T})  where { T }              = eltype(T)
 eltype(::Type{PermutationLayer{T}}) where { T }              = eltype(T)
 
 # forward pass through the permutation layer
-function _forward(layer::PermutationLayer, input::Array{T,1}) where { T <: Real } 
+function _forward(layer::PermutationLayer, input::AbstractVector{ <: Real })
 
     # fetch variables
     P = getP(layer)
@@ -75,11 +75,11 @@ function _forward(layer::PermutationLayer, input::Array{T,1}) where { T <: Real 
     return result
     
 end
-forward(layer::PermutationLayer, input::Array{T,1}) where { T <: Real } = _forward(layer, input)
-Broadcast.broadcasted(::typeof(forward), layer::PermutationLayer, input::Array{Array{T,1},1}) where { T <: Real } = broadcast(_forward, Ref(layer), input)
+forward(layer::PermutationLayer, input::AbstractVector{ <: Real }) = _forward(layer, input)
+Broadcast.broadcasted(::typeof(forward), layer::PermutationLayer, input::AbstractVector{ <: AbstractVector{ <: Real } }) = broadcast(_forward, Ref(layer), input)
 
 # inplace forward pass through the permutation layer
-function forward!(output::Array{T1,1}, layer::PermutationLayer, input::Array{T2,1}) where { T1 <: Real, T2 <: Real }
+function forward!(output::AbstractVector{ <: Real }, layer::PermutationLayer, input::AbstractVector{ <: Real })
 
     # fetch variables
     P = getP(layer)
@@ -90,7 +90,7 @@ function forward!(output::Array{T1,1}, layer::PermutationLayer, input::Array{T2,
 end
 
 # backward pass through the permutation layer
-function _backward(layer::PermutationLayer, output::Array{T,1}) where { T <: Real }
+function _backward(layer::PermutationLayer, output::AbstractVector{ <: Real })
 
     # fetch variables
     P = getP(layer)
@@ -102,11 +102,11 @@ function _backward(layer::PermutationLayer, output::Array{T,1}) where { T <: Rea
     return result
     
 end
-backward(layer::PermutationLayer, output::Array{T,1}) where { T <: Real } = _backward(layer, output)
-Broadcast.broadcasted(::typeof(backward), layer::PermutationLayer, output::Array{Array{T,1},1}) where { T <: Real } = broadcast(_backward, Ref(layer), output)
+backward(layer::PermutationLayer, output::AbstractVector{ <: Real }) = _backward(layer, output)
+Broadcast.broadcasted(::typeof(backward), layer::PermutationLayer, output::AbstractVector{ <: AbstractVector{ <: Real } }) = broadcast(_backward, Ref(layer), output)
 
 # inplace backward pass through the additive coupling layer
-function backward!(input::Array{T1,1}, layer::PermutationLayer, output::Array{T2,1}) where { T1 <: Real, T2 <: Real }
+function backward!(input::AbstractVector{ <: Real }, layer::PermutationLayer, output::AbstractVector{ <: Real })
 
     # fetch variables
     P = getP(layer)
@@ -117,40 +117,40 @@ function backward!(input::Array{T1,1}, layer::PermutationLayer, output::Array{T2
 end
 
 # jacobian of the additive coupling layer
-function _jacobian(layer::PermutationLayer, input::Array{T1,1}) where { T1 <: Real }
+function _jacobian(layer::PermutationLayer, input::AbstractVector{ <: Real })
 
     # return result
     return getP(layer)
     
 end
-jacobian(layer::PermutationLayer, input::Array{T,1}) where { T <: Real } = _jacobian(layer, input)
[…]
