+++ src/nodes/flow/flow_models/flow_model.jl
@@ -148,7 +148,7 @@ end
 forward(model::CompiledFlowModel, input::AbstractVector{T}) where { T <: Real } = _forward(model, input)
 
 # for broadcasting over forward, fix the model for multiple inputs
-Broadcast.broadcasted(::typeof(forward), model::CompiledFlowModel, input::AbstractVector{AbstractVector{T}}) where { T <: Real } = broadcast(_forward, Ref(model), input)
+Broadcast.broadcasted(::typeof(forward), model::CompiledFlowModel, input::AbstractVector{ <: AbstractVector{ <: Real } }) = broadcast(_forward, Ref(model), input)
 
 # inplace forward pass through the Flow model
 function forward!(output::AbstractVector{T1}, model::CompiledFlowModel, input::AbstractVector{T2}) where { T1 <: Real, T2 <: Real }
@@ -207,7 +207,7 @@ end
 backward(model::CompiledFlowModel, output::AbstractVector{T}) where { T <: Real } = _backward(model, output)
 
 # for broadcasting over backward, fix the model for multiple inputs
-Broadcast.broadcasted(::typeof(backward), model::CompiledFlowModel, output::AbstractVector{AbstractVector{T}}) where { T <: Real } = broadcast(_backward, Ref(model), output)
+Broadcast.broadcasted(::typeof(backward), model::CompiledFlowModel, output::AbstractVector{ <: AbstractVector{ <: Real } }) = broadcast(_backward, Ref(model), output)
 
 # inplace backward pass through the Flow model
 function backward!(input::AbstractVector{T1}, model::CompiledFlowModel, output::AbstractVector{T2}) where { T1 <: Real, T2 <: Real }
@@ -273,7 +273,7 @@ end
 forward_jacobian(model::CompiledFlowModel, input::AbstractVector{T}) where { T <: Real } = _forward_jacobian(model, input)
 
 # for broadcasting over forward_jacobian, fix the model for multiple inputs
-Broadcast.broadcasted(::typeof(forward_jacobian), model::CompiledFlowModel, input::AbstractVector{AbstractVector{T}}) where { T <: Real } = broadcast(_jacobian, Ref(model), input)
+Broadcast.broadcasted(::typeof(forward_jacobian), model::CompiledFlowModel, input::AbstractVector{ <: AbstractVector{ <: Real } }) = broadcast(_jacobian, Ref(model), input)
 
 # inplace forward_jacobian of the Flow model
 function forward_jacobian!(output::AbstractVector{T1}, J::Matrix{T2}, model::CompiledFlowModel, input::AbstractVector{T3}) where { T1 <: Real, T2 <: Real, T3 <: Real }
@@ -370,7 +370,7 @@ end
 backward_inv_jacobian(model::CompiledFlowModel, output::AbstractVector{T}) where { T <: Real } = _backward_inv_jacobian(model, output)
 
 # for broadcasting over backward inverse jacobian, fix the model for multiple inputs
-Broadcast.broadcasted(::typeof(backward_inv_jacobian), model::CompiledFlowModel, output::AbstractVector{AbstractVector{T}}) where { T <: Real } = broadcast(_backward_inv_jacobian, Ref(model), output)
+Broadcast.broadcasted(::typeof(backward_inv_jacobian), model::CompiledFlowModel, output::AbstractVector{ <: AbstractVector{ <: Real } }) = broadcast(_backward_inv_jacobian, Ref(model), output)
 
 # inplace inverse backward jacobian of the Flow model
[…]

[…]
