+++ src/nodes/flow/flow_models/flow_model.jl
@@ -128,13 +128,13 @@ nr_params(layers::Tuple{}) = return 0
 
 
 # forward pass through the Flow model
-function _forward(model::CompiledFlowModel, input::AbstractVector{T1}) where { T1 <: Real }
+function _forward(model::CompiledFlowModel, input::AbstractVector{T}) where { T <: Real }
 
     # promote type for allocating output
-    T = promote_type(eltype(model), T1)
+    Ti = promote_type(eltype(model), T)
 
     # allocate space for result
-    output = zeros(T, size(input))
+    output = zeros(Ti, size(input))
 
     # perform calculations 
     forward!(output, model, input)
@@ -145,22 +145,22 @@ function _forward(model::CompiledFlowModel, input::AbstractVector{T1}) where { T
 end
 
 # when calling forward, redirect to _forward
-forward(model::CompiledFlowModel, input::AbstractVector{T}) where { T <: Real } = _forward(model, input)
+forward(model::CompiledFlowModel, input::AbstractVector{ <: Real }) = _forward(model, input)
 
 # for broadcasting over forward, fix the model for multiple inputs
 Broadcast.broadcasted(::typeof(forward), model::CompiledFlowModel, input::AbstractVector{ <: AbstractVector{ <: Real } }) = broadcast(_forward, Ref(model), input)
 
 # inplace forward pass through the Flow model
-function forward!(output::AbstractVector{T1}, model::CompiledFlowModel, input::AbstractVector{T2}) where { T1 <: Real, T2 <: Real }
+function forward!(output::AbstractVector{ <: Real }, model::CompiledFlowModel, input::AbstractVector{T}) where { T <: Real }
 
     # fetch layers
     layers = getlayers(model)
 
     # promote type for allocating output
-    T = promote_type(eltype(model), T2)
+    Ti = promote_type(eltype(model), T)
 
     # decouple changing input from actual input
-    input_new = zeros(T, size(input))
+    input_new = zeros(Ti, size(input))
     input_new .= input
 
     # perform forward pass over all layers
@@ -169,7 +169,7 @@ function forward!(output::AbstractVector{T1}, model::CompiledFlowModel, input::A
 end
 
 # inplace forward pass through a tuple of layers
-function forward!(output::AbstractVector{T1}, layers::T, input::AbstractVector{T2}) where { T1 <: Real, T2 <: Real,  T <: NTuple{N,AbstractLayer} where N}
+function forward!(output::AbstractVector{ <: Real }, layers::T, input::AbstractVector{ <: Real }) where { T <: NTuple{N,AbstractLayer} where N}
 
     # perform pass over first layer
     forward!(output, first(layers), input)
@@ -183,17 +183,17 @@ function forward!(output::AbstractVector{T1}, layers::T, input::AbstractVector{T
 end
 
 # when no layers are left, stop the inplace recursion
-forward!(output::AbstractVector{T1}, layers::Tuple{}, input::AbstractVector{T2}) where { T1 <: Real, T2 <: Real } = return
+forward!(output::AbstractVector{ <: Real }, layers::Tuple{}, input::AbstractVector{ <: Real }) = return
 
 
 # backward pass through the Flow model
-function _backward(model::CompiledFlowModel, output::AbstractVector{T1}) where { T1 <: Real }
+function _backward(model::CompiledFlowModel, output::AbstractVector{T}) where { T <: Real }
 
     # promote type for allocating output
[…]
