### 2021-11-11 0e88df9
update length definiton of permutation matrix
