### 2021-11-11 29f246b
use flatten from TupleTools
