### 2021-11-11 5ed8820
update helper functions
