+++ src/nodes/flow/layers/additive_coupling_layer.jl
@@ -1,11 +1,6 @@
 ## TODO: fix for non-unity partition dimensions
 
 export AdditiveCouplingLayer
-export getf, getflow, getdim
-export forward, forward!, backward, backward!
-export jacobian, inv_jacobian
-export det_jacobian, absdet_jacobian, logdet_jacobian, logabsdet_jacobian
-export detinv_jacobian, absdetinv_jacobian, logdetinv_jacobian, logabsdetinv_jacobian
 
 import Base: eltype
 
