+++ test/nodes/flow/coupling_flows/test_planar_flow.jl
@@ -1,7 +1,11 @@
 module FlowNodeCouplingFlowsPlanarFlowTest
 
 using Test
-using ReactiveMP 
+using ReactiveMP
+using ReactiveMP: getdim, getu, getb, getw, getall, setu!, setb!, setw!
+using ReactiveMP: forward, forward!, jacobian, jacobian!, inv_jacobian, inv_jacobian!
+using ReactiveMP: det_jacobian, absdet_jacobian, logdet_jacobian, logdet_jacobian, logabsdet_jacobian
+
 
 @testset "Planar Flow" begin
 
