+++ src/rules/flow/marginals.jl
@@ -48,7 +48,7 @@ end
 
     # calculate sigma points/vectors    
     sqrtΣ = sqrt((L + λ)*Σ_out)
-    χ = Vector{Vector{Float64}}(undef, 2*L + 1)
+    χ = Vector{Vector{T}}(undef, 2*L + 1)
     for k = 1:length(χ)
         χ[k] = copy(μ_out)
     end
