### 2021-11-12 dd42e13
fix(): use softmax from StatsFuns
