+++ .github/workflows/ci.yml
@@ -37,6 +37,8 @@ jobs:
             ${{ runner.os }}-test-
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@latest
+        env:
+            PYTHON: ""
       - uses: julia-actions/julia-runtest@latest
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
@@ -79,11 +81,15 @@ jobs:
       - uses: julia-actions/setup-julia@v1
         env:
           PYTHON: ""
+      - uses: julia-actions/julia-buildpkg@latest
+        env:
+            PYTHON: ""
         with:
           version: '1.6'
       - run: |
           julia --project=docs -e '
             using Pkg
+            Pkg.build("PyCall")
             Pkg.develop(PackageSpec(path=pwd()))
             Pkg.instantiate()'
       - run: julia --project=docs docs/make.jl
