+++ .github/workflows/ci.yml
@@ -89,9 +89,9 @@ jobs:
       - run: |
           julia --project=docs -e '
             using Pkg
-            Pkg.build("PyCall")
             Pkg.develop(PackageSpec(path=pwd()))
-            Pkg.instantiate()'
+            Pkg.instantiate()
+            Pkg.build("PyCall")'
       - run: julia --project=docs docs/make.jl
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
