+++ .github/workflows/ci.yml
@@ -79,20 +79,18 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
-        env:
-          PYTHON: ""
-      - uses: julia-actions/julia-buildpkg@latest
-        env:
-            PYTHON: ""
+      - uses: julia-actions/julia-buildpkg@latest          
         with:
           version: '1.6'
       - run: |
           julia --project=docs -e '
+            ENV["PYTHON"]=""
             using Pkg
             Pkg.develop(PackageSpec(path=pwd()))
             Pkg.instantiate()
-            Pkg.build("PyCall")'
+            Pkg.build("PyPlot")'
       - run: julia --project=docs docs/make.jl
         env:
+          PYTHON: ""
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
