+++ docs/Project.toml
@@ -11,6 +11,7 @@ PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 ReactiveMP = "a194aa59-28ba-4574-a09c-4a745416d6e3"
 Rocket = "df971d30-c9d6-4b37-b8ff-e965b2cb3a40"
+StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 
 [compat]
 Documenter = "0.27.7"
