### 2021-11-16 #91 @matbesancon
I am not sure why this happens, you're the second one to have the issue. In the meantime, the hack @gdalle found was to copy the checklist to their own comment and edit it. Will let you know when we find a fix
