### 2021-11-18 #91 @torfjelde
* [ ]  **Repository:** Is the source code for this software available at the [repository url](https://github.com/biaslab/ReactiveMP.jl)?

* [ ]  **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
