### 2021-11-18 #91 @torfjelde
Btw, I opened an issue in the repo regarding the benchmark comparison to Turing.jl. It doesn't make the claims in the paper incorrect (the perf-diff is still an the order of magnitudes on conjugate models), so no worries wrt. the paper submission :+1:
