### 2021-11-18 #91 @bvdmitri
Yes! I saw it and that is very useful! My laptop is in the office - I will definitely rerun these benchmarks tomorrow! I PM'ed you in Julia Slack regarding other models I use for benchmarking (they are not present in the main repository, but in my private one).
