### 2021-11-19 #91 @bvdmitri
Dear @torfjelde , first of all thank you for reviewing our work! We appreciate your effort and time you spent on it.
I will just leave here a couple of remarks about your comments in https://github.com/JuliaCon/proceedings-review/issues/91#issuecomment-972963611 .

I do think we can improve both paper content and documentation a bit based on your comments.

> all existing factor graph frameworks [...] use preset message sequence schedules

This is an indeed quite a statement without a context. You correctly assumed that we meant "As far as the authors are aware". We can definitely change it in our final version.

> The comparison to a general-purpose PPL, e.g. Turing.jl, might give the impression that ReactiveMP.jl supports a larger family of models than it really does. Though it's good that the performance claim is limited to conjugate models:

Hm, that might be true. I'm not sure we explicitly outline somewhere in our documentation the scope of the package. This is due to the fact that `ReactiveMP.jl` is a successor of `ForneyLab.jl`. `ForneyLab.jl` does support inference in basically most of models that Turing.jl does with the help of nonlinear node and sampling. `ReactiveMP.jl` is supposed to be sort of `ForneyLab.jl 2.0`, but we didn't port and didn't reimplement all of the existing functionality from `ForneyLab.jl`. That is true that currently ReactiveMP.jl is limited to a specific family of models, namely conjugate state space models. More or less representative list of supported models (not all) can be found in `demo/` folder (but you've probably already seen it).

I think we could be more explicit here! I can change it in README and repo documentation. What would you prefer? Would like me to outline it in the paper content as well? It is not a problem :)

Regarding performance claims, we test only models we can test ¯\_(ツ)_/¯. Since in most of our models we use analytical solutions for messages and we do not use sampling I think we are guaranteed to be faster than sampling-based solutions (though it fallbacks again to the discussion that we do not support all models).

> In general, any valid snippet of Julia code can be used inside the @model block.

Ah, I see now that this statement might be a bit confusing. Here we mean that you can use `if` statements, `for` statements, you can define `lambda`-functions in your model etc.. For example `Soss.jl` does not allow `for` blocks. We did not want to imply that inference will run without any errors in this statement. We may change this to be more clear.

> unless one has defined nodes for each of the operations present at the top-level

You understood it correctly. Package consists of a set of standard nodes and message computations rules. We are constantly adding new nodes and rules, but it takes some time :) Currently if something does not work you either get a "missing node" error or "missing rule". It does not necessarily mean that package cannot run inference at all, but maybe we just missed some rules or still looking for a time to implement it.

Package is very young in that sense. We have another PhD project that is aimed to expand the scope of the package to more models and reduce number of such "missing something" errors. We also have plans to integrate with Turing.jl and Soss.jl at some point.

> Finally, though I recognize that this would probably require quite a bit of effort/time: it would be really awesome to have comparisons to other software based on message-passing from outside the Julia community, e.g. Infer.NET.

That is true, I actually have comparison with Infer.Net in my private repos. The problem here is that I'm not an expert in C# at all. In addition, model specifications in C# are huge - they don't have very nice `@macro` lambdas :).  I wrote a lot of code just for a setup and to make it look like in Julia and I'm not sure I did everything correct. Infer.NET gives reasonably performance btw (on models I managed to test), it is somewhere in between ReactiveMP and Turing.

Let me know what do you think and I will adjust the paper's content a bit. Thanks again for your time and comments!
