### 2021-11-19 cf84d3f
docs(): Update LGSSM benchmark and documentation introduction
