### 2021-11-19 #91 @torfjelde
> This is an indeed quite a statement without a context. You correctly assumed that we meant "As far as the authors are aware". We can definitely change it in our final version.

Perfect!

> Hm, that might be true. I'm not sure we explicitly outline somewhere in our documentation the scope of the package. This is due to the fact that ReactiveMP.jl is a successor of ForneyLab.jl. ForneyLab.jl does support inference in basically most of models that Turing.jl does with the help of nonlinear node and sampling. ReactiveMP.jl is supposed to be sort of ForneyLab.jl 2.0, but we didn't port and didn't reimplement all of the existing functionality from ForneyLab.jl. That is true that currently ReactiveMP.jl is limited to a specific family of models, namely conjugate state space models. More or less representative list of supported models (not all) can be found in demo/ folder (but you've probably already seen it).

> EDIT: Ah, we do mention it in our README, but not in the documentation. I will change it as well.

> I think we could be more explicit here! I can change it in README and repo documentation. What would you prefer? Would like me to outline it in the paper content as well? It is not a problem :)

Ah I missed the part in the README! And indeed, the statements in the README are exactly what I had in mind! And regarding the paper: as I mentioned above, just a brief sentence or even half a sentence about the restrictions ReactiveMP.jl would make the presentation even better. Something along the lines of "Even though ReactiveMP.jl does not currently have support for _all_ models supported by more general-purpose PPLs, for the ones we do have support the performance is superior" (not _exactly_ this statement, but you get the idea).

And because I feel the need to say this: I would say the exact same thing if Turing.jl wasn't even mentioned, e.g. if you were comparing to Stan instead. It's more just that for someone not super-familiar with the topic, reading the paper one could possibly walk away with the impression that "ReactiveMP.jl supports the same kind of models as Turing.jl/Stan, and ReactiveMP.jl is faster on conjugate models", and then a user installs ReactiveMP.jl and tries something rather trivial such as:

```julia
mymul(a, b) = a * b # or any user-defined function

@model function linear_gaussian_ssm_smoothing(n, A, B, P, Q)
    ...
    for t in 2:n
        x[t] ~ MvGaussianMeanCovariance(mymul(A, x[t - 1]), P)
        y[t] ~ MvGaussianMeanCovariance(mymul(B, x[t]), Q)
    end

    return x, y
end
```

and then

```julia
model, vars = linear_gaussian_ssm_smoothing(n, A, B, P, Q);
```

results in

```
Unknown functional form 'mymul' used for node specification.

[...]
```

In contrast, in a general purpose PPL, e.g. Turing.jl or (AFAIK) Gen.jl, the above would _just work_.

But, as I said, I'm probably overly sensitive to this because of my previous experience interacting with the community; I've often found myself saying "Ehh for this problem Turing.jl is probably not the right choice; you want something more specialized." (and I'll be sure to recommend ReactiveMP.jl for models such as Linear SSM in high-perf-required settings going forward) or people just being generally quite confused as to why there are so many PPLs in Julia because we're not being clear enough regarding our pros and cons, and so I find myself gravitating towards being a bit clearer regarding this. (We can also definitively do a better job of this in Turing.jl!)

> Regarding performance claims, we test only models we can test ¯_(ツ)_/¯. Since in most of our models we use analytical solutions for messages and we do not use sampling I think we are guaranteed to be faster than sampling-based solutions (though it fallbacks again to the discussion that we do not support all models).

I think you performance claims are completely fine! My previous comment was _endorsing_ the fact that you explicitly mention you're comparing conjugacy models:)

> Ah, I see now that this statement might be a bit confusing. Here we mean that you can use if statements, for statements, you can define lambda-functions in your model etc.. For example Soss.jl does not allow for blocks. We did not want to imply that inference will run without any errors in this statement. We may change this to be more clear.

Sounds good!

> Package is very young in that sense. We have another PhD project that is aimed to expand the scope of the package to more models and reduce number of such "missing something" errors. We also have plans to integrate with Turing.jl and Soss.jl at some point.

As mentioned in DMs, I'm very much looking forward to this!

> That is true, I actually have comparison with Infer.Net in my private repos. The problem here is that I'm not an expert in C# at all. In addition, model specifications in C# are huge - they don't have very nice @macro lambdas :). I wrote a lot of code just for a setup and to make it look like in Julia and I'm not sure I did everything correct. Infer.NET gives reasonably performance btw (on models I managed to test), it is somewhere in between ReactiveMP and Turing.

Yeah I figured this was probably the issue, hence why I said "though I realize this will be probably require a fair bit of effort". So this is completely fine!
