### 2021-11-23 21331a0
CompatHelper: bump compat for SpecialFunctions to 2, (keep existing compat)
