+++ Project.toml
@@ -37,7 +37,7 @@ MacroTools = "0.5"
 Optim = "1.0.0"
 PositiveFactorizations = "0.2"
 Rocket = "1.3.10"
-SpecialFunctions = "1.4"
+SpecialFunctions = "1.4, 2"
 StaticArrays = "1.2"
 StatsBase = "0.33"
 StatsFuns = "0.9"
