### 2021-11-23 #91 @bvdmitri
@torfjelde I've update my paper based on your comments, you can quickly see changes here: [commit1](https://github.com/biaslab/ReactiveMP.jl/commit/77c786b982276c112b2ddedc31115948ec5539dd), [commit2](https://github.com/biaslab/ReactiveMP.jl/commit/35b298e25a30f09f0cad551ff99645ee7c740140).
