### 2021-11-23 #91 @bvdmitri
@matbesancon

Do we have any deadlines for JuliaCon2021 submission process? I would like to submit a full article about `ReactiveMP` (which goes very deep into details about reactive message passing and is very different from this submission) in the JMLR somewhere before January. They do have the following statement in their submission guide:

> Submissions to JMLR cannot have been published previously in any other journal. We will consider submissions that have been published at workshops or conferences. In these cases, we expect the JMLR submission to cite the prior work, go into much greater depth and to extend the published results in a substantive way.
...
We will also consider concurrent submissions of papers that are under review at conferences, provided the conference explicitly allows for this.
...

So I either need to cite the prior work (which is this paper) or I need an explicit permission from JuliaCon 2021 Proceedings for a cross-review.
