### 2021-11-28 fcbc53c
fix(): Fix broken compilation with FastGaussQuadrature 0.4.8
