+++ Project.toml
@@ -29,7 +29,7 @@ DataStructures = "0.17, 0.18"
 Distributions = "0.24, 0.25"
 DomainIntegrals = "0.3.2"
 DomainSets = "0.5.2"
-FastGaussQuadrature = "0.4"
+FastGaussQuadrature = "< 0.4.8, ^0.4.9"
 ForwardDiff = "0.10"
 LazyArrays = "0.21, 0.22"
 LoopVectorization = "0.12"
