### 2021-11-28 12d9c75
update(): Bump version to 1.2.0
