+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ReactiveMP"
 uuid = "a194aa59-28ba-4574-a09c-4a745416d6e3"
 authors = ["Dmitry Bagaev <bvdmitri@gmail.com>"]
-version = "1.1.0"
+version = "1.2.0"
 
 [deps]
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
@@ -29,7 +29,7 @@ DataStructures = "0.17, 0.18"
 Distributions = "0.24, 0.25"
 DomainIntegrals = "0.3.2"
 DomainSets = "0.5.2"
-FastGaussQuadrature = "< 0.4.8, ^0.4.9"
+FastGaussQuadrature = "0.4"
 ForwardDiff = "0.10"
 LazyArrays = "0.21, 0.22"
 LoopVectorization = "0.12"
