+++ test/distributions/test_sample_list.jl
@@ -312,7 +312,7 @@ import ReactiveMP: approximate_prod_with_sample_list
     end
 
     @testset "Iteration utilities" begin 
-        rng = MersenneTwister(123)
+        rng = MersenneTwister(1234)
 
         uni_distribution = Uniform(-10rand(rng), 10rand(rng))
 
@@ -386,7 +386,7 @@ import ReactiveMP: approximate_prod_with_sample_list
             (x = NormalMeanPrecision(3.0, 7.0), y = NormalMeanVariance(-4.0, 6.0), mean_tol = [ 1e-1, 1e-1, 1e-1 ], cov_tol = [ 1e-1, 1e-1, 1e-1 ], entropy_tol = [ 1e-1, 1e-1, 1e-1 ]),
             (x = NormalMeanVariance(3.0, 7.0), y = NormalWeightedMeanPrecision(4.0, 6.0), mean_tol = [ 1e-1, 1e-1, 1e-1 ], cov_tol = [ 1e-1, 1e-1, 1e-1 ], entropy_tol = [ 1e-1, 1e-1, 1e-1 ]),
             (x = GammaShapeRate(3.0, 7.0), y = GammaShapeScale(4.0, 6.0), mean_tol = [ 1e-1, 1e-1, 1e-1 ], cov_tol = [ 1e-1, 1e-1, 1e-1 ], entropy_tol = [ 3e-1, 3e-1, 3e-1 ]),
-            (x = MvNormalMeanCovariance(10rand(rng, 4), posdefm(rng, 4)), y = MvNormalMeanPrecision(10rand(rng, 4), posdefm(rng, 4)), mean_tol = [ 6e-1, 6e-1, 6e-1 ], cov_tol = [ 5e1, 5e1, 5e1 ], entropy_tol = [ 4e-1, 4e-1, 4e-1 ]),
+            (x = MvNormalMeanCovariance(10rand(rng, 4), posdefm(rng, 4)), y = MvNormalMeanPrecision(10rand(rng, 4), posdefm(rng, 4)), mean_tol = [ 6e-1, 6e-1, 6e-1 ], cov_tol = [ 6e1, 6e1, 6e1 ], entropy_tol = [ 4e-1, 4e-1, 4e-1 ]),
             (x = Wishart(10.0, posdefm(rng, 3)), y = Wishart(5.0, posdefm(rng, 3)), mean_tol = [ 7e-1, 7e-1, 7e-1 ], cov_tol = [ 3e2, 3e2, 3e2 ], entropy_tol = [ 1e-1, 1e-1, 1e-1 ]),
         ]
         
