### 2021-11-28 5574cd9
test(): use StableRNG is sample list tests
