### 2021-12-01 bc20148
ci(): add Julia 1.7
