### 2021-12-01 d2b2957
fix(): Fix performance regression in HGF
