+++ src/nodes/gcv.jl
@@ -19,24 +19,30 @@ const DefaultGCVNodeMetadata = GCVMetadata(GaussHermiteCubature(20))
 default_meta(::Type{ GCV }) = DefaultGCVNodeMetadata
 
 @average_energy GCV (q_y_x::MultivariateNormalDistributionsFamily, q_z::NormalDistributionsFamily, q_κ::Any, q_ω::Any) = begin
-    m, c = mean(q_y_x), cov(q_y_x)
+    y_x_mean, y_x_cov = mean_cov(q_y_x)
+    z_mean, z_var     = mean_var(q_z)
+    κ_mean, κ_var     = mean_var(q_κ)
+    ω_mean, ω_var     = mean_var(q_ω)
 
-    ksi = (mean(q_κ) ^ 2) * var(q_z) + (mean(q_z) ^ 2) * var(q_κ) + var(q_κ) * var(q_z)
-    psi = (m[2] - m[1]) ^ 2 + c[1, 1] + c[2, 2] - c[1, 2] - c[2, 1]
-    A = exp(-mean(q_ω) + var(q_ω) / 2)
-    B = exp(-mean(q_κ) * mean(q_z) + ksi / 2)
+    ksi = (κ_mean ^ 2) * z_var + κ_var * ((z_mean ^ 2) + z_var)
+    psi = @inbounds (y_x_mean[2] - y_x_mean[1]) ^ 2 + y_x_cov[1, 1] + y_x_cov[2, 2] - y_x_cov[1, 2] - y_x_cov[2, 1]
+    A = exp(-ω_mean + ω_var / 2)
+    B = exp(-κ_mean * z_mean + ksi / 2)
 
-    0.5 * log2π + 0.5 * (mean(q_z) * mean(q_κ) + mean(q_ω)) + 0.5 * (psi * A * B)
+    0.5 * log2π + 0.5 * (z_mean * κ_mean + ω_mean) + 0.5 * (psi * A * B)
 end
 
-@average_energy GCV (q_y::NormalDistributionsFamily,q_x::NormalDistributionsFamily, q_z::NormalDistributionsFamily, q_κ::Any, q_ω::Any) = begin
-    my,vy = mean(q_y), cov(q_y)
-    mx,vx = mean(q_x), cov(q_x)
+@average_energy GCV (q_y::NormalDistributionsFamily, q_x::NormalDistributionsFamily, q_z::NormalDistributionsFamily, q_κ::Any, q_ω::Any) = begin
+    y_mean, y_var = mean_var(q_y)
+    x_mean, x_var = mean_var(q_x)
+    z_mean, z_var = mean_var(q_z)
+    κ_mean, κ_var = mean_var(q_κ)
+    ω_mean, ω_var = mean_var(q_ω)
 
-    ksi = (mean(q_κ) ^ 2) * var(q_z) + (mean(q_z) ^ 2) * var(q_κ) + var(q_κ) * var(q_z)
-    psi = (my - mx) ^ 2 + vy+vx
-    A = exp(-mean(q_ω) + var(q_ω) / 2)
-    B = exp(-mean(q_κ) * mean(q_z) + ksi / 2)
+    ksi = (κ_mean ^ 2) * z_var + (z_mean ^ 2) * κ_var + κ_var * z_var
+    psi = (y_mean - x_mean) ^ 2 + y_var+x_var
+    A = exp(-ω_mean + ω_var / 2)
+    B = exp(-κ_mean * z_mean + ksi / 2)
 
-    0.5 * log2π + 0.5 * (mean(q_z) * mean(q_κ) + mean(q_ω)) + 0.5 * (psi * A * B)
+    0.5 * log2π + 0.5 * (z_mean * κ_mean + ω_mean) + 0.5 * (psi * A * B)
 end
