+++ src/rules/gcv/k.jl
@@ -7,7 +7,7 @@ export rule
     z_mean, z_var     = mean_var(q_z)
 
     A = exp(-ω_mean + ω_var / 2)
-    psi = (y_x_mean[1] - y_x_mean[2]) ^ 2 + y_x_var[1, 1] + y_x_var[2, 2] - y_x_var[1, 2] - y_x_var[2, 1]
+    psi = @inbounds (y_x_mean[1] - y_x_mean[2]) ^ 2 + y_x_var[1, 1] + y_x_var[2, 2] - y_x_var[1, 2] - y_x_var[2, 1]
 
     a = z_mean
     b = psi * A
