### 2021-12-03 05e5747
update(): Bump version to 1.2.1
