### 2021-12-04 7438811
add optimized code for scalar outer product
