### 2021-12-04 d3e776f
optimize dot product rules
