+++ src/rules/dot_product/out.jl
@@ -6,5 +6,5 @@ end
 @rule typeof(dot)(:out, Marginalisation) (m_in1::PointMass, m_in2::NormalDistributionsFamily, meta::AbstractCorrection) = begin 
     A = mean(m_in1)
     in2_mean, in2_cov = mean_cov(m_in2)
-    return NormalMeanVariance(dot(A, in2_mean), A' * in2_cov * A)
+    return NormalMeanVariance(dot(A, in2_mean), dot(A, in2_cov, A))
 end
