+++ src/algebra/standard_basis_vector.jl
@@ -1,0 +1,209 @@
+export StandardBasisVector
+
+import Base: *
+import LinearAlgebra: dot, transpose, mul!
+
+using LinearAlgebra
+
+@doc raw"""
+`StandardBasisVector{len, ind}()` creates a standard Cartesian basis vector of zeros of length `len` with a single one at index `ind`.
+
+An example is the 3-dimensional standard basis vector for the first dimension
+```math
+e = \begin{bmatrix} 1 \\ 0 \\ 0 \end{bmatrix}
+```
+Which can be constructed by calling `e = StandardBasisVector(3,1)`
+"""
+struct StandardBasisVector{N, T} <: AbstractVector{T}
+    function StandardBasisVector(len::Int, ind::Int)
+        @assert ind >= 0 && len >= 1 && ind <= len
+        return new{len, ind}()
+    end
+end
+
+# extensions of base functionality
+Base.size(e::StandardBasisVector{N, T})    where { N, T } = (N,)
+Base.size(e::StandardBasisVector{N, T}, d) where { N, T } = d::Integer == 1 ? N : 1
+Base.length(e::StandardBasisVector{N, T})  where { N, T } = N
+
+function Base.getindex(e::StandardBasisVector{N,T}, i::Int) where { N, T }
+    @assert i > 0 && i <= N
+    if T == i
+        return 1
+    else 
+        return 0
+    end
+end
+
+function Base.show(io::IO, ::MIME"text/plain", ::StandardBasisVector{N,T}) where { N, T }
+    if N < 10
+        x = zeros(N)
+        x[T] = 1
+        print(io, x)
+    else
+        if T == 1
+            print(io, "[1, 0, 0, ...]")
+        elseif T == 2
+            print(io, "[0, 1, 0, ...]")
+        else
+            print(io, "[..., 0, 1, 0, ...]\n         ^\n         $T")
+        end
+    end
+end
+function Base.show(io::IO, ::StandardBasisVector{N,T}) where { N, T }
+    if N < 10
+        x = zeros(N)
+        x[T] = 1
+        print(io, x)
+    else
+        compact = get(io, :compact, false)
+        if T == 1
+            print(io, "[1, 0, 0, ...]")
+        elseif T == 2
+            print(io, "[0, 1, 0, ...]")
+        else
+            if compact
+                print(io, "[..., 0, 1, 0, ...]\n         ^\n         $T")
+            else
+                print(io, "[0, ..., 1 at index $T, ..., 0]")
+            end
+        end
+    end
+end
+
+# get index function
+getind(e::StandardBasisVector{N, T}) where { N, T }= T
+LinearAlgebra.adjoint(e::StandardBasisVector{N, T}) where { N, T } = Adjoint{Int64, StandardBasisVector{N, T}}(e)
+
+# standard basis vector - scalar
+function Base.:*(e::StandardBasisVector{N,T}, x::T1) where { N, T, T1 <: Real}
+    y = zeros(T1, N)
+    y[T] = x
+    return y
+end
+
+function Base.:*(x::T1, e::StandardBasisVector{N,T}) where { N, T, T1 <: Real}
+    y = zeros(T1, N)
+    y[T] = x
+    return y
+end
+
+function Base.:*(::Adjoint{T2, StandardBasisVector{N,T}}, x::T1) where { N, T, T1 <: Real, T2}
+    y = zeros(T1, 1, N)
+    y[T] = x
+    return y
+end
+
[…]
