### 2021-12-04 12fd5a6
add default scalar outer product function
