### 2021-12-04 d55d7b1
Optimize average energy mvnormal
