### 2021-12-08 057d2d1
Halve memory usage of cholinv
