### 2021-12-09 517f751
add joint cholesky inv and logdet function
