### 2021-12-10 571e86f
add chollogdet function
