+++ src/algebra/cholesky.jl
@@ -20,38 +20,25 @@ chollogdet(x)           = logdet(cholesky(PositiveFactorizations.Positive, Hermi
 chollogdet(x::Diagonal) = logdet(x)
 chollogdet(x::Real)     = logdet(x)
 
-function cholinv_logdet(x::AbstractMatrix{T}) where { T <: LinearAlgebra.BlasFloat } 
+function cholinv_logdet(x) 
     # calculate cholesky decomposition
     y = cholesky(PositiveFactorizations.Positive, Hermitian(x))
-    
-    # calculate logdeterminant of cholesky decomposition
-    L = y.L
-    z = zero(eltype(y))
-    @inbounds @simd for k ∈ 1:size(L,1)
-        z += log(L[k,k])
-    end
-    z *= 2
-    
+      
     # return inverse and log-determinant
-    return inv(y), z
+    return inv(y), logdet(y)
 end
 function cholinv_logdet(x::AbstractMatrix{T}) where { T <: LinearAlgebra.BlasFloat } 
     # calculate cholesky decomposition
     y = cholesky(PositiveFactorizations.Positive, Hermitian(x))
     
     # calculate logdeterminant of cholesky decomposition
-    L = y.L
-    z = zero(eltype(y))
-    @inbounds @simd for k ∈ 1:size(L,1)
-        z += log(L[k,k])
-    end
-    z *= 2
+    ly = logdet(y)
     
     # calculate inplace inverse of A and store in y.factors
     LinearAlgebra.inv!(y)
     
     # return inverse and log-determinant
-    return y.factors, z
+    return y.factors, ly
 end
 cholinv_logdet(x::Diagonal) = Diagonal(inv.(diag(x))), mapreduce(z -> log(z), +, diag(x))
 cholinv_logdet(x::Real)     = inv(x), log(abs(x))
