### 2021-12-10 d49e918
update logdetcov functions
