### 2021-12-10 e5b9df3
make avg energy calculations more efficient
