### 2021-12-10 biaslab/ReactiveMP.jl#71 @ludgerpaehler
Mathematical Equations in README
The mathematical equations in the README are sadly not legible when using GitHub's dark mode. I.e. the equations are rendered in black, on a dark grey background.
