### 2021-12-11 biaslab/ReactiveMP.jl#71 @bvdmitri
Thanks for opening this issue! We'll take a look on it!
