### 2021-12-12 #91 @ludgerpaehler
@matbesancon @bvdmitri

An excellent package, with exemplary documentation. In reviewing the individual checkpoints the only very minor "fault" I could find was with the README, and its equation rendering under GitHub's dark mode (which should be quite easy to fix I believe), i.e. see

https://github.com/biaslab/ReactiveMP.jl/issues/71

The paper is well written, but I would recommend the addition of a few articles at some points in the paper. Going through the points in the order of appearance in the paper:

In the abstract:
> ".. of variational family of distributions"  -> "of **the** variational family of distributions"

In section 2 *Message Passing*:
> ".. we exploit reactive programming approach" -> "we exploit **the** reactive programming approach.."

In section 3 *Reactive Message Passing*:
> "We present ReactiveMP.jl package .." -> "We present **the** ReactiveMP.jl package.."

> "In our experiments new implementation scales comfortably ..." -> "In our experiments **our/this** new implementation scales comfortably.."

> ", supports real-time data inference." -> ", **and** support real-time data inference"

> "Moreover, ReactiveMP.jl API supports various processing models..." -> "Moreover, ReactiveMP.jl's API supports various processing models.."

In section 4 *Conclusions* (this is more of a style suggestion):

I would recommend to consider switching the "that" for a "which" in the following sentence.
> We developed ReactiveMP.jl as a package that enables developers to builld ... -> We developed ReactiveMP.jl as a package which enables developers to build..

With the minor README rendering "issue" resolved and the non-style suggestions included, I would recommend direct acceptance to the proceedings.
