### 2021-12-13 be70d26
use backsubstitution for parameter conversion
