### 2021-12-13 biaslab/ReactiveMP.jl#71 @bvdmitri
@ludgerpaehler  Hm, after brief googling it looks like that this issue is not easy to fix at the moment:

https://stackoverflow.com/questions/66247653/get-the-user-selected-color-theme-in-github-readme
https://github.community/t/support-theme-context-for-images-in-light-vs-dark-mode/147981/81?page=2
https://githubmemory.com/repo/greg7mdp/parallel-hashmap/issues/85
and many other people struggle with the exact same issue.

It is probably possible to upload an image in the repository with just white background so it renders in dark mode. But I'll keep it opened and try to find another solution.
