+++ src/rules/multiplication/in.jl
@@ -2,8 +2,46 @@ export rule
 
 @rule typeof(*)(:in, Marginalisation) (m_out::PointMass, m_A::PointMass) = PointMass(mean(m_in1) * mean(m_in2))
 
-@rule typeof(*)(:in, Marginalisation) (m_out::F, m_A::PointMass, meta::AbstractCorrection) where { F <: NormalDistributionsFamily } = begin
+# if A is a matrix, then the input is multivariate
+@rule typeof(*)(:in, Marginalisation) (m_out::F, m_A::PointMass{<:AbstractMatrix}, meta::AbstractCorrection) where { F <: NormalDistributionsFamily } = begin
     A = mean(m_A)
-    W = correction!(meta, A' * precision(m_out) * A)
-    return convert(promote_variate_type(F, NormalWeightedMeanPrecision), A' * weightedmean(m_out), W)
+    ξ_out, W_out = weightedmean_precision(m_out)
+    W = correction!(meta, A' * W_out * A)
+    return convert(promote_variate_type(F, NormalWeightedMeanPrecision), A' * ξ_out, W)
+end
+
+# if A is a vector, then the input is univariate
+@rule typeof(*)(:in, Marginalisation) (m_out::F, m_A::PointMass{<:AbstractVector}, meta::AbstractCorrection) where { F <: MultivariateNormalDistributionsFamily } = begin
+    A = mean(m_A)
+    ξ_out, W_out = weightedmean_precision(m_out)
+    W = correction!(meta, dot(A, W_out, A))
+    return convert(promote_variate_type(F, NormalWeightedMeanPrecision), dot(A, ξ_out), W)
+end
+
+# if A is a scalar, then the input is either univariate or multivariate
+@rule typeof(*)(:in, Marginalisation) (m_out::F, m_A::PointMass{<:Real}, meta::AbstractCorrection) where { F <: NormalDistributionsFamily } = begin
+    A = mean(m_A)
+    ξ_out, W_out = weightedmean_precision(m_out)
+    W = correction!(meta, A^2 * W_out)
+    return convert(promote_variate_type(F, NormalWeightedMeanPrecision), A * ξ_out, W)
+end
+
+# specialized versions for mean-covariance parameterization
+@rule typeof(*)(:in, Marginalisation) (m_out::F, m_A::PointMass{<:AbstractMatrix}, meta::AbstractCorrection) where { F <: MvNormalMeanCovariance } = begin
+    A = mean(m_A)
+    μ_out, Σ_out = mean_cov(m_out)
+
+    z = cholesky(PositiveFactorizations.Positive, Σ_out)
+    tmp = A' / z
+    W = correction!(meta, tmp * A)
+    return convert(promote_variate_type(F, NormalWeightedMeanPrecision), tmp *μ_out, W)
+end
+@rule typeof(*)(:in, Marginalisation) (m_out::F, m_A::PointMass{<:AbstractVector}, meta::AbstractCorrection) where { F <: MvNormalMeanCovariance } = begin
+    A = mean(m_A)
+    μ_out, Σ_out = mean_cov(m_out)
+
+    z = cholesky(PositiveFactorizations.Positive, Σ_out)
+    tmp = A' / z
+    W = correction!(meta, dot(tmp, A))
+    return convert(promote_variate_type(F, NormalWeightedMeanPrecision), dot(tmp, μ_out), W)
 end
