+++ src/rules/multiplication/out.jl
@@ -4,5 +4,6 @@ export rule
 
 @rule typeof(*)(:out, Marginalisation) (m_A::PointMass, m_in::F) where { F <: NormalDistributionsFamily } = begin
     A = mean(m_A)
-    return convert(promote_variate_type(F, NormalMeanVariance), A * mean(m_in), A * cov(m_in) * A')
+    μ_in, Σ_in = mean_cov(m_in)
+    return convert(promote_variate_type(F, NormalMeanVariance), A * μ_in, A * Σ_in * A')
 end
