### 2021-12-17 9a50b73
speed up mvnormal rules with wishart
