### 2021-12-17 43d2ce2
halve allocations in prod
