### 2021-12-21 0a74238
simplify update rules
