### 2021-12-23 3b5ca6e
feat(): extend multiplication node rules
