### 2021-12-23 a9841d4
use convolve for addition node
