+++ src/rules/addition/out.jl
@@ -1,4 +1,8 @@
 # distritbutions
+@rule typeof(+)(:out, Marginalisation) (m_in1::Distribution, m_in2::Distribution) = begin
+    return convolve(m_in1, m_in2)
+end
+
 @rule typeof(+)(:out, Marginalisation) (m_in1::UnivariateNormalDistributionsFamily, m_in2::UnivariateNormalDistributionsFamily) = begin
     min1, vin1 = mean_var(m_in1)
     min2, vin2 = mean_var(m_in2)
