### 2021-12-23 2ebf2a2
fix breaking change of multiplication update rules
