### 2021-12-23 a5f72c8
simplify rules subtraction node
