### 2022-01-03 0cf94da
optimize addition/subtraction rules
