+++ test/rules/addition/test_in1.jl
@@ -50,8 +50,8 @@ import ReactiveMP: @test_rules
             (input = (m_out = MvNormalMeanCovariance([-4.0, 3.0], [3.0 2.0; 2.0 4.0]), m_in2 = PointMass([1.0, 1.0])),  output = MvNormalMeanCovariance([-5.0, 2.0], [3.0 2.0; 2.0 4.0])),
             (input = (m_out = MvNormalMeanPrecision([-4.0, 3.0], [3.0 2.0; 2.0 4.0]),  m_in2 = PointMass([3.0, 2.0])),  output = MvNormalMeanPrecision([-7.0, 1.0], [3.0 2.0; 2.0 4.0])),
             (input = (m_out = MvNormalMeanPrecision([-4.0, 3.0], [3.0 2.0; 2.0 4.0]),  m_in2 = PointMass([-2.0, 3.0])), output = MvNormalMeanPrecision([-2.0, 0.0], [3.0 2.0; 2.0 4.0])),
-            (input = (m_out = MvNormalWeightedMeanPrecision([1.0, 3.0], [ 1.0 0.0; 0.0 1.0 ]), m_in2 = PointMass([ 2.0, 7.0 ])),  output = MvNormalMeanCovariance([ -1.0, -4.0 ], [ 1.0 0.0; 0.0 1.0 ])),
-            (input = (m_out = MvNormalWeightedMeanPrecision([2.0, 4.0], [ 2.0 0.0; 0.0 2.0 ]), m_in2 = PointMass([ 1.0, 3.0 ])),  output = MvNormalMeanCovariance([ 0.0, -1.0 ], [ 1/2 0.0; 0.0 1/2 ])),
+            (input = (m_out = MvNormalWeightedMeanPrecision([1.0, 3.0], [ 1.0 0.0; 0.0 1.0 ]), m_in2 = PointMass([ 2.0, 7.0 ])),  output = MvNormalWeightedMeanPrecision([ -1.0, -4.0 ], [ 1.0 0.0; 0.0 1.0 ])),
+            (input = (m_out = MvNormalWeightedMeanPrecision([2.0, 4.0], [ 2.0 0.0; 0.0 2.0 ]), m_in2 = PointMass([ 1.0, 3.0 ])),  output = MvNormalWeightedMeanPrecision([ 0.0, -2.0 ], [ 2.0 0.0; 0.0 2.0 ])),
         ]
 
     end
@@ -63,8 +63,8 @@ import ReactiveMP: @test_rules
             (input = (m_out = PointMass([1.0, 1.0]),    m_in2 = MvNormalMeanCovariance([-4.0, 3.0], [3.0 2.0; 2.0 4.0])),   output = MvNormalMeanCovariance([5.0, -2.0], [3.0 2.0; 2.0 4.0])),
             (input = (m_out = PointMass([1.0, 1.0]),    m_in2 = MvNormalMeanPrecision([-4.0, 3.0], [3.0 2.0; 2.0 4.0])),    output = MvNormalMeanPrecision([5.0, -2.0], [3.0 2.0; 2.0 4.0])),
             (input = (m_out = PointMass([-2.0, 1.0]),   m_in2 = MvNormalMeanPrecision([-4.0, 3.0], [3.0 2.0; 2.0 4.0])),    output = MvNormalMeanPrecision([2.0, -2.0], [3.0 2.0; 2.0 4.0])),
[…]
