+++ src/rules/addition/marginals.jl
@@ -1,10 +1,12 @@
 
 @marginalrule typeof(+)(:in1_in2) (m_out::UnivariateNormalDistributionsFamily, m_in1::UnivariateNormalDistributionsFamily, m_in2::PointMass) = begin
-    return (in1 = prod(ProdAnalytical(), NormalMeanVariance(mean(m_out) - mean(m_in2), var(m_out)), m_in1), in2 = m_in2)
+    mout, vout = mean_var(m_out)
+    return (in1 = prod(ProdAnalytical(), NormalMeanVariance(mout - mean(m_in2), vout), m_in1), in2 = m_in2)
 end
 
 @marginalrule typeof(+)(:in1_in2) (m_out::UnivariateNormalDistributionsFamily, m_in1::PointMass, m_in2::UnivariateNormalDistributionsFamily) = begin
-    return (in1 = m_in1, in2 = prod(ProdAnalytical(), NormalMeanVariance(mean(m_out) - mean(m_in1), var(m_out)), m_in2))
+    mout, vout = mean_var(m_out)
+    return (in1 = m_in1, in2 = prod(ProdAnalytical(), NormalMeanVariance(mout - mean(m_in1), vout), m_in2))
 end
 
 @marginalrule typeof(+)(:in1_in2) (m_out::MvNormalMeanCovariance, m_in1::MultivariateNormalDistributionsFamily, m_in2::PointMass) = begin
@@ -19,7 +21,9 @@ end
 
 @marginalrule typeof(+)(:in1_in2) (m_out::MvNormalWeightedMeanPrecision, m_in1::MultivariateNormalDistributionsFamily, m_in2::PointMass) = begin
     xiout, Wout = weightedmean_precision(m_out)
-    return (in1 = prod(ProdAnalytical(), MvNormalWeightedMeanPrecision(xiout - Wout*mean(m_in2), Wout), m_in1), in2 = m_in2)
+    tmp = -Wout*mean(m_in2)
+    tmp .+= xiout
+    return (in1 = prod(ProdAnalytical(), MvNormalWeightedMeanPrecision(tmp, Wout), m_in1), in2 = m_in2)
 end
 
 @marginalrule typeof(+)(:in1_in2) (m_out::MvNormalMeanCovariance, m_in1::PointMass, m_in2::MultivariateNormalDistributionsFamily) = begin
@@ -34,13 +38,35 @@ end
 
 @marginalrule typeof(+)(:in1_in2) (m_out::MvNormalWeightedMeanPrecision, m_in1::PointMass, m_in2::MultivariateNormalDistributionsFamily) = begin
     xiout, Wout = weightedmean_precision(m_out)
-    return (in1 = m_in1, in2 = prod(ProdAnalytical(), MvNormalWeightedMeanPrecision(xiout - Wout * mean(m_in1), Wout), m_in2))
+    tmp = -Wout * mean(m_in1)
+    tmp .+= xiout
+    return (in1 = m_in1, in2 = prod(ProdAnalytical(), MvNormalWeightedMeanPrecision(tmp, Wout), m_in2))
 end
 
-@marginalrule typeof(+)(:in1_in2) (m_out::NormalDistributionsFamily, m_in1::NormalDistributionsFamily, m_in2::NormalDistributionsFamily) = begin
+@marginalrule typeof(+)(:in1_in2) (m_out::UnivariateNormalDistributionsFamily, m_in1::UnivariateNormalDistributionsFamily, m_in2::UnivariateNormalDistributionsFamily) = begin
     xi_out, W_out = weightedmean_precision(m_out)
     xi_in1, W_in1 = weightedmean_precision(m_in1)
[…]
