### 2022-01-06 8b0163c
update marginal rules mvnormal
