+++ src/rules/subtraction/marginals.jl
@@ -1,10 +1,12 @@
 
 @marginalrule typeof(-)(:in1_in2) (m_out::UnivariateNormalDistributionsFamily, m_in1::UnivariateNormalDistributionsFamily, m_in2::PointMass) = begin
-    return (in1 = prod(ProdAnalytical(), NormalMeanVariance(mean(m_out) + mean(m_in2), var(m_out)), m_in1), in2 = m_in2)
+    mout, vout = mean_var(m_out)
+    return (in1 = prod(ProdAnalytical(), NormalMeanVariance(mout + mean(m_in2), voud), m_in1), in2 = m_in2)
 end
 
 @marginalrule typeof(-)(:in1_in2) (m_out::UnivariateNormalDistributionsFamily, m_in1::PointMass, m_in2::UnivariateNormalDistributionsFamily) = begin
-    return (in1 = m_in1, in2 = prod(ProdAnalytical(), NormalMeanVariance(mean(m_out) - mean(m_in1), var(m_out)), m_in2))
+    mout, vout = mean_var(m_out)
+    return (in1 = m_in1, in2 = prod(ProdAnalytical(), NormalMeanVariance(mout - mean(m_in1), vout), m_in2))
 end
 
 @marginalrule typeof(-)(:in1_in2) (m_out::MvNormalMeanCovariance, m_in1::MultivariateNormalDistributionsFamily, m_in2::PointMass) = begin
