+++ src/rules/subtraction/out.jl
@@ -1,3 +1,3 @@
-@rule typeof(-)(:out, Marginalisation) (m_in1::Any, m_in2::Any) = begin
+@rule typeof(-)(:out, Marginalisation) (m_in1::Any, m_in2::Any, meta::Any) = begin
     @call_rule typeof(+)(:in1, Marginalisation) (m_out = m_in1, m_in2 = m_in2, meta = meta)
 end
