+++ src/rules/subtraction/in1.jl
@@ -1,3 +1,3 @@
-@rule typeof(-)(:in1, Marginalisation) (m_out::Any, m_in2::Any) = begin
+@rule typeof(-)(:in1, Marginalisation) (m_out::Any, m_in2::Any, meta::Any) = begin
     @call_rule typeof(+)(:out, Marginalisation) (m_in1 = m_out, m_in2 = m_in2, meta = meta)
 end
