+++ src/algebra/cholesky.jl
@@ -3,33 +3,33 @@ export cholinv, cholsqrt
 using LinearAlgebra
 using PositiveFactorizations
 
-cholinv(x)           = inv(cholesky(PositiveFactorizations.Positive, Hermitian(x)))
+cholinv(x)           = inv(fastcholesky(x))
 cholinv(x::Diagonal) = Diagonal(inv.(diag(x)))
 cholinv(x::Real)     = inv(x)
 function cholinv(x::AbstractMatrix{T}) where { T <: LinearAlgebra.BlasFloat }
-    y = cholesky(PositiveFactorizations.Positive, Hermitian(x))
+    y = fastcholesky(x)
     LinearAlgebra.inv!(y)
     return y.factors
 end
 
-cholsqrt(x)           = Matrix(cholesky(PositiveFactorizations.Positive, Hermitian(x)).L)
+cholsqrt(x)           = Matrix(fastcholesky(x).L)
 cholsqrt(x::Diagonal) = Diagonal(sqrt.(diag(x)))
 cholsqrt(x::Real)     = sqrt(x)
 
-chollogdet(x)           = logdet(cholesky(PositiveFactorizations.Positive, Hermitian(x)))
+chollogdet(x)           = logdet(fastcholesky(x))
 chollogdet(x::Diagonal) = logdet(x)
 chollogdet(x::Real)     = logdet(x)
 
 function cholinv_logdet(x) 
     # calculate cholesky decomposition
-    y = cholesky(PositiveFactorizations.Positive, Hermitian(x))
+    y = fastcholesky(x)
       
     # return inverse and log-determinant
     return inv(y), logdet(y)
 end
 function cholinv_logdet(x::AbstractMatrix{T}) where { T <: LinearAlgebra.BlasFloat } 
     # calculate cholesky decomposition
-    y = cholesky(PositiveFactorizations.Positive, Hermitian(x))
+    y = fastcholesky(x)
     
     # calculate logdeterminant of cholesky decomposition
     ly = logdet(y)
@@ -42,3 +42,60 @@ function cholinv_logdet(x::AbstractMatrix{T}) where { T <: LinearAlgebra.BlasFlo
 end
 cholinv_logdet(x::Diagonal) = Diagonal(inv.(diag(x))), mapreduce(z -> log(z), +, diag(x))
 cholinv_logdet(x::Real)     = inv(x), log(abs(x))
+
+function fastcholesky(mat::AbstractMatrix)
+    try
+        A = copy(mat)
+        return fastcholesky!(A)
+    catch
+        return cholesky(PositiveFactorizations.Positive, Hermitian(mat))
+    end
+end
+
+function fastcholesky!(A::AbstractMatrix)
+    nbrows, nbcols = size(A)
+    @inbounds for col=1:nbcols
+        @inbounds @simd for idx in 1:col-1
+            A[col, col] -= A[col, idx]^2;
+        end
+        A[col, col] = sqrt(A[col, col])
+
+        @inbounds for row in col+1: nbrows
+            @inbounds @simd for idx in 1:col-1
+                A[row, col] -= A[row, idx]*A[col, idx]
+            end
+            A[row, col] /= A[col, col]
+        end
+    end
+    return Cholesky(LowerTriangular(A))
+end
+
+function fastcholesky!(A::AbstractMatrix{T}) where { T <: LinearAlgebra.BlasFloat }
+    # blocked version (https://arxiv.org/pdf/1812.02056.pdf)
+
+    # step size
+    s = 250
+
+    n = size(A,1)
+    z = 1
+    @inbounds for c in 1:n
+
+        if c == z + s
[…]
