### 2022-01-06 bcd19ec
update(): update basis vector impl
