+++ src/algebra/standard_basis_vector.jl
@@ -6,202 +6,217 @@ import LinearAlgebra: dot, transpose, mul!
 using LinearAlgebra
 
 @doc raw"""
-`StandardBasisVector{len, ind}()` creates a standard Cartesian basis vector of zeros of length `len` with a single one at index `ind`.
+
+    StandardBasisVector{T, len, ind}(scale::T)
+
+`StandardBasisVector` creates a standard Cartesian basis vector of zeros of length `len` with a single element `scale` at index `ind`.
 
 An example is the 3-dimensional standard basis vector for the first dimension
 ```math
 e = \begin{bmatrix} 1 \\ 0 \\ 0 \end{bmatrix}
 ```
-Which can be constructed by calling `e = StandardBasisVector(3,1)`
+Which can be constructed by calling `e = StandardBasisVector(3, 1, 1)`
 """
-struct StandardBasisVector{N, T} <: AbstractVector{T}
-    function StandardBasisVector(len::Int, ind::Int)
-        @assert ind >= 0 && len >= 1 && ind <= len
-        return new{len, ind}()
-    end
+struct StandardBasisVector{N, I, T} <: AbstractVector{T} 
+    scale :: T
 end
+
+function StandardBasisVector(N::Int, I::Int, scale::T = one(Int)) where { T <: Real }
+    @assert N >= 0 && (1 <= I <= N)
+    return StandardBasisVector{N, I, T}(scale)
+end
+
+Base.eltype(::StandardBasisVector{N, I, T})         where { N, I, T } = T
+Base.eltype(::Type{ StandardBasisVector{N, I, T} }) where { N, I, T } = T
 
 # extensions of base functionality
-Base.size(e::StandardBasisVector{N, T})    where { N, T } = (N,)
-Base.size(e::StandardBasisVector{N, T}, d) where { N, T } = d::Integer == 1 ? N : 1
-Base.length(e::StandardBasisVector{N, T})  where { N, T } = N
+Base.size(::StandardBasisVector{N})    where { N } = (N,)
+Base.size(::StandardBasisVector{N}, d) where { N } = d::Integer == 1 ? N : 1
+Base.length(::StandardBasisVector{N})  where { N } = N
 
-function Base.getindex(e::StandardBasisVector{N,T}, i::Int) where { N, T }
-    @assert i > 0 && i <= N
-    if T == i
-        return 1
-    else 
-        return 0
-    end
+Base.@propagate_inbounds function Base.getindex(e::StandardBasisVector{N, I}, i::Int) where { N, I }
+    @boundscheck checkbounds(e, i)
+    return ifelse(I === i, e.scale, zero(eltype(e)))
 end
 
-function Base.show(io::IO, ::MIME"text/plain", ::StandardBasisVector{N,T}) where { N, T }
+function Base.show(io::IO, ::MIME"text/plain", e::StandardBasisVector{N, I, T}) where { N, I, T }
     if N < 10
-        x = zeros(N)
-        x[T] = 1
+        x = zeros(T, N)
+        x[I] = e.scale
         print(io, x)
     else
-        if T == 1
-            print(io, "[1, 0, 0, ...]")
-        elseif T == 2
-            print(io, "[0, 1, 0, ...]")
+        if I == 1
+            print(io, "[$(e.scale), 0, 0, ...]")
+        elseif I == 2
+            print(io, "[0, $(e.scale), 0, ...]")
         else
[…]
