+++ test/runtests.jl
@@ -77,7 +77,9 @@ end
     end
 
     addtests("algebra/test_correction.jl")
+    addtests("algebra/test_helpers.jl")
     addtests("algebra/test_permutation_matrix.jl")
+    addtests("algebra/test_standard_basis_vector.jl")
 
     addtests("test_math.jl")
     addtests("test_helpers.jl")
