### 2022-01-06 0a179d0
test(): more tests for basis vector
