### 2022-01-07 d11fa70
test(): more tests for basis vector
