+++ src/algebra/standard_basis_vector.jl
@@ -17,29 +17,33 @@ e = \begin{bmatrix} 1 \\ 0 \\ 0 \end{bmatrix}
 ```
 Which can be constructed by calling `e = StandardBasisVector(3, 1, 1)`
 """
-struct StandardBasisVector{N, I, T} <: AbstractVector{T} 
-    scale :: T
+struct StandardBasisVector{T} <: AbstractVector{T} 
+    length :: Int
+    index  :: Int
+    scale  :: T
 end
 
-function StandardBasisVector(N::Int, I::Int, scale::T = one(Int)) where { T <: Real }
-    @assert N >= 0 && (1 <= I <= N)
-    return StandardBasisVector{N, I, T}(scale)
+function StandardBasisVector(length::Int, index::Int, scale::T = one(Int)) where { T <: Real }
+    @assert length >= 0 && (1 <= index <= length)
+    return StandardBasisVector{T}(length, index, scale)
 end
 
-Base.eltype(::StandardBasisVector{N, I, T})         where { N, I, T } = T
-Base.eltype(::Type{ StandardBasisVector{N, I, T} }) where { N, I, T } = T
+Base.eltype(::StandardBasisVector{T})         where { T } = T
+Base.eltype(::Type{ StandardBasisVector{T} }) where { T } = T
 
 # extensions of base functionality
-Base.size(::StandardBasisVector{N})    where { N } = (N,)
-Base.size(::StandardBasisVector{N}, d) where { N } = d::Integer == 1 ? N : 1
-Base.length(::StandardBasisVector{N})  where { N } = N
+Base.size(e::StandardBasisVector)    = (length(e), )
+Base.size(e::StandardBasisVector, d) = d::Integer == 1 ? length(e) : 1
+Base.length(e::StandardBasisVector)  = e.length
 
-Base.@propagate_inbounds function Base.getindex(e::StandardBasisVector{N, I}, i::Int) where { N, I }
+Base.@propagate_inbounds function Base.getindex(e::StandardBasisVector, i::Int)
     @boundscheck checkbounds(e, i)
-    return ifelse(I === i, e.scale, zero(eltype(e)))
+    return ifelse(getind(e) === i, e.scale, zero(eltype(e)))
 end
 
-function Base.show(io::IO, ::MIME"text/plain", e::StandardBasisVector{N, I, T}) where { N, I, T }
+function Base.show(io::IO, ::MIME"text/plain", e::StandardBasisVector{T}) where { T }
+    N = length(e)
+    I = getind(e)
     if N < 10
         x = zeros(T, N)
         x[I] = e.scale
@@ -55,7 +59,9 @@ function Base.show(io::IO, ::MIME"text/plain", e::StandardBasisVector{N, I, T}) 
     end
 end
 
-function Base.show(io::IO, e::StandardBasisVector{N, I, T}) where { N, I, T }
+function Base.show(io::IO, e::StandardBasisVector{T}) where { T }
+    N = length(e)
+    I = getind(e)
     if N < 10
         x = zeros(T, N)
         x[I] = e.scale
@@ -77,50 +83,47 @@ function Base.show(io::IO, e::StandardBasisVector{N, I, T}) where { N, I, T }
 end
 
 # get index function
-getind(::StandardBasisVector{N, I}) where { N, I } = I
+getind(e::StandardBasisVector) = e.index
 
 LinearAlgebra.adjoint(e::S) where { S <: StandardBasisVector } = Adjoint{eltype(S), S}(e)
 
[…]
