+++ src/nodes/autoregressive.jl
@@ -87,9 +87,10 @@ function ar_unit end
 ar_unit(::Type{ V }, order) where { V <: VariateForm } = ar_unit(Float64, V, order)
 
 function ar_unit(::Type{T}, ::Type{Multivariate}, order) where { T <: Real }
-    c    = zeros(T, order)
-    c[1] = one(T)
-    return c
+    return StandardBasisVector(order, 1, one(T))
+    # c    = zeros(T, order)
+    # c[1] = one(T)
+    # return c
 end
 
 function ar_unit(::Type{T}, ::Type{Univariate}, order) where { T <: Real }
