### 2022-01-07 9a696d4
test(): tests for average energy
