+++ test/nodes/test_normal_mean_precision.jl
@@ -1,0 +1,109 @@
+module NormalMeanPrecisionNodeTest
+
+using Test
+using ReactiveMP
+using Random
+
+import ReactiveMP: make_node
+
+@testset "NormalMeanPrecisionNode" begin
+
+    @testset "Creation" begin
+        node = make_node(NormalMeanPrecision)
+
+        @test functionalform(node)          === NormalMeanPrecision
+        @test sdtype(node)                  === Stochastic()
+        @test name.(interfaces(node))       === (:out, :μ, :τ)
+        @test factorisation(node)           === ((1, 2, 3), )
+        @test localmarginalnames(node)      === (:out_μ_τ, )
+        @test metadata(node)                === nothing
+
+        node = make_node(NormalMeanPrecision, meta = 1)
+
+        @test functionalform(node)          === NormalMeanPrecision
+        @test sdtype(node)                  === Stochastic()
+        @test name.(interfaces(node))       === (:out, :μ, :τ)
+        @test factorisation(node)           === ((1, 2, 3), )
+        @test localmarginalnames(node)      === (:out_μ_τ, )
+        @test metadata(node)                === 1
+
+        node = make_node(NormalMeanPrecision, factorisation = ((1, ), (2, 3)))
+
+        @test functionalform(node)          === NormalMeanPrecision
+        @test sdtype(node)                  === Stochastic()
+        @test name.(interfaces(node))       === (:out, :μ, :τ)
+        @test factorisation(node)           === ((1, ), (2, 3, ))
+        @test localmarginalnames(node)      === (:out, :μ_τ, )
+        @test metadata(node)                === nothing
+    end
+
+    @testset "AverageEnergy" begin 
+
+        begin 
+            q_out = PointMass(1.0)
+            q_μ   = NormalMeanPrecision(1.0, 1.0)
+            q_τ   = GammaShapeRate(1.5, 1.5)
+
+            for N in (NormalMeanPrecision, NormalMeanVariance, NormalWeightedMeanPrecision), G in (GammaShapeRate, GammaShapeScale)
+                marginals = (Marginal(q_out, false, false), Marginal(convert(N, q_μ), false, false), Marginal(convert(G, q_τ), false, false))
+                @test score(AverageEnergy(), NormalMeanPrecision, Val{(:out, :μ, :τ)}, marginals, nothing) ≈ 1.6034261002694663
+            end
+        end
+
+        begin 
+            q_out = NormalMeanPrecision(1.0, 1.0)
+            q_μ   = NormalMeanPrecision(1.0, 1.0)
+            q_τ   = GammaShapeRate(1.5, 1.5)
+
+            for N in (NormalMeanPrecision, NormalMeanVariance, NormalWeightedMeanPrecision), G in (GammaShapeRate, GammaShapeScale)
+                marginals = (Marginal(q_out, false, false), Marginal(convert(N, q_μ), false, false), Marginal(convert(G, q_τ), false, false))
+                @test score(AverageEnergy(), NormalMeanPrecision, Val{(:out, :μ, :τ)}, marginals, nothing) ≈ 2.1034261002694663
+            end
+        end
+
+        begin 
+            q_out = PointMass(0.956629)
+            q_μ   = NormalMeanPrecision(0.255332, 0.762870)
+            q_τ   = GammaShapeRate(0.93037, 0.79312)
+
+            for N in (NormalMeanPrecision, NormalMeanVariance, NormalWeightedMeanPrecision), G in (GammaShapeRate, GammaShapeScale)
[…]
