### 2022-01-07 6375e0e
fix(): fix for mul_inplace
