+++ test/algebra/test_helpers.jl
@@ -8,6 +8,18 @@ using LinearAlgebra
 
 @testset "Helpers" begin
 
+    @testset "mirrorlog" begin 
+        rng  = MersenneTwister(1234)
+        vals = rand(rng, 10) 
+        @test ReactiveMP.mirrorlog.(vals) == map(x -> log(1 - x), vals)
+    end
+
+    @testset "xtlog" begin 
+        rng  = MersenneTwister(1234)
+        vals = rand(rng, 10) 
+        @test ReactiveMP.xtlog.(vals) == map(x -> x * log(x), vals)
+    end
+
     @testset "negate_inplace!" begin 
         rng = MersenneTwister(1234)
 
