+++ src/rules/mv_normal_mean_precision/marginals.jl
@@ -13,16 +13,29 @@ end
     xi_m, W_m  = weightedmean_precision(m_μ)
 
     W_bar = mean(m_Λ)
-    
+
+    T = promote_type(eltype(W_bar), eltype(W_y), eltype(W_m))
     d = length(xi_y)
-    Λ = repeat(W_bar, 2, 2)
-    view(Λ, 1:d, d+1:2*d) .*= -1
-    view(Λ, d+1:2*d, 1:d) .*= -1
-    view(Λ, 1:d, 1:d) .+= W_y
-    view(Λ, d+1:2*d, d+1:2*d) .+= W_m
+    Λ = Matrix{T}(undef, (2*d, 2*d))
+    @inbounds for k2 in 1:d
+        @inbounds for k1 in 1:d
+            tmp1 = W_bar[k1,k2]
+            tmp2 = -tmp1
+            k1d = k1+d
+            k2d = k2+d
+            Λ[k1,k2] = tmp1 + W_y[k1,k2]
+            Λ[k1d,k2] = tmp2
+            Λ[k1,k2d] = tmp2
+            Λ[k1d,k2d] = tmp1 + W_m[k1,k2]
+        end
+    end
 
     ξ  = [ xi_y; xi_m ]
     
+    # naive:
+    # Λ  = [ W_y + W_bar -W_bar; -W_bar W_m + W_bar ]
+    # ξ  = [ xi_y; xi_m ]
+
     return (out_μ = MvNormalWeightedMeanPrecision(ξ, Λ), Λ = m_Λ)
 end
 
@@ -32,14 +45,27 @@ end
 
     W_bar = mean(q_Λ)
     
+    T = promote_type(eltype(W_bar), eltype(W_y), eltype(W_m))
     d = length(xi_y)
-    Λ = repeat(W_bar, 2, 2)
-    view(Λ, 1:d, d+1:2*d) .*= -1
-    view(Λ, d+1:2*d, 1:d) .*= -1
-    view(Λ, 1:d, 1:d) .+= W_y
-    view(Λ, d+1:2*d, d+1:2*d) .+= W_m
+    Λ = Matrix{T}(undef, (2*d, 2*d))
+    @inbounds for k2 in 1:d
+        @inbounds for k1 in 1:d
+            tmp1 = W_bar[k1,k2]
+            tmp2 = -tmp1
+            k1d = k1+d
+            k2d = k2+d
+            Λ[k1,k2] = tmp1 + W_y[k1,k2]
+            Λ[k1d,k2] = tmp2
+            Λ[k1,k2d] = tmp2
+            Λ[k1d,k2d] = tmp1 + W_m[k1,k2]
+        end
+    end
     
     ξ  = [ xi_y; xi_m ]
+
+    # naive:
+    # Λ  = [ W_y + W_bar -W_bar; -W_bar W_m + W_bar ]
+    # ξ  = [ xi_y; xi_m ]
     
     return MvNormalWeightedMeanPrecision(ξ, Λ)
 end
