+++ src/rules/mv_normal_mean_covariance/marginals.jl
@@ -16,12 +16,25 @@ end
     
     xi = [ xi_out; xi_m ]
 
+    T = promote_type(eltype(W_bar), eltype(W_out), eltype(W_m))
     d = length(xi_out)
-    W = repeat(W_bar, 2, 2)
-    view(W, 1:d, d+1:2*d) .*= -1
-    view(W, d+1:2*d, 1:d) .*= -1
-    view(W, 1:d, 1:d) .+= W_out
-    view(W, d+1:2*d, d+1:2*d) .+= W_m
+    Λ = Matrix{T}(undef, (2*d, 2*d))
+    @inbounds for k2 in 1:d
+        @inbounds for k1 in 1:d
+            tmp1 = W_bar[k1,k2]
+            tmp2 = -tmp1
+            k1d = k1+d
+            k2d = k2+d
+            Λ[k1,k2] = tmp1 + W_out[k1,k2]
+            Λ[k1d,k2] = tmp2
+            Λ[k1,k2d] = tmp2
+            Λ[k1d,k2d] = tmp1 + W_m[k1,k2]
+        end
+    end
+
+    # naive:
+    # xi = [ xi_out; xi_m ]
+    # W  = [ W_out + W_bar -W_bar; -W_bar W_m + W_bar ]
 
     return (out_μ = MvNormalWeightedMeanPrecision(xi, W), Σ = m_Σ)
 end
@@ -34,12 +47,25 @@ end
     
     xi = [ xi_out; xi_m ]
     
+    T = promote_type(eltype(W_bar), eltype(W_out), eltype(W_m))
     d = length(xi_out)
-    W = repeat(W_bar, 2, 2)
-    view(W, 1:d, d+1:2*d) .*= -1
-    view(W, d+1:2*d, 1:d) .*= -1
-    view(W, 1:d, 1:d) .+= W_out
-    view(W, d+1:2*d, d+1:2*d) .+= W_m
+    Λ = Matrix{T}(undef, (2*d, 2*d))
+    @inbounds for k2 in 1:d
+        @inbounds for k1 in 1:d
+            tmp1 = W_bar[k1,k2]
+            tmp2 = -tmp1
+            k1d = k1+d
+            k2d = k2+d
+            Λ[k1,k2] = tmp1 + W_out[k1,k2]
+            Λ[k1d,k2] = tmp2
+            Λ[k1,k2d] = tmp2
+            Λ[k1d,k2d] = tmp1 + W_m[k1,k2]
+        end
+    end
+
+    # naive:
+    # xi = [ xi_out; xi_m ]
+    # W  = [ W_out + W_bar -W_bar; -W_bar W_m + W_bar ]
     
     return MvNormalWeightedMeanPrecision(xi, W)
 end
