+++ test/algebra/test_standard_basis_vector.jl
@@ -76,7 +76,7 @@ using LinearAlgebra
                 end
             end
         end
-
+    end
     # Different sizes
     @testset begin 
 
