### 2022-01-10 f08157f
feat(): rename logmean to mean(::log)
