+++ src/rules/bernoulli/out.jl
@@ -9,7 +9,7 @@ end
 end
 
 @rule Bernoulli(:out, Marginalisation) (q_p::Beta, ) = begin
-    rho_1 = clamp(exp(logmean(q_p)), tiny, huge)
+    rho_1 = clamp(exp(mean(log, q_p)), tiny, huge)
     rho_2 = clamp(exp(mirroredlogmean(q_p)), tiny, huge)
     return Bernoulli(rho_1 / (rho_1 + rho_2))
 end
