+++ src/nodes/categorical.jl
@@ -1,4 +1,4 @@
 
 @node Categorical Stochastic [ out, p ]
 
-@average_energy Categorical (q_out::Categorical, q_p::Any) = -sum(probvec(q_out) .* logmean(q_p))
+@average_energy Categorical (q_out::Categorical, q_p::Any) = -sum(probvec(q_out) .* mean(log, q_p))
