+++ src/distributions/beta.jl
@@ -14,7 +14,7 @@ function prod(::ProdAnalytical, left::Beta, right::Beta)
     return Beta(left_a + right_a - one(T), left_b + right_b - one(T))
 end
 
-function logmean(dist::Beta) 
+function mean(::typeof(log), dist::Beta) 
     a, b = params(dist)
     return digamma(a) - digamma(a + b)
 end
