+++ src/distributions/sample_list.jl
@@ -184,7 +184,8 @@ function weightedmean_precision(sl::SampleList)
 end
 
 weightedmean(sl::SampleList)    = first(weightedmean_precision(sl))
-logmean(sl::SampleList)         = sample_list_logmean(variate_form(sl), sl)
+
+mean(::typeof(log), sl::SampleList) = sample_list_logmean(variate_form(sl), sl)
 meanlogmean(sl::SampleList)     = sample_list_meanlogmean(variate_form(sl), sl)
 mirroredlogmean(sl::SampleList) = sample_list_mirroredlogmean(variate_form(sl), sl)
 
