### 2022-01-10 1f1f931
feat(): rename meanlogmean to mean(::xtlog)
