+++ src/distributions/gamma_shape_rate.jl
@@ -40,7 +40,7 @@ function loggammamean(dist::GammaShapeRate)
     return 0.5 * (log2π - (digamma(a) - log(b))) + mean(dist) * (-1 + digamma(a + 1) - log(b))
 end
 
-function meanlogmean(dist::GammaShapeRate)
+function mean(::typeof(xtlog), dist::GammaShapeRate)
     a, b = params(dist)
     return mean(dist) * (digamma(a + 1) - log(b))
 end
