+++ test/distributions/test_gamma.jl
@@ -150,9 +150,9 @@ using Random
             @test pdf(left, 10.0)    ≈ pdf(right, 10.0)
             @test logpdf(left, 1.0)  ≈ logpdf(right, 1.0)
             @test logpdf(left, 10.0) ≈ logpdf(right, 10.0)
-            @test logmean(left)      ≈ logmean(right)
+            @test mean(log, left)    ≈ mean(log, right)
             @test loggammamean(left) ≈ loggammamean(right)
-            @test meanlogmean(left)  ≈ meanlogmean(right)
+            @test mean(xtlog, left)  ≈ mean(xtlog, right)
         end
 
         types = ReactiveMP.union_types(GammaDistributionsFamily{Float64})
