### 2022-01-10 b325ae1
feat(): rename mirroredlogmean to mean(::typeof(mirrorlog))
