+++ test/distributions/test_pointmass.jl
@@ -52,9 +52,9 @@ import ReactiveMP.MacroHelpers: @test_inferred
             @test @test_inferred(Type{T}, eltype(dist)) == T
 
             @test_throws ErrorException probvec(dist) 
-            @test @test_inferred(T, logmean(dist))         == log(scalar)
+            @test @test_inferred(T, mean(log, dist))       == log(scalar)
             @test @test_inferred(T, mean(inv, dist))       == inv(scalar)
-            @test @test_inferred(T, mirroredlogmean(dist)) == log(one(scalar) - scalar)
+            @test @test_inferred(T, mean(mirrorlog, dist)) == log(one(scalar) - scalar)
             @test @test_inferred(T, loggammamean(dist))    == loggamma(scalar)
         end
     end
@@ -100,9 +100,9 @@ import ReactiveMP.MacroHelpers: @test_inferred
             @test @test_inferred(Type{T}, eltype(dist)) == T
 
             @test @test_inferred(AbstractVector{T}, probvec(dist))      == vector
-            @test @test_inferred(AbstractVector{T}, logmean(dist))      == log.(vector)
+            @test @test_inferred(AbstractVector{T}, mean(log, dist))    == log.(vector)
             @test_throws ErrorException mean(inv, dist)
-            @test_throws ErrorException mirroredlogmean(dist)
+            @test_throws ErrorException mean(mirrorlog, dist)
             @test @test_inferred(AbstractVector{T}, loggammamean(dist)) == loggamma.(vector)
         end
     end
@@ -150,9 +150,9 @@ import ReactiveMP.MacroHelpers: @test_inferred
             @test_throws ErrorException precision(dist)
 
             @test_throws ErrorException probvec(dist)
-            @test @test_inferred(AbstractMatrix{T}, logmean(dist))      == log.(matrix)
+            @test @test_inferred(AbstractMatrix{T}, mean(log, dist))  == log.(matrix)
             @test @test_inferred(AbstractMatrix{T}, mean(inv, dist))  ≈ cholinv(matrix)
-            @test_throws ErrorException mirroredlogmean(dist)
+            @test_throws ErrorException mean(mirrorlog, dist)
             @test @test_inferred(AbstractMatrix{T}, loggammamean(dist)) == loggamma.(matrix)
         end
     end
