+++ test/distributions/test_dirichlet_matrix.jl
@@ -46,10 +46,10 @@ using Random
         @test entropy(MatrixDirichlet([ 0.2 3.4; 5.0 11.0; 0.2 0.6 ])) ≈ -11.444984495104693
     end
 
-    @testset "logmean" begin
-        @test logmean(MatrixDirichlet([ 1.0 1.0; 1.0 1.0; 1.0 1.0 ])) ≈ [-1.5000000000000002 -1.5000000000000002; -1.5000000000000002 -1.5000000000000002; -1.5000000000000002 -1.5000000000000002]
-        @test logmean(MatrixDirichlet([ 1.2 3.3; 4.0 5.0; 2.0 1.1 ])) ≈ [-2.1920720408623637 -1.1517536610071326; -0.646914475838374 -0.680458481634953; -1.480247809171707 -2.6103310904778305]
-        @test logmean(MatrixDirichlet([ 0.2 3.4; 5.0 11.0; 0.2 0.6 ])) ≈ [-6.879998107291004 -1.604778825293528; -0.08484054226701443 -0.32259407259407213; -6.879998107291004 -4.214965875553984]
+    @testset "mean(::typeof(log))" begin
+        @test mean(log, MatrixDirichlet([ 1.0 1.0; 1.0 1.0; 1.0 1.0 ])) ≈ [-1.5000000000000002 -1.5000000000000002; -1.5000000000000002 -1.5000000000000002; -1.5000000000000002 -1.5000000000000002]
+        @test mean(log, MatrixDirichlet([ 1.2 3.3; 4.0 5.0; 2.0 1.1 ])) ≈ [-2.1920720408623637 -1.1517536610071326; -0.646914475838374 -0.680458481634953; -1.480247809171707 -2.6103310904778305]
+        @test mean(log, MatrixDirichlet([ 0.2 3.4; 5.0 11.0; 0.2 0.6 ])) ≈ [-6.879998107291004 -1.604778825293528; -0.08484054226701443 -0.32259407259407213; -6.879998107291004 -4.214965875553984]
     end
 
     @testset "prod" begin
