### 2022-01-10 c6225f6
feat(): rename loggammamean to mean(::typeof(loggamma))
