+++ src/constraints/prod/prod_generic.jl
@@ -61,8 +61,6 @@ Base.size(product::DistProduct)      = error("size() is not defined for $(produc
 probvec(product::DistProduct)         = error("probvec() is not defined for $(product). DistProduct structure has to be approximated and cannot be used in inference procedure.")
 weightedmean(product::DistProduct)    = error("weightedmean() is not defined for $(product). DistProduct structure has to be approximated and cannot be used in inference procedure.")
 
-loggammamean(product::DistProduct)    = error("loggammamean() is not defined for $(product). DistProduct structure has to be approximated and cannot be used in inference procedure.")
-
 variate_form(::P)                         where { P <: DistProduct } = variate_form(P)
 variate_form(::Type{ DistProduct{L, R} }) where { L, R }             = _check_dist_product_variate_form(variate_form(L), variate_form(R))
 

[…]
