+++ test/distributions/test_dirichlet.jl
@@ -37,7 +37,7 @@ using Random
     end
 
     @testset "mean(::typeof(log))" begin
-        @test mean(log, Dirichlet([ 1.0, 1.0, 1.0 ])) ≈ [-1.5000000000000002, -1.5000000000000002, -1.5000000000000002]åå
+        @test mean(log, Dirichlet([ 1.0, 1.0, 1.0 ])) ≈ [-1.5000000000000002, -1.5000000000000002, -1.5000000000000002]
         @test mean(log, Dirichlet([ 1.1, 2.0, 2.0 ])) ≈ [-1.9517644694670657, -1.1052251939575213, -1.1052251939575213]
         @test mean(log, Dirichlet([ 3.0, 1.2, 5.0 ])) ≈ [-1.2410879175727905, -2.4529121492634465, -0.657754584239457]
     end
