+++ src/rules/mv_normal_mean_covariance/marginals.jl
@@ -18,17 +18,17 @@ end
 
     T = promote_type(eltype(W_bar), eltype(W_out), eltype(W_m))
     d = length(xi_out)
-    Λ = Matrix{T}(undef, (2*d, 2*d))
+    W = Matrix{T}(undef, (2*d, 2*d))
     @inbounds for k2 in 1:d
-        @inbounds for k1 in 1:d
+        for k1 in 1:d
             tmp1 = W_bar[k1,k2]
             tmp2 = -tmp1
             k1d = k1+d
             k2d = k2+d
-            Λ[k1,k2] = tmp1 + W_out[k1,k2]
-            Λ[k1d,k2] = tmp2
-            Λ[k1,k2d] = tmp2
-            Λ[k1d,k2d] = tmp1 + W_m[k1,k2]
+            W[k1,k2] = tmp1 + W_out[k1,k2]
+            W[k1d,k2] = tmp2
+            W[k1,k2d] = tmp2
+            W[k1d,k2d] = tmp1 + W_m[k1,k2]
         end
     end
 
@@ -49,17 +49,17 @@ end
     
     T = promote_type(eltype(W_bar), eltype(W_out), eltype(W_m))
     d = length(xi_out)
-    Λ = Matrix{T}(undef, (2*d, 2*d))
+    W = Matrix{T}(undef, (2*d, 2*d))
     @inbounds for k2 in 1:d
-        @inbounds for k1 in 1:d
+        for k1 in 1:d
             tmp1 = W_bar[k1,k2]
             tmp2 = -tmp1
             k1d = k1+d
             k2d = k2+d
-            Λ[k1,k2] = tmp1 + W_out[k1,k2]
-            Λ[k1d,k2] = tmp2
-            Λ[k1,k2d] = tmp2
-            Λ[k1d,k2d] = tmp1 + W_m[k1,k2]
+            W[k1,k2] = tmp1 + W_out[k1,k2]
+            W[k1d,k2] = tmp2
+            W[k1,k2d] = tmp2
+            W[k1d,k2d] = tmp1 + W_m[k1,k2]
         end
     end
 
