### 2022-01-10 87a4990
feat(): tests for average energy of gaussians
