+++ test/nodes/test_mv_normal_mean_precision.jl
@@ -42,7 +42,7 @@ import ReactiveMP: make_node
         begin 
             q_out = PointMass([ 1.0, 1.0 ])
             q_μ   = MvNormalMeanPrecision([ 1.0, 1.0 ], [ 1.0 0.0; 0.0 1.0 ])
-            q_Λ   = Wishart(4, [ 2.0 0.0; 0.0 2.0 ])
+            q_Λ   = Wishart(3, [ 2.0 0.0; 0.0 2.0 ])
 
             for N in (MvNormalMeanPrecision, MvNormalMeanCovariance, MvNormalWeightedMeanPrecision), G in (Wishart, )
                 marginals = (Marginal(q_out, false, false), Marginal(convert(N, q_μ), false, false), Marginal(convert(G, q_Λ), false, false))
@@ -50,58 +50,89 @@ import ReactiveMP: make_node
             end
         end
 
-        # begin 
-        #     q_out = NormalMeanPrecision(1.0, 1.0)
-        #     q_μ   = NormalMeanPrecision(1.0, 1.0)
-        #     q_τ   = GammaShapeRate(1.5, 1.5)
+        begin 
+            q_out = PointMass([0.3873049736301686, 0.6250550344669505, 0.9865681758036855])
+            q_μ   = MvNormalMeanPrecision([0.21393068431529905, 0.5271266845217295, 0.0032162989237940476], [1.1178940049080035 1.004643185289189 0.5726053202206195; 1.004643185289189 0.9113807582278761 0.515956570765438; 0.5726053202206195 0.515956570765438 0.3402071079911484])
+            q_Λ   = Wishart(4, [0.4720138588203935 0.49760446060666863 0.4405032929933653; 0.49760446060666863 0.6757755009955487 0.636955076077564; 0.4405032929933653 0.636955076077564 0.6660144475917664])
 
-        #     for N in (NormalMeanPrecision, NormalMeanVariance, NormalWeightedMeanPrecision), G in (GammaShapeRate, GammaShapeScale)
-        #         marginals = (Marginal(q_out, false, false), Marginal(convert(N, q_μ), false, false), Marginal(convert(G, q_τ), false, false))
-        #         @test score(AverageEnergy(), NormalMeanPrecision, Val{(:out, :μ, :τ)}, marginals, nothing) ≈ 2.1034261002694663
-        #     end
-        # end
+            for N in (MvNormalMeanPrecision, MvNormalMeanCovariance, MvNormalWeightedMeanPrecision), G in (Wishart, )
+                marginals = (Marginal(q_out, false, false), Marginal(convert(N, q_μ), false, false), Marginal(convert(G, q_Λ), false, false))
+                @test score(AverageEnergy(), MvNormalMeanPrecision, Val{(:out, :μ, :Λ)}, marginals, nothing) ≈ 57.362404928342
+            end
+        end
 
-        # begin 
-        #     q_out = PointMass(0.956629)
-        #     q_μ   = NormalMeanPrecision(0.255332, 0.762870)
-        #     q_τ   = GammaShapeRate(0.93037, 0.79312)
+        begin 
+            q_out = MvNormalMeanPrecision([0.8625589157256603, 0.6694783342639599], [1.0014322413749484 0.7989099036521625; 0.7989099036521625 1.0976639268696966])
+            q_μ   = MvNormalMeanPrecision([0.9334416739853251, 0.38318522701093105], [0.21867945696266933 0.5704895781120056; 0.5704895781120056 1.5321190185800933])
+            q_Λ   = Wishart(3, [1.2509510680086597 1.3662166757006484; 1.3662166757006484 1.4925451500802907])
 
[…]
