+++ src/nodes/mv_normal_mean_covariance.jl
@@ -7,84 +7,39 @@ conjugate_type(::Type{ <: MvNormalMeanCovariance }, ::Type{ Val{ :Σ } })   = In
 
 # default method for mean-field assumption
 @average_energy MvNormalMeanCovariance (q_out::Any, q_μ::Any, q_Σ::Any) = begin
-    # naive: (ndims(q_out) * log2π + logdet(mean(q_Σ)) + tr(cholinv(mean(q_Σ))*(v_out + v_mean + (m_out - m_mean)*(m_out - m_mean)'))) / 2
+    # naive: (ndims(q_out) * log2π + mean(logdet, q_Σ) + tr(cholinv(mean(q_Σ))*(v_out + v_mean + (m_out - m_mean)*(m_out - m_mean)'))) / 2
     dim = ndims(q_out)
 
     m_mean, v_mean = mean_cov(q_μ)
     m_out, v_out   = mean_cov(q_out)
-    m_Σ            = mean(q_Σ)
+    inv_m_Σ        = mean(inv, q_Σ)
 
-    inv_m_Σ, result = cholinv_logdet(m_Σ)
-
+    result = zero(promote_type(eltype(q_out), eltype(q_μ), eltype(q_Σ)))
+    result += mean(logdet, q_Σ)
     result += dim * log2π
-    @turbo for k1 ∈ 1:dim, k2 ∈ 1:dim   # optimize trace operation (indices can be interchanges because of symmetry)
+    @inbounds for k1 in 1:dim, k2 in 1:dim   # optimize trace operation (indices can be interchanges because of symmetry)
         result += inv_m_Σ[k1,k2] * (v_out[k1,k2] + v_mean[k1,k2] + (m_out[k2]-m_mean[k2]) * (m_out[k1] - m_mean[k1]))
     end
     result /= 2
+    
     return result
-    
-end
-
-# specialized method for mean-field assumption with q_Σ::InverseWishart
-@average_energy MvNormalMeanCovariance (q_out::Any, q_μ::Any, q_Σ::InverseWishart) = begin
-    dim = ndims(q_out)
-
-    m_mean, v_mean = mean_cov(q_μ)
-    m_out, v_out   = mean_cov(q_out)
-    S_Σ = q_Σ.S.mat                     # work with S matrix to prevent extra allocation of mean matrix
-    r_Σ = q_Σ.df - dim + 1              
-    
-    inv_S_Σ, result = cholinv_logdet(S_Σ)
-    inv_S_Σ .*= r_Σ                     # compensate for working with S Matrix
-    
-    result += log2π * dim
-    result -= log(r_Σ) * dim            # compensate for working with S matrix
-    @turbo for k1 ∈ 1:dim, k2 ∈ 1:dim   # optimize trace operation (indices can be interchanges because of symmetry)
-        result += inv_S_Σ[k1,k2] * (v_out[k1,k2] + v_mean[k1,k2] + (m_out[k2]-m_mean[k2]) * (m_out[k1] - m_mean[k1]))
-    end
-    result /= 2
-    return result
-
 end
 
 # default method for structured mean-field assumption
 @average_energy MvNormalMeanCovariance (q_out_μ::Any, q_Σ::Any) = begin
-    # naive: @views (d*log2π + logdet(mean(q_Σ)) + tr(cholinv(mean(q_Σ))*( V[1:d,1:d] - V[1:d,d+1:end] - V[d+1:end,1:d] + V[d+1:end,d+1:end] + (m[1:d] - m[d+1:end])*(m[1:d] - m[d+1:end])' ))) / 2
[…]

[…]
