### 2022-01-10 e6c4896
test(): fix for phantom error in loopvec
