### 2022-01-11 0534f18
update(): Bump version to 1.3.0
