### 2022-01-24 92ace74
Add sketch inference
