### 2022-01-24 0bf492c
feat(): inference API [ WIP ]
