### 2022-01-24 7e20237
WIP: Update inference
