### 2022-01-25 8cd2ab1
fix bug multiplication node
