### 2022-01-26 3bdd42e
Allow for default meta data specification
