+++ src/scale_factor_helpers.jl
@@ -1,0 +1,10 @@
+export ScaleFactorMeta, ScaledMessage
+
+abstract type AbstractMeta end
+
+struct ScaleFactorMeta <: AbstractMeta end
+
+struct ScaledMessage{T}
+    message :: T
+    scale   :: Float64
+end
