### 2022-01-26 7909468
Add Scale Factor update rules for Normal and Multiplication
