+++ src/rules/mv_normal_mean_covariance/mean.jl
@@ -8,6 +8,25 @@
     return MvNormalMeanCovariance(m_out_mean, m_out_cov + mean(m_Σ))
 end
 
+# Message with scale factor (BP case)  
+@rule MvNormalMeanCovariance(:μ, Marginalisation) (m_out::PointMass, m_Σ::PointMass, meta::ScaleFactorMeta) = begin
+    message = @call_rule MvNormalMeanCovariance(:μ, Marginalisation) (m_out = m_out, m_Σ = m_Σ) 
+    scale = 0.0
+    return ScaledMessage(message,scale)
+end
+
+@rule MvNormalMeanCovariance(:μ, Marginalisation) (m_out::MultivariateNormalDistributionsFamily, m_Σ::PointMass, meta::ScaleFactorMeta) = begin
+    message = @call_rule MvNormalMeanCovariance(:μ, Marginalisation) (m_out = m_out, m_Σ = m_Σ) 
+    scale = 0.0
+    return ScaledMessage(message,scale)
+end
+
+@rule MvNormalMeanCovariance(:μ, Marginalisation) (m_out::ScaledMessage, m_Σ::PointMass, meta::ScaleFactorMeta) = begin 
+    message = @call_rule MvNormalMeanCovariance(:μ, Marginalisation) (m_out = m_out.message, m_Σ = m_Σ) 
+    scale = m_out.scale
+    return ScaledMessage(message, scale)
+end
+
 # Variational                       # 
 # --------------------------------- #
 @rule MvNormalMeanCovariance(:μ, Marginalisation) (q_out::Any, q_Σ::Any) = MvNormalMeanCovariance(mean(q_out), mean(q_Σ))
