### 2022-01-26 bd1d94b
feat(): init constraints
