+++ src/constraints/spec/factorisation_spec.jl
@@ -1,0 +1,136 @@
+
+import Base: merge!, show, ==, *, first, last, hash
+
+abstract type FactorisationSpecEntryIndex end
+
+struct FactorisationSpecEntryExact <: FactorisationSpecEntryIndex end
+struct FactorisationSpecEntryIndexed <: FactorisationSpecEntryIndex end
+struct FactorisationSpecEntryRanged <: FactorisationSpecEntryIndex end
+struct FactorisationSpecEntrySplitRanged <: FactorisationSpecEntryIndex end
+
+struct SplittedRange{ R <: AbstractRange }
+    range :: R
+end
+
+Base.first(range::SplittedRange) = first(range.range)
+Base.last(range::SplittedRange)  = last(range.range)
+
+struct FactorisationSpecEntry{I}
+    symbol :: Symbol
+    index  :: I
+end
+
+name(entry::FactorisationSpecEntry) = entry.symbol
+
+Base.show(io::IO, entry::FactorisationSpecEntry) = show(io, indextype(entry), entry)
+
+Base.show(io, ::FactorisationSpecEntryExact, entry::FactorisationSpecEntry) = print(io, entry.symbol)
+Base.show(io, ::FactorisationSpecEntryIndexed, entry::FactorisationSpecEntry) = print(io, entry.symbol, "[", entry.index, "]")
+Base.show(io, ::FactorisationSpecEntryRanged, entry::FactorisationSpecEntry) = print(io, entry.symbol, "[", entry.index, "]")
+Base.show(io, ::FactorisationSpecEntrySplitRanged, entry::FactorisationSpecEntry) = print(io, entry.symbol, "[", first(entry.index), "]..", entry.symbol, "[", last(entry.index), "]")
+
+function Base.:(==)(left::FactorisationSpecEntry, right::FactorisationSpecEntry)
+    return left.symbol == right.symbol && left.index == right.index
+end
+
+indextype(spec::FactorisationSpecEntry) = indextype(spec, spec.index)
+
+indextype(::FactorisationSpecEntry, index::Nothing)       = FactorisationSpecEntryExact()
+indextype(::FactorisationSpecEntry, index::Integer)       = FactorisationSpecEntryIndexed()
+indextype(::FactorisationSpecEntry, index::AbstractRange) = FactorisationSpecEntryRanged()
+indextype(::FactorisationSpecEntry, index::SplittedRange) = FactorisationSpecEntrySplitRanged()
+
+Base.merge!(left::NTuple{N, FactorisationSpecEntry}, right::FactorisationSpecEntry) where N = TupleTools.setindex(left, merge!(left[end], right), lastindex(left))
+Base.merge!(left::FactorisationSpecEntry, right::NTuple{N, FactorisationSpecEntry}) where N = TupleTools.setindex(right, merge!(left, right[begin]), firstindex(right))
+
+function Base.merge!(left::NTuple{N1, FactorisationSpecEntry}, right::NTuple{N2, FactorisationSpecEntry}) where { N1, N2 }
+    return TupleTools.insertat(left, lastindex(left), (merge!(left[end], right[begin]), right[begin + 1:end]...))
+end
+
+function Base.merge!(left::FactorisationSpecEntry, right::FactorisationSpecEntry)
+    if left.symbol !== right.symbol
+        error("Cannot merge factorisation specification entries with different names $(left) and $(right)")
+    end
+    return merge!(indextype(left), indextype(right), left, right) 
+end
+
+function Base.merge!(::FactorisationSpecEntryIndex, ::FactorisationSpecEntryIndex, left::FactorisationSpecEntry, right::FactorisationSpecEntry)
+    error("Cannot merge factorisation specification entries $(left) and $(right)")
+end
+
[…]
