+++ test/constraints/spec/test_factorisation_spec.jl
@@ -1,0 +1,93 @@
+module ReactiveMPFactorisationSpecTest 
+
+using Test
+using ReactiveMP 
+
+import ReactiveMP: FactorisationSpecEntryIndex
+import ReactiveMP: FactorisationSpecEntryExact, FactorisationSpecEntryIndexed, FactorisationSpecEntryRanged, FactorisationSpecEntrySplitRanged
+import ReactiveMP: SplittedRange
+import ReactiveMP: FactorisationSpecEntry, FactorisationSpec, FactorisationSpecNode
+import ReactiveMP: indextype
+
+@testset "show" begin 
+    @test repr(FactorisationSpecEntry(:z, nothing)) == "z"
+    @test repr(FactorisationSpecEntry(:x, nothing)) == "x"
+    @test repr(FactorisationSpecEntry(:z, 1)) == "z[1]"
+    @test repr(FactorisationSpecEntry(:x, 3)) == "x[3]"
+    @test repr(FactorisationSpecEntry(:z, 1:4)) == "z[1:4]"
+    @test repr(FactorisationSpecEntry(:x, 6:9)) == "x[6:9]"
+    @test repr(FactorisationSpecEntry(:z, SplittedRange(1:4))) == "z[1]..z[4]"
+    @test repr(FactorisationSpecEntry(:x, SplittedRange(6:9))) == "x[6]..x[9]"
+
+    @test repr(FactorisationSpec(( FactorisationSpecEntry(:z, 1), FactorisationSpecEntry(:s, 1) ))) == "q(z[1], s[1])"
+    @test repr(FactorisationSpec(( FactorisationSpecEntry(:z, 3), FactorisationSpecEntry(:s, 1) ))) == "q(z[3], s[1])"
+    @test repr(FactorisationSpec(( FactorisationSpecEntry(:z, 3), FactorisationSpecEntry(:s, nothing) ))) == "q(z[3], s)"
+    @test repr(FactorisationSpec(( FactorisationSpecEntry(:z, 1:2), FactorisationSpecEntry(:s, 1:2) ))) == "q(z[1:2], s[1:2])"
+    @test repr(FactorisationSpec(( FactorisationSpecEntry(:z, 1:2), FactorisationSpecEntry(:s, 1:2) ))) == "q(z[1:2], s[1:2])"
+    @test repr(FactorisationSpec(( FactorisationSpecEntry(:z, SplittedRange(1:2)), FactorisationSpecEntry(:s, nothing) ))) == "q(z[1]..z[2], s)"
+end
+
+@testset "indextype" begin 
+    
+    @test indextype(FactorisationSpecEntry(:z, nothing)) === FactorisationSpecEntryExact()
+    @test indextype(FactorisationSpecEntry(:z, 1)) === FactorisationSpecEntryIndexed()
+    @test indextype(FactorisationSpecEntry(:z, 1:2)) === FactorisationSpecEntryRanged()
+    @test indextype(FactorisationSpecEntry(:z, SplittedRange(1:2))) === FactorisationSpecEntrySplitRanged()
+
+end
+
+@testset "merge!" begin 
+
+    @test_throws ErrorException merge!(FactorisationSpecEntry(:z, nothing), FactorisationSpecEntry(:s, nothing))
+    @test merge!(FactorisationSpecEntry(:s, nothing), FactorisationSpecEntry(:s, nothing)) == FactorisationSpecEntry(:s, nothing)
+    @test_throws ErrorException merge!(FactorisationSpecEntry(:z, nothing), FactorisationSpecEntry(:z, 1))
[…]
