### 2022-01-26 8fabcf6
feat(): factorisation constraints WIP
