+++ src/constraints/spec/spec.jl
@@ -1,29 +1,26 @@
 
+import Base: show
+
 struct Constraints 
-    nodes :: Dict{FactorisationSpec, FactorisationSpecNode}
+    factorisation :: Dict{FactorisationSpec, FactorisationSpecNode}
 
     Constraints() = new(Dict{FactorisationSpec, FactorisationSpecNode}())
 end
 
-
-make_factorisation_spec_entry(constraints::Constraints, args...) = FactorisationSpecEntry(args...)
-
-function search_factorisation_spec(constraints::Constraints, args::FactorisationSpecEntry...) 
-    spec = FactorisationSpec(args)
-    if haskey(constraints.nodes, spec)
-        return constraints.nodes[spec]
-    else
-        return spec
+function Base.show(io::IO, constraints::Constraints)
+    for (key, value) in constraints.factorisation
+        println(io, key, " => ", value)
     end
 end
 
-function add_factorisation_node(constraints::Constraints, key::FactorisationSpec, entries::Union{ FactorisationSpec, FactorisationSpecNode }...)
-    if haskey(constraints.nodes, key)
-        error("Factorisation spec for $(key) exists already.")
-    end
-    childnodes = filter(d -> d isa FactorisationSpecNode, entries)
-    childspec  = filter(d -> d isa FactorisationSpec, entries)
-    node = FactorisationSpecNode(key, childnodes, childspec)
-    constraints.nodes[key] = node
+make_factorisation_spec_entry(constraints::Constraints, args...)                   = FactorisationSpecEntry(args...)
+make_factorisation_spec(constraints::Constraints, args::FactorisationSpecEntry...) = FactorisationSpec(args)
+
+add_factorisation_node(constraints::Constraints, key::FactorisationSpec, entries::FactorisationSpec) = add_factorisation_node(constraints, key, (entries, ))
+
+function add_factorisation_node(constraints::Constraints, key::FactorisationSpec, entries::NTuple{N, FactorisationSpec}) where N
+    !(haskey(constraints.factorisation, key)) || error("Factorisation spec for $(key) exists already.")
+    node = FactorisationSpecNode(entries)
+    constraints.factorisation[key] = node
     return constraints
 end
