### 2022-01-27 39263c5
feat(): better show methods and allow single block constraints specification
