+++ src/ReactiveMP.jl
@@ -12,9 +12,6 @@ include("message.jl")
 include("marginal.jl")
 include("distributions.jl")
 
-include("constraints/spec/factorisation_spec.jl")
-include("constraints/spec/spec.jl")
-
 include("constraints/prod/prod_analytical.jl")
 include("constraints/prod/prod_generic.jl")
 include("constraints/prod/prod_preserve_type.jl")
@@ -129,6 +126,9 @@ include("nodes/multiplication.jl")
 include("rules/prototypes.jl")
 
 include("actors/prior.jl")
+
+include("constraints/spec/factorisation_spec.jl")
+include("constraints/spec/spec.jl")
 
 include("model.jl")
 
