### 2022-01-27 c1eb1b5
feat(): validate factorisation constraints specification
