+++ src/constraints/spec/factorisation_spec.jl
@@ -21,6 +21,7 @@ struct FactorisationSpecEntry{I}
 end
 
 name(entry::FactorisationSpecEntry) = entry.symbol
+getindex(entry::FactorisationSpecEntry) = entry.index
 
 Base.show(io::IO, entry::FactorisationSpecEntry) = show(io, indextype(entry), entry)
 
@@ -73,11 +74,37 @@ function Base.merge!(::FactorisationSpecEntrySplitRanged, ::FactorisationSpecEnt
     return FactorisationSpecEntry(left.symbol, SplittedRange(first(left.index):last(right.index)))
 end
 
+function validate(entry::FactorisationSpecEntry, variable::AbstractVariable) 
+    (name(entry) === name(variable)) || error("`validate` expects name of factorisation specification entry to be the same as the `variable` name")
+    return validate(indextype(entry), entry, variable)
+end
+
+function validate(entry::FactorisationSpecEntry, variable::AbstractArray{ <: AbstractVariable }) 
+    (name(entry) === name(first(variable))) || error("`validate` expects name of factorisation specification entry to be the same as the `variable` name")
+    return validate(indextype(entry), entry, variable)
+end
+
+validate(::FactorisationSpecEntryExact, entry::FactorisationSpecEntry, variable::RandomVariable) = true
+validate(::FactorisationSpecEntryExact, entry::FactorisationSpecEntry, variable::AbstractArray{ <: RandomVariable }) = true
+validate(::FactorisationSpecEntryIndexed, entry::FactorisationSpecEntry, variable::AbstractVariable) = error("Single variable $(name(variable)) cannot be indexed in factorisation specification entry $(entry)")
+validate(::FactorisationSpecEntryIndexed, entry::FactorisationSpecEntry, variable::AbstractArray{ <: RandomVariable }) = true
+validate(::FactorisationSpecEntryRanged, entry::FactorisationSpecEntry, variable::AbstractVariable) = error("Single variable $(name(variable)) cannot be range-indexed in factorisation specification entry $(entry)")
+validate(::FactorisationSpecEntryRanged, entry::FactorisationSpecEntry, variable::AbstractArray{ <: RandomVariable }) = (first(getindex(entry)) >= firstindex(variable) && last(getindex(entry)) <= lastindex(variable)) || error("Index out of bounds for variable $(name(first(variable))) in factorisation specification entry $(entry)")
+validate(::FactorisationSpecEntrySplitRanged, entry::FactorisationSpecEntry, variable::AbstractVariable) = error("Single variable $(name(variable)) cannot be splitrange-indexed in factorisation specification entry $(entry)")
+validate(::FactorisationSpecEntrySplitRanged, entry::FactorisationSpecEntry, variable::AbstractArray{ <: RandomVariable }) = (first(getindex(entry)) >= firstindex(variable) && last(getindex(entry)) <= lastindex(variable)) || error("Index out of bounds for variable $(name(first(variable))) in factorisation specification entry $(entry)")
+
+validate(::FactorisationSpecEntryIndex, entry::FactorisationSpecEntry, variable::ConstVariable) = error("Constant $(name(variable)) should not be present in the factorisation constraints specification")
+validate(::FactorisationSpecEntryIndex, entry::FactorisationSpecEntry, variable::AbstractArray{ <: ConstVariable }) = error("Constant $(name(first(variable))) should not be present in the factorisation constraints specification")
+validate(::FactorisationSpecEntryIndex, entry::FactorisationSpecEntry, variable::DataVariable) = error("Data variable $(name(variable)) should not be present in the factorisation constraints specification")
[…]
