### 2022-01-27 7114713
feat(): more sanity checks
