+++ src/model.jl
@@ -5,7 +5,7 @@ export activate!, repeat!
 export UntilConvergence
 # export MarginalsEagerUpdate, MarginalsPureUpdate
 
-import Base: show, getindex, haskey
+import Base: show, getindex, haskey, firstindex, lastindex
 
 # Marginals update strategies 
 
@@ -98,6 +98,15 @@ end
 function Base.haskey(model::Model, symbol::Symbol)
     return haskey(getvardict(model), symbol)
 end
+
+firstindex(model::Model, symbol::Symbol) = firstindex(model, getindex(model, symbol))
+lastindex(model::Model, symbol::Symbol)  = lastindex(model, getindex(model, symbol))
+
+firstindex(::Model, ::AbstractVariable) = typemin(Int64)
+lastindex(::Model, ::AbstractVariable)  = typemax(Int64)
+
+firstindex(::Model, variables::AbstractVector{ <: AbstractVariable }) = firstindex(variables)
+lastindex(::Model, variables::AbstractVector{ <: AbstractVariable })  = lastindex(variables)
 
 add!(vardict::Dict, name::Symbol, entity) = vardict[name] = entity
 
