### 2022-01-27 3db71a5
feat(): extra checks for factorisation specification
