### 2022-01-27 #91 @matbesancon
@bvdmitri can you archive the repo on zenodo, precisely that corresponds to the version used in the paper
