### 2022-01-28 #91 @bvdmitri
@matbesancon We archived our repo on Zenodo here https://zenodo.org/record/5913617 . All statements in the paper are valid for the latest release too (we even improved our performance better and it became faster since the last JuliaCon).
