### 2022-01-28 #91 @matbesancon
@bvdmitri can you take a look at the paper produced and let us know if all good on your side?
